Studies on the Association Between Supervision Quality and Various Outcomes in Social, Mental Health, and Child Welfare Workers

Results from 5 studies examining the association between various measures of supervision quality and various work-related outcomes in social, mental health, and child welfare workers.

dat.aloe2013

Format

The data frame contains the following columns:

study	character	study author(s) and year

n	integer	sample size

tval	numeric	t-statistic for the test of the association/predictor

preds	integer	number of predictors included in the regression model

R2	numeric	the coefficient of determination (i.e., R-squared value) of the regression model

Details

The dataset is based on studies that used regression models to examine the association between some measure of perceived supervision quality (e.g., the quality of the relationship with one's supervisor) and some work-related outcome (e.g., job satisfaction) in social, mental health, and child welfare workers. The dataset was extracted from Aloe and Thompson (2013), which in turn is a subset of the studies included in the meta-analysis by Mor Barak et al. (2009).

The dataset can be used to illustrate the meta-analysis of regression models, using measures such as the (semi-)partial correlation coefficient. For this, the t-statistic from the regression model for the association (i.e., predictor) of interest was extracted from each regression model (tval), as well as the sample size (n), the number of predictors included in the regression model (preds), and the coefficient of determination (i.e., R-squared value) of the regression model (R2). Based on this information, the (semi-)partial correlation coefficient can be computed for each study, as well as its corresponding sampling variance. These values can then be meta-analyzed using standard methods.

Source

Aloe, A. M., & Thompson, C. G. (2013). The synthesis of partial effect sizes. Journal of the Society for Social Work and Research, 4(4), 390–405. https://doi.org/10.5243/jsswr.2013.24

References

Mor Barak, M. E., Travis, D. J., Pyun, H., & Xie, B. (2009). The impact of supervision on worker outcomes: A meta-analysis. Social Service Review, 83(1), 3–32. https://doi.org/10.1086/599028

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

social work, (semi-)partial correlations, meta-regression

Examples

### copy data into 'dat' and examine data
dat <- dat.aloe2013
dat
#>                      study   n  tval preds    R2
#> 1         Abu-Bader (2000) 218  4.61     4 0.240
#> 2       Cole et al. (2004) 232  6.19     7 0.455
#> 3 Wallach & Mueller (2006) 156  4.07     6 0.500
#> 4     Weaver et al. (2007) 382 -0.77    19 0.327
#> 5             Acker (2004) 259  1.16    15 0.117

### load metafor package
suppressPackageStartupMessages(library(metafor))

### compute the partial correlation coefficients and corresponding sampling variances
dat <- escalc(measure="PCOR", ti=tval, ni=n, mi=preds, data=dat)
dat
#> 
#>                      study   n  tval preds    R2      yi     vi 
#> 1         Abu-Bader (2000) 218  4.61     4 0.240  0.3012 0.0039 
#> 2       Cole et al. (2004) 232  6.19     7 0.455  0.3822 0.0032 
#> 3 Wallach & Mueller (2006) 156  4.07     6 0.500  0.3163 0.0054 
#> 4     Weaver et al. (2007) 382 -0.77    19 0.327 -0.0404 0.0027 
#> 5             Acker (2004) 259  1.16    15 0.117  0.0742 0.0041 
#> 

### random-effects model
res <- rma(yi, vi, data=dat)
res
#> 
#> Random-Effects Model (k = 5; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0298 (SE = 0.0237)
#> tau (square root of estimated tau^2 value):      0.1725
#> I^2 (total heterogeneity / total variability):   88.91%
#> H^2 (total variability / sampling variability):  9.01
#> 
#> Test for Heterogeneity:
#> Q(df = 4) = 40.1986, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub    
#>   0.2048  0.0820  2.4993  0.0124  0.0442  0.3655  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### mixed-effects meta-regression model examining the relationship between the partial
### correlation coefficients and the number of predictors included in the models
res <- rma(yi, vi, mods = ~ preds, data=dat)
res
#> 
#> Mixed-Effects Model (k = 5; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0014 (SE = 0.0043)
#> tau (square root of estimated tau^2 value):             0.0374
#> I^2 (residual heterogeneity / unaccounted variability): 26.42%
#> H^2 (unaccounted variability / sampling variability):   1.36
#> R^2 (amount of heterogeneity accounted for):            95.31%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 3) = 3.8516, p-val = 0.2779
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 25.1818, p-val < .0001
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb    ci.ub      
#> intrcpt    0.4832  0.0658   7.3389  <.0001   0.3542   0.6122  *** 
#> preds     -0.0269  0.0054  -5.0182  <.0001  -0.0374  -0.0164  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### compute the r-to-z transformed partial correlation coefficients and their variances
dat <- escalc(measure="ZPCOR", ti=tval, ni=n, mi=preds, data=dat)
dat
#> 
#>                      study   n  tval preds    R2      yi     vi 
#> 1         Abu-Bader (2000) 218  4.61     4 0.240  0.3108 0.0047 
#> 2       Cole et al. (2004) 232  6.19     7 0.455  0.4026 0.0045 
#> 3 Wallach & Mueller (2006) 156  4.07     6 0.500  0.3275 0.0068 
#> 4     Weaver et al. (2007) 382 -0.77    19 0.327 -0.0405 0.0028 
#> 5             Acker (2004) 259  1.16    15 0.117  0.0743 0.0041 
#> 

### random-effects model
res <- rma(yi, vi, data=dat)
res
#> 
#> Random-Effects Model (k = 5; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0324 (SE = 0.0261)
#> tau (square root of estimated tau^2 value):      0.1799
#> I^2 (total heterogeneity / total variability):   88.25%
#> H^2 (total variability / sampling variability):  8.51
#> 
#> Test for Heterogeneity:
#> Q(df = 4) = 38.1200, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub    
#>   0.2108  0.0859  2.4537  0.0141  0.0424  0.3791  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### back-transformation to the partial correlation scale
predict(res, transf=transf.ztor)
#> 
#>    pred  ci.lb  ci.ub   pi.lb  pi.ub 
#>  0.2077 0.0424 0.3620 -0.1781 0.5381 
#> 

### compute the semi-partial correlation coefficients and their variances
dat <- escalc(measure="SPCOR", ti=tval, ni=n, mi=preds, r2i=R2, data=dat)
dat
#> 
#>                      study   n  tval preds    R2      yi     vi 
#> 1         Abu-Bader (2000) 218  4.61     4 0.240  0.2754 0.0033 
#> 2       Cole et al. (2004) 232  6.19     7 0.455  0.3053 0.0023 
#> 3 Wallach & Mueller (2006) 156  4.07     6 0.500  0.2358 0.0032 
#> 4     Weaver et al. (2007) 382 -0.77    19 0.327 -0.0332 0.0018 
#> 5             Acker (2004) 259  1.16    15 0.117  0.0699 0.0034 
#> 

### random-effects model
res <- rma(yi, vi, data=dat)
res
#> 
#> Random-Effects Model (k = 5; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0194 (SE = 0.0157)
#> tau (square root of estimated tau^2 value):      0.1393
#> I^2 (total heterogeneity / total variability):   87.97%
#> H^2 (total variability / sampling variability):  8.31
#> 
#> Test for Heterogeneity:
#> Q(df = 4) = 38.6380, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub    
#>   0.1692  0.0666  2.5406  0.0111  0.0387  0.2996  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>