Studies on the Effectiveness of Oral Anticoagulants in Patients with Coronary Artery Disease

Results from 34 trials examining the effectiveness of oral anticoagulants in patients with coronary artery disease.

dat.anand1999

Format

The data frame contains the following columns:

study	character	author(s) or trial name

year	numeric	publication year

intensity	character	intensity of anticoagulation (low, medium, or high)

asp.t	numeric	concomitant use of aspirin in the treatment group (0 = no, 1 = yes)

asp.c	numeric	concomitant use of aspirin in the control group (0 = no, 1 = yes)

ai	numeric	number of deaths in the treatment group

n1i	numeric	number of patients in the treatment group

ci	numeric	number of deaths in the control group

n2i	numeric	number of patients in the control group

Details

The dataset includes the results from 34 randomized clinical trials that examined the effectiveness of oral anticoagulants in patients with coronary artery disease. The results given here are focused on the total mortality in the treatment versus control groups.

Note

Strictly speaking, there are only 31 trials, since Breddin et al. (1980) and ATACS (1990) are multiarm trials.

According to a correction, dat.anand1999$ci[29] should be 1. But then dat.anand1999$ci[21] would also have to be 1 (if these data indeed refer to the same control group). This appears contradictory, so this correction was not made.

Source

Anand, S. S., & Yusuf, S. (1999). Oral anticoagulant therapy in patients with coronary artery disease: A meta-analysis. Journal of the American Medical Association, 282(21), 2058–2067. https://doi.org/10.1001/jama.282.21.2058

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, cardiology, odds ratios, Mantel-Haenszel method

Examples

### copy data into 'dat' and examine data
dat <- dat.anand1999
dat
#>                            study year intensity asp.t asp.c  ai  n1i  ci  n2i
#> 1                MacMillan et al 1960      high     0     0   8   27   0   23
#> 2                  Borchegrevink 1960      high     0     0   1  103   8  100
#> 3                  Clausen et al 1961      high     0     0  15   93  13   99
#> 4                  Harvald et al 1961      high     0     0  34  145  45  171
#> 5  Apenstrom and Korsan-Bengtsen 1964      high     0     0  39  118  50  113
#> 6                   Conrad et al 1964      high     0     0   9   52   8   34
#> 7                Wasserman et al 1966      high     0     0  12   77  15   70
#> 8                 Loeliger et al 1967      high     0     0   8  128  11  112
#> 9                   Lovell et al 1967      high     0     0  33  178  39  172
#> 10                  Seaman et al 1969      high     0     0  36   88  31   87
#> 11                Sorensen et al 1969      high     0     0  30  156  43  120
#> 12       Sixty Plus Reinfarction 1980      high     0     0  51  439  69  439
#> 13                         WARIS 1990      high     0     0  94  607 123  607
#> 14                        ASPECT 1994      high     0     0 170 1700 189 1704
#> 15               Meuwissen et al 1969      high     0     0   1   68   8   70
#> 16           Drapkin and Merskey 1974      high     0     0 111  745 166  782
#> 17                 Breddin et al 1980      high     0     0  39  320  32  309
#> 18                 Breddin et al 1980      high     0     1  39  320  27  317
#> 19                       CABADAS 1993      high     0     1   3  307   8  309
#> 20               Eritsland et al 1996      high     0     1   5  319   9  291
#> 21                         ATACS 1990      high     0     1   1   24   0   32
#> 22                 McEnany et al 1982  moderate     0     1   1   68   1   71
#> 23                  Kraska et al 1981  moderate     0     1   5   60   7   60
#> 24                         EPSIM 1982  moderate     0     1  67  652  72  651
#> 25                          COOP 1969  moderate     0     0 120  385 114  350
#> 26             MRC Anticoagulant 1964  moderate     0     0  29  195  40  188
#> 27                Williams et al 1986  moderate     0     0   1   51   4   51
#> 28                 McEnany et al 1982  moderate     0     0   1   68   3   77
#> 29                         ATACS 1990      high     1     1   0   37   0   32
#> 30                 OASIS Pilot 2 1998  moderate     1     1   2   98   5   99
#> 31                    ATACS-Main 1994  moderate     1     1   2  105   2  109
#> 32                   OASIS Pilot 1998       low     1     1   2  155   3  154
#> 33                     Post-CABG 1997       low     1     1  28  674  39  677
#> 34                          CARS 1997       low     1     1 118 3382 102 3393

### load metafor package
library(metafor)

### High-Intensity OA vs Control
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity=="high" & asp.t==0 & asp.c==0), digits=2)
#> 
#> Equal-Effects Model (k = 17)
#> 
#> I^2 (total heterogeneity / total variability):  41.89%
#> H^2 (total variability / sampling variability): 1.72
#> 
#> Test for Heterogeneity: 
#> Q(df = 16) = 27.54, p-val = 0.04
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.26  0.06  -4.65  <.01  -0.37  -0.15 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.77   0.69   0.86 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 21.49, df = 1,  p-val < 0.01
#> Tarone's Test for Heterogeneity: X^2 = 34.88, df = 16, p-val < 0.01
#> 

### High- or Moderate-Intensity OA vs Aspirin
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity %in% c("high","moderate") & asp.t==0 & asp.c==1), digits=2)
#> 
#> Equal-Effects Model (k = 7)
#> 
#> I^2 (total heterogeneity / total variability):  16.57%
#> H^2 (total variability / sampling variability): 1.20
#> 
#> Test for Heterogeneity: 
#> Q(df = 6) = 7.19, p-val = 0.30
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.03  0.13  -0.26  0.80  -0.30   0.23 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.97   0.74   1.26 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 0.04, df = 1, p-val = 0.85
#> Tarone's Test for Heterogeneity: X^2 = 8.05, df = 6, p-val = 0.23
#> 

### Moderate-Intensity OA vs Control
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity=="moderate" & asp.t==0 & asp.c==0), digits=2)
#> 
#> Equal-Effects Model (k = 4)
#> 
#> I^2 (total heterogeneity / total variability):  5.55%
#> H^2 (total variability / sampling variability): 1.06
#> 
#> Test for Heterogeneity: 
#> Q(df = 3) = 3.18, p-val = 0.37
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.20  0.13  -1.48  0.14  -0.46   0.06 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.82   0.63   1.07 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 1.99, df = 1, p-val = 0.16
#> Tarone's Test for Heterogeneity: X^2 = 3.35, df = 3, p-val = 0.34
#> 

### High- or Moderate-Intensity OA and Aspirin vs Aspirin
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity %in% c("high","moderate") & asp.t==1 & asp.c==1), digits=2)
#> Warning: Some yi/vi values are NA.
#> 
#> Equal-Effects Model (k = 3)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 0.55
#> 
#> Test for Heterogeneity: 
#> Q(df = 1) = 0.55, p-val = 0.46
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.55  0.63  -0.87  0.38  -1.80   0.69 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.57   0.17   1.99 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 0.33, df = 1, p-val = 0.56
#> Tarone's Test for Heterogeneity: X^2 = 0.56, df = 1, p-val = 0.45
#> 

### Low-Intensity OA and Aspirin vs Aspirin
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity=="low" & asp.t==1 & asp.c==1), digits=2)
#> 
#> Equal-Effects Model (k = 3)
#> 
#> I^2 (total heterogeneity / total variability):  37.85%
#> H^2 (total variability / sampling variability): 1.61
#> 
#> Test for Heterogeneity: 
#> Q(df = 2) = 3.22, p-val = 0.20
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.03  0.12  0.27  0.79  -0.20   0.27 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.03   0.82   1.30 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 0.04, df = 1, p-val = 0.84
#> Tarone's Test for Heterogeneity: X^2 = 3.25, df = 2, p-val = 0.20
#>