Mortality Outcomes with Hydroxychloroquine and Chloroquine in COVID-19 from an International Collaborative Meta-Analysis of Randomized Trials

Results from 33 trials examining the effectiveness of hydroxychloroquine or chloroquine in patients with COVID-19.

dat.axfors2021

Format

The data frame contains the following columns:

id	character	registry number

acronym	character	shortened registry number

patient_setting	character	patient setting

blinding_exact	character	study blinding

high_dose	character	high or low dose of medication

Published	character	publication status

hcq_cq	character	medication type (hcq = hydroxychloroquine or cq = chloroquine)

hcq_arm_event	numeric	number of deaths in the treatment group

hcq_arm_total	numeric	number of patients in the treatment group

control_arm_event	numeric	number of deaths in the control group

control_arm_total	numeric	number of patients in the control group

Control	character	control group type (Standard of Care or Placebo)

Details

The dataset includes the results from 33 published and unpublished randomized clinical trials that examined the effectiveness of hydroxychloroquine or chloroquine in patients with COVID-19. The results given here are focused on the total mortality in the treatment versus control groups.

References

Axfors, C., Schmitt, A. M., Janiaud, P., van’t Hooft, J., Abd-Elsalam, S., Abdo, E. F., Abella, B. S., Akram, J., Amaravadi, R. K., Angus, D. C., Arabi, Y. M., Azhar, S., Baden, L. R., Baker, A. W., Belkhir, L., Benfield, T., Berrevoets, M. A. H., Chen, C.-P., Chen, T.-C., … Hemkens, L. G. (2021). Mortality outcomes with hydroxychloroquine and chloroquine in COVID-19 from an international collaborative meta-analysis of randomized trials. Nature Communications, 12(1), 2349. https://doi.org/10.1038/s41467-021-22446-z

Source

Axfors, C., Schmitt, A., Janiaud, P., van ’t Hooft, J., Moher, D., Goodman, S., … Hemkens, L. G. (2021, March 9). Hydroxychloroquine and chloroquine for survival in COVID-19: An international collaborative meta-analysis of randomized trials. https://doi.org/10.17605/OSF.IO/QESV4

Author

W. Kyle Hamilton whamilton@ucmerced.edu https://kylehamilton.com

Examples

# copy data into 'dat' and examine data
dat <- dat.axfors2021
dat

#>                  id             acronym patient_setting            covid_status
#> 1       NCT04353336         NCT04353336       Inpatient               Confirmed
#> 2       NCT04329923               PATCH       Inpatient               Confirmed
#> 3       NCT04338698           PROTECT C       Inpatient               Confirmed
#> 4       NCT04338698           PROTECT A       Inpatient               Confirmed
#> 5       NCT04338698           PROTECT B       Inpatient               Confirmed
#> 6       NCT04345289              CCAP-1       Inpatient               Confirmed
#> 7       NCT04308668           COVID-PEP      Outpatient Confirmed and Suspected
#> 8       NCT04384380         NCT04384380       Inpatient               Confirmed
#> 9       NCT04316377         NO COVID-19       Inpatient               Confirmed
#> 10      NCT04325893         NCT04325893       Inpatient               Confirmed
#> 11      NCT04342221             COV-HCQ       Inpatient               Confirmed
#> 12      NCT02735707           REMAP-CAP             ICU Confirmed and Suspected
#> 13 ChiCTR2000029868    ChiCTR2000029868       Inpatient               Confirmed
#> 14      NCT04381936            RECOVERY       Inpatient Confirmed and Suspected
#> 15      NCT04342650    CloroCOVID19II B      Outpatient               Suspected
#> 16      NCT04323527    CloroCOVID19II A       Inpatient               Suspected
#> 17      NCT04261517         NCT04261517       Inpatient               Confirmed
#> 18      NCT04369742               TEACH       Inpatient               Confirmed
#> 19      NCT04315896               HYDRA       Inpatient               Confirmed
#> 20 ChiCTR2000031204    ChiCTR2000031204       Inpatient               Confirmed
#> 21      NCT04333654         NCT04333654      Outpatient               Confirmed
#> 22      NCT04345692        OAHU-COVID19       Inpatient               Confirmed
#> 23      NCT04335552       NCT04335552 A       Inpatient               Confirmed
#> 24      NCT04335552       NCT04335552 B       Inpatient               Confirmed
#> 25           NL8490             ARCHAIC       Inpatient               Confirmed
#> 26           NL8490             ARCHAIC       Inpatient               Confirmed
#> 27      NCT04322123         Coalition I       Inpatient Confirmed and Suspected
#> 28 ChiCTR2000030054    ChiCTR2000030054       Inpatient               Confirmed
#> 29 ChiCTR2000030054    ChiCTR2000030054       Inpatient               Confirmed
#> 30      NCT04304053 BCN PEP CoV-2 Study      Outpatient               Confirmed
#> 31 ChiCTR2000029559    ChiCTR2000029559       Inpatient               Confirmed
#> 32      NCT04340544              COMIHY      Outpatient               Confirmed
#> 33      NCT04491994         NCT04491994       Inpatient               Confirmed
#>            blinding_exact high_dose     Published hcq_cq hcq_arm_total
#> 1                    None        no     Published    hcq            97
#> 2  Participant, Caregiver        no Not published    hcq            15
#> 3            Investigator        no Not published    hcq            64
#> 4            Investigator        no Not published    hcq            62
#> 5            Investigator        no Not published    hcq            59
#> 6  Participant, Caregiver        no Not published    hcq             1
#> 7  Participant, Caregiver        no     Published    hcq           244
#> 8                    None        no     Published    hcq            21
#> 9                    None        no     Published    hcq            27
#> 10 Participant, Caregiver        no Not published    hcq           124
#> 11 Participant, Caregiver        no Not published    hcq            13
#> 12                   None       yes Not published    hcq            61
#> 13                   None        no     Published    hcq            75
#> 14                   None       yes     Published    hcq          1561
#> 15 Participant, Caregiver        no Not published     cq            78
#> 16 Participant, Caregiver        no Not published     cq            41
#> 17                   None        no     Published    hcq            15
#> 18 Participant, Caregiver        no     Published    hcq            67
#> 19 Participant, Caregiver        no Not published    hcq            75
#> 20            Participant        no Not published     cq            18
#> 21 Participant, Caregiver        no Not published    hcq             5
#> 22                   None        no Not published    hcq            10
#> 23                   None        no Not published    hcq             4
#> 24                   None        no Not published    hcq             2
#> 25                   None        no Not published     cq             5
#> 26                   None        no Not published    hcq             4
#> 27                   None        no     Published    hcq           221
#> 28                   None        no     Published    hcq            18
#> 29                   None        no     Published     cq            18
#> 30                   None        no     Published    hcq           136
#> 31 Participant, Caregiver        no     Published    hcq            31
#> 32 Participant, Caregiver        no Not published    hcq             8
#> 33                   none        no     Published    hcq           349
#>    hcq_arm_event control_arm_total control_arm_event          Control
#> 1              6                97                 5 Standard of Care
#> 2              0                15                 0          Placebo
#> 3              0                64                 0 Standard of Care
#> 4              0                63                 1 Standard of Care
#> 5              1                61                 2 Standard of Care
#> 6              0                 1                 0          Placebo
#> 7              1               247                 1          Placebo
#> 8              0                12                 0 Standard of Care
#> 9              1                26                 1 Standard of Care
#> 10             6               123                11          Placebo
#> 11             1                14                 0          Placebo
#> 12            17                81                22 Standard of Care
#> 13             0                75                 0 Standard of Care
#> 14           418              3155               788 Standard of Care
#> 15             0                74                 0          Placebo
#> 16            17                41                12          Placebo
#> 17             0                15                 0 Standard of Care
#> 18             7                61                 6          Placebo
#> 19            28                77                29          Placebo
#> 20             0                17                 0          Placebo
#> 21             0                 3                 0          Placebo
#> 22             2                 6                 0 Standard of Care
#> 23             1                 2                 0 Standard of Care
#> 24             1                 3                 2 Standard of Care
#> 25             1                 3                 0 Standard of Care
#> 26             2                 3                 0 Standard of Care
#> 27             7               227                 6 Standard of Care
#> 28             0                12                 0 Standard of Care
#> 29             0                12                 0 Standard of Care
#> 30             0               157                 0 Standard of Care
#> 31             0                31                 0 Standard of Care
#> 32             0                 8                 0          Placebo
#> 33             0               151                 0 Standard of Care

# \dontrun{

# load metafor package
require(metafor)

# calculate log odds ratios and corresponding sampling variances
dat <- escalc(measure="OR", ai=hcq_arm_event, n1i=hcq_arm_total,
                            ci=control_arm_event, n2i=control_arm_total, data=dat)

# meta-analysis Hydroxychloroquine
res_hcq <- rma(yi, vi, subset=(hcq_cq=="hcq"), slab = id, data=dat)
print(res_hcq, digits=2)

#> 
#> Random-Effects Model (k = 28; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0 (SE = 0.04)
#> tau (square root of estimated tau^2 value):      0
#> I^2 (total heterogeneity / total variability):   0.00%
#> H^2 (total variability / sampling variability):  1.00
#> 
#> Test for Heterogeneity:
#> Q(df = 27) = 6.03, p-val = 1.00
#> 
#> Model Results:
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.08  0.07  1.18  0.24  -0.05   0.20    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# meta-analysis Chloroquine
res_cq <- rma(yi, vi, subset=(hcq_cq=="cq"), slab = id, data=dat)
print(res_cq, digits=2)

#> 
#> Random-Effects Model (k = 5; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0 (SE = 1.22)
#> tau (square root of estimated tau^2 value):      0
#> I^2 (total heterogeneity / total variability):   0.00%
#> H^2 (total variability / sampling variability):  1.00
#> 
#> Test for Heterogeneity:
#> Q(df = 4) = 0.38, p-val = 0.98
#> 
#> Model Results:
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.46  0.42  1.10  0.27  -0.36   1.29    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }