Studies on Ketotifen Alone or as Additional Medication for Long-Term Control of Asthma and Wheeze in Children

Results from 10 trials reporting the physicians' judgement on the overall efficacy of ketotifen for long-term control of asthma and wheeze in children.

dat.bassler2004

Format

The data frame contains the following columns:

study	character	study label

Ee	integer	number of children with treament success (ketotifen group)

Ne	integer	number of children (ketotifen group)

Ec	integer	number of children with treament success (control group)

Nc	integer	number of children (control group)

blind	character	blinding of clinicians

Details

Results from 10 trials reporting the physicians' judgement on the overall efficacy of Ketotifen for long-term control of asthma and wheeze in children. A prespecified subgroup analysis was conducted to evaluate whether the treatment effect is different in trials with adequate blinding compared to trials with inadequate / unclear blinding.

This data set is used as an example in Schwarzer et al. (2015).

Source

Bassler D., Mitra A. A. D., Ducharme F. M., Forster J., & Schwarzer, G. (2004). Ketotifen alone or as additional medication for long-term control of asthma and wheeze in children. Cochrane Database of Systematic Reviews, 1, CD001384. https://doi.org/10.1002/14651858.CD001384.pub2

References

Schwarzer, G., Carpenter, J. R., & Rücker, G. (2015). Meta-analysis with R. Cham, Switzerland: Springer.

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

risk ratios, medicine, subgroup analysis

Examples

### Show full data set
dat.bassler2004
#>                 study Ee Ne Ec Nc             blind
#> 1           Chay 1992  1 10  6 10 Adequate blinding
#> 2        Rackham 1989 31 68 38 65 Adequate blinding
#> 3    Van Asperen 1992 16 52 19 51 Adequate blinding
#> 4          Croce 1995 19 39 17 36    Method unclear
#> 5  de Benedictis 1990  7 34 35 41    Method unclear
#> 6          Longo 1986 10 18 15 18    Method unclear
#> 7        Montoya 1988  6 20 14 20    Method unclear
#> 8        Mulhern 1982  6 16  8 15    Method unclear
#> 9              Salmon  8 28 16 34    Method unclear
#> 10        Spicak 1983  9 25 20 25    Method unclear

### Load meta package
suppressPackageStartupMessages(library("meta"))

### Use DerSimonian-Laird estimator (which was the default in meta in the year 2015).
### Furthermore, print meta-analysis results with two digits.
oldset <- settings.meta(method.tau = "DL", digits = 2)

### Calculate experimental and control event rates
with(dat.bassler2004, summary(Ee / Ne))
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.1000  0.2893  0.3338  0.3433  0.4357  0.5556 
with(dat.bassler2004, summary(Ec / Nc))
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.3725  0.4875  0.5923  0.6220  0.7750  0.8537 

### Conduct meta-analysis using the inverse variance method
mb3 <- metabin(Ee, Ne, Ec, Nc, method = "I",
               data = dat.bassler2004, studlab = study)
mb3
#> Number of studies: k = 10
#> Number of observations: o = 625 (o.e = 310, o.c = 315)
#> Number of events: e = 301
#> 
#>                        RR       95%-CI     z  p-value
#> Common effect model  0.65 [0.55; 0.78] -4.81 < 0.0001
#> Random effects model 0.60 [0.46; 0.79] -3.64   0.0003
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 0.0915 [0.0000; 0.6662]; tau = 0.3025 [0.0000; 0.8162]
#>  I^2 = 52.3% [2.2%; 76.7%]; H = 1.45 [1.01; 2.07]
#> 
#> Test of heterogeneity:
#>      Q d.f. p-value
#>  18.87    9  0.0263
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q

### Conduct subgroup analysis comparing trials with adequate blinding
### to trials with inadequate or unclear blinding
mb3s <- update(mb3, subgroup = blind, print.subgroup.name = FALSE)
mb3s
#> Number of studies: k = 10
#> Number of observations: o = 625 (o.e = 310, o.c = 315)
#> Number of events: e = 301
#> 
#>                        RR       95%-CI     z  p-value
#> Common effect model  0.65 [0.55; 0.78] -4.81 < 0.0001
#> Random effects model 0.60 [0.46; 0.79] -3.64   0.0003
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 0.0915 [0.0000; 0.6662]; tau = 0.3025 [0.0000; 0.8162]
#>  I^2 = 52.3% [2.2%; 76.7%]; H = 1.45 [1.01; 2.07]
#> 
#> Test of heterogeneity:
#>      Q d.f. p-value
#>  18.87    9  0.0263
#> 
#> Results for subgroups (common effect model):
#>                     k   RR       95%-CI     Q   I^2
#> Adequate blinding   3 0.77 [0.58; 1.01]  2.49 19.7%
#> Method unclear      7 0.59 [0.47; 0.74] 14.29 58.0%
#> 
#> Test for subgroup differences (common effect model):
#>                    Q d.f. p-value
#> Between groups  2.09    1  0.1483
#> Within groups  16.79    8  0.0324
#> 
#> Results for subgroups (random effects model):
#>                     k   RR       95%-CI  tau^2    tau
#> Adequate blinding   3 0.75 [0.53; 1.08] 0.0237 0.1541
#> Method unclear      7 0.56 [0.39; 0.79] 0.1282 0.3580
#> 
#> Test for subgroup differences (random effects model):
#>                   Q d.f. p-value
#> Between groups 1.40    1  0.2367
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q

### Conduct subgroup analysis assuming common between-study variance in subgroups
mb3s.c <- update(mb3s, tau.common = TRUE)
mb3s.c
#> Number of studies: k = 10
#> Number of observations: o = 625 (o.e = 310, o.c = 315)
#> Number of events: e = 301
#> 
#>                        RR       95%-CI     z  p-value
#> Common effect model  0.65 [0.55; 0.78] -4.81 < 0.0001
#> Random effects model 0.60 [0.46; 0.79] -3.64   0.0003
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 0.0915 [0.0000; 0.6662]; tau = 0.3025 [0.0000; 0.8162]
#>  I^2 = 52.3% [2.2%; 76.7%]; H = 1.45 [1.01; 2.07]
#> 
#> Quantifying residual heterogeneity (with 95%-CIs):
#>  tau^2 = 0.1028; tau = 0.3207; I^2 = 52.3% [0.0%; 77.6%]; H = 1.45 [1.00; 2.11]
#> 
#> Test of heterogeneity:
#>      Q d.f. p-value
#>  18.87    9  0.0263
#> 
#> Results for subgroups (common effect model):
#>                     k   RR       95%-CI     Q   I^2
#> Adequate blinding   3 0.77 [0.58; 1.01]  2.49 19.7%
#> Method unclear      7 0.59 [0.47; 0.74] 14.29 58.0%
#> 
#> Test for subgroup differences (common effect model):
#>                    Q d.f. p-value
#> Between groups  2.09    1  0.1483
#> Within groups  16.79    8  0.0324
#> 
#> Results for subgroups (random effects model):
#>                     k   RR       95%-CI  tau^2    tau
#> Adequate blinding   3 0.72 [0.43; 1.21] 0.1028 0.3207
#> Method unclear      7 0.56 [0.40; 0.78] 0.1028 0.3207
#> 
#> Test for subgroup differences (random effects model):
#>                    Q d.f. p-value
#> Between groups  0.64    1  0.4245
#> Within groups  16.79    8  0.0324
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#>   (assuming common tau^2 in subgroups)
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q

### Use previous settings
settings.meta(oldset)