dat.bcg {metadat}	R Documentation

Studies on the Effectiveness of the BCG Vaccine Against Tuberculosis

Description

Results from 13 studies examining the effectiveness of the Bacillus Calmette-Guerin (BCG) vaccine against tuberculosis.

Usage


dat.bcg

Format

The data frame contains the following columns:

trial 	 numeric 	 trial number

author 	 character 	 author(s)

year 	 numeric 	 publication year

tpos 	 numeric 	 number of TB positive cases in the treated (vaccinated) group

tneg 	 numeric 	 number of TB negative cases in the treated (vaccinated) group

cpos 	 numeric 	 number of TB positive cases in the control (non-vaccinated) group

cneg 	 numeric 	 number of TB negative cases in the control (non-vaccinated) group

ablat 	 numeric 	 absolute latitude of the study location (in degrees)

alloc 	 character 	 method of treatment allocation (random, alternate, or systematic assignment)

Details

The 13 studies provide data in terms of \(2 \times 2\) tables in the form:

TB positive 	 TB negative

vaccinated group 	 tpos 	 tneg

control group 	 cpos 	 cneg

The goal of the meta-analysis was to examine the overall effectiveness of the BCG vaccine for preventing tuberculosis and to examine moderators that may potentially influence the size of the effect.

The dataset has been used in several publications to illustrate meta-analytic methods (see ‘References’).

Concepts

medicine, risk ratios, meta-regression

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Colditz, G. A., Brewer, T. F., Berkey, C. S., Wilson, M. E., Burdick, E., Fineberg, H. V., & Mosteller, F. (1994). Efficacy of BCG vaccine in the prevention of tuberculosis: Meta-analysis of the published literature. Journal of the American Medical Association, 271(9), 698–702. https://doi.org/10.1001/jama.1994.03510330076038

References

Berkey, C. S., Hoaglin, D. C., Mosteller, F., & Colditz, G. A. (1995). A random-effects regression model for meta-analysis. Statistics in Medicine, 14(4), 395–411. https://doi.org/10.1002/sim.4780140406

van Houwelingen, H. C., Arends, L. R., & Stijnen, T. (2002). Advanced methods in meta-analysis: Multivariate approach and meta-regression. Statistics in Medicine, 21(4), 589–624. https://doi.org/10.1002/sim.1040

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

Examples


### copy data into 'dat' and examine data
dat <- dat.bcg
dat

## Not run: 

### load metafor package
library(metafor)

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat)
dat

### random-effects model
res <- rma(yi, vi, data=dat)
res

### average risk ratio with 95% CI
predict(res, transf=exp)

### mixed-effects model with absolute latitude and publication year as moderators
res <- rma(yi, vi, mods = ~ ablat + year, data=dat)
res

### predicted average risk ratios for 10-60 degrees absolute latitude
### holding the publication year constant at 1970
predict(res, newmods=cbind(seq(from=10, to=60, by=10), 1970), transf=exp)

### note: the interpretation of the results is difficult because absolute
### latitude and publication year are strongly correlated (the more recent
### studies were conducted closer to the equator)
plot(ablat ~ year, data=dat, pch=19, xlab="Publication Year", ylab="Absolute Lattitude")
cor(dat$ablat, dat$year)


## End(Not run)

[Package metadat version 1.2-0 Index]