Studies on Treatments for Periodontal Disease

Results from 5 trials comparing surgical and non-surgical treatments for medium-severity periodontal disease one year after treatment.

dat.berkey1998

Format

The data frame contains the following columns:

trial	numeric	trial number

author	character	study author(s)

year	numeric	publication year

ni	numeric	number of patients

outcome	character	outcome (PD = probing depth; AL = attachment level)

yi	numeric	observed mean difference in outcome (surgical versus non-surgical)

vi	numeric	corresponding sampling variance

v1i	numeric	variances and covariances of the observed effects

v2i	numeric	variances and covariances of the observed effects

Details

The dataset includes the results from 5 trials that compared surgical and non-surgical methods for the treatment of medium-severity periodontal disease. Reported outcomes include the change in probing depth (PD) and attachment level (AL) one year after the treatment. The outcome measure used for this meta-analysis was the (raw) mean difference, calculated in such a way that positive values indicate that surgery was more effective than non-surgical treatment in decreasing the probing depth and increasing the attachment level (so, the results from the various trials indicate that surgery is preferable for reducing the probing depth, while non-surgical treatment is preferable for increasing the attachment level). Since each trial provides effect size estimates for both outcomes, the estimates are correlated. A multivariate model can be used to meta-analyze the two outcomes simultaneously.

The v1i and v2i values are the variances and covariances of the observed effects. In particular, for each study, variables v1i and v2i form a \(2 \times 2\) variance-covariance matrix of the observed effects, with the diagonal elements corresponding to the sampling variances of the mean differences (the first for probing depth, the second for attachment level) and the off-diagonal value corresponding to the covariance of the two mean differences. Below, the full (block diagonal) variance-covariance for all studies is constructed from these two variables.

Source

Berkey, C. S., Antczak-Bouckoms, A., Hoaglin, D. C., Mosteller, F., & Pihlstrom, B. L. (1995). Multiple-outcomes meta-analysis of treatments for periodontal disease. Journal of Dental Research, 74(4), 1030–1039. https://doi.org/10.1177/00220345950740040201

Berkey, C. S., Hoaglin, D. C., Antczak-Bouckoms, A., Mosteller, F., & Colditz, G. A. (1998). Meta-analysis of multiple outcomes by regression with random effects. Statistics in Medicine, 17(22), 2537–2550. https://doi.org/10.1002/(sici)1097-0258(19981130)17:22<2537::aid-sim953>3.0.co;2-c

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, dentistry, raw mean differences, multivariate models

Examples

### copy data into 'dat' and examine data
dat <- dat.berkey1998
dat
#> 
#>    trial           author year ni outcome      yi     vi    v1i    v2i 
#> 1      1 Pihlstrom et al. 1983 14      PD  0.4700 0.0075 0.0075 0.0030 
#> 2      1 Pihlstrom et al. 1983 14      AL -0.3200 0.0077 0.0030 0.0077 
#> 3      2    Lindhe et al. 1982 15      PD  0.2000 0.0057 0.0057 0.0009 
#> 4      2    Lindhe et al. 1982 15      AL -0.6000 0.0008 0.0009 0.0008 
#> 5      3   Knowles et al. 1979 78      PD  0.4000 0.0021 0.0021 0.0007 
#> 6      3   Knowles et al. 1979 78      AL -0.1200 0.0014 0.0007 0.0014 
#> 7      4  Ramfjord et al. 1987 89      PD  0.2600 0.0029 0.0029 0.0009 
#> 8      4  Ramfjord et al. 1987 89      AL -0.3100 0.0015 0.0009 0.0015 
#> 9      5    Becker et al. 1988 16      PD  0.5600 0.0148 0.0148 0.0072 
#> 10     5    Becker et al. 1988 16      AL -0.3900 0.0304 0.0072 0.0304 
#> 

### load metafor package
library(metafor)

### construct block diagonal var-cov matrix of the observed outcomes based on variables v1i and v2i
V <- vcalc(vi=1, cluster=author, rvars=c(v1i, v2i), data=dat)

### fit multiple outcomes (meta-regression) model (with REML estimation)
res <- rma.mv(yi, V, mods = ~ outcome - 1, random = ~ outcome | trial, struct="UN", data=dat)
print(res, digits=3)
#> 
#> Multivariate Meta-Analysis Model (k = 10; method: REML)
#> 
#> Variance Components:
#> 
#> outer factor: trial   (nlvls = 5)
#> inner factor: outcome (nlvls = 2)
#> 
#>            estim   sqrt  k.lvl  fixed  level 
#> tau^2.1    0.033  0.181      5     no     AL 
#> tau^2.2    0.012  0.108      5     no     PD 
#> 
#>     rho.AL  rho.PD    AL  PD 
#> AL       1             -   5 
#> PD   0.609       1    no   - 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 8) = 128.227, p-val < .001
#> 
#> Test of Moderators (coefficients 1:2):
#> QM(df = 2) = 108.862, p-val < .001
#> 
#> Model Results:
#> 
#>            estimate     se    zval   pval   ci.lb   ci.ub      
#> outcomeAL    -0.339  0.088  -3.859  <.001  -0.512  -0.167  *** 
#> outcomePD     0.353  0.059   6.006  <.001   0.238   0.469  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### test/estimate difference between the two outcomes
anova(res, X=c(1,-1))
#> 
#> Hypothesis:                             
#> 1: outcomeAL - outcomePD = 0 
#> 
#> Results:
#>    estimate     se    zval   pval 
#> 1:  -0.6926 0.0744 -9.3120 <.0001 
#> 
#> Test of Hypothesis:
#> QM(df = 1) = 86.7140, p-val < .0001
#> 

### fit model including publication year as moderator for both outcomes (with ML estimation)
res <- rma.mv(yi, V, mods = ~ outcome + outcome:I(year - 1983) - 1,
              random = ~ outcome | trial, struct="UN", data=dat, method="ML")
print(res, digits=3)
#> 
#> Multivariate Meta-Analysis Model (k = 10; method: ML)
#> 
#> Variance Components:
#> 
#> outer factor: trial   (nlvls = 5)
#> inner factor: outcome (nlvls = 2)
#> 
#>            estim   sqrt  k.lvl  fixed  level 
#> tau^2.1    0.025  0.158      5     no     AL 
#> tau^2.2    0.008  0.090      5     no     PD 
#> 
#>     rho.AL  rho.PD    AL  PD 
#> AL       1             -   5 
#> PD   0.659       1    no   - 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 6) = 125.756, p-val < .001
#> 
#> Test of Moderators (coefficients 1:4):
#> QM(df = 4) = 143.430, p-val < .001
#> 
#> Model Results:
#> 
#>                           estimate     se    zval   pval   ci.lb   ci.ub      
#> outcomeAL                   -0.335  0.079  -4.261  <.001  -0.489  -0.181  *** 
#> outcomePD                    0.348  0.052   6.694  <.001   0.246   0.450  *** 
#> outcomeAL:I(year - 1983)    -0.011  0.024  -0.445  0.656  -0.059   0.037      
#> outcomePD:I(year - 1983)     0.001  0.015   0.063  0.950  -0.029   0.031      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>