Studies on Gender Differences in Grant and Fellowship Awards

Results from 21 studies on gender differences in grant and fellowship awards.

dat.bornmann2007

Format

The data frame contains the following columns:

study	character	study reference

obs	numeric	observation within study

doctype	character	document type

gender	character	gender of the study authors

year	numeric	(average) cohort year

org	character	funding organization / program

country	character	country of the funding organization / program

type	character	fellowship or grant application

discipline	character	discipline / field

waward	numeric	number of women who received a grant/fellowship award

wtotal	numeric	number of women who applied for an award

maward	numeric	number of men who received a grant/fellowship award

mtotal	numeric	number of men who applied for an award

Details

The studies in this dataset examine whether the chances of receiving a grant or fellowship award differs for men and women. Note that many studies provide multiple comparisons (e.g., for different years / cohorts / disciplines). A multilevel meta-analysis model can be used to account for the multilevel structure in these data.

Source

Bornmann, L., Mutz, R., & Daniel, H. (2007). Gender differences in grant peer review: A meta-analysis. Journal of Informetrics, 1(3), 226--238. https://doi.org/10.1016/j.joi.2007.03.001

References

Marsh, H. W., Bornmann, L., Mutz, R., Daniel, H.-D., & O'Mara, A. (2009). Gender effects in the peer reviews of grant proposals: A comprehensive meta-analysis comparing traditional and multilevel approaches. Review of Educational Research, 79(3), 1290--1326. https://doi.org/10.3102/0034654309334143

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Examples

### copy data into 'dat' and examine data
dat <- dat.bornmann2007
head(dat, 16)

#>                  study obs doctype gender   year  org   country       type
#> 1        Ackers (2000)   1    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 2        Ackers (2000)   2    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 3        Ackers (2000)   3    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 4        Ackers (2000)   4    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 5        Ackers (2000)   5    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 6        Ackers (2000)   6    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 7        Ackers (2000)   7    Grey    M&F 1996.0 MSCA    Europe Fellowship
#> 8  Allmendinger (2002)   1 Article    M&F 1993.0  DFG    Europe      Grant
#> 9  Allmendinger (2002)   2 Article    M&F 1994.0  DFG    Europe      Grant
#> 10 Allmendinger (2002)   3 Article    M&F 1995.0  DFG    Europe      Grant
#> 11 Allmendinger (2002)   4 Article    M&F 1996.0  DFG    Europe      Grant
#> 12 Allmendinger (2002)   5 Article    M&F 1997.0  DFG    Europe      Grant
#> 13 Allmendinger (2002)   6 Article    M&F 1998.0  DFG    Europe      Grant
#> 14 Allmendinger (2002)   7 Article    M&F 1999.0  DFG    Europe      Grant
#> 15      Bazeley (1998)   1 Article      F 1995.0  ARC Australia      Grant
#> 16     Bornmann (2005)   1 Article      M 1992.5  BIF    Europe Fellowship
#>                   discipline waward wtotal maward mtotal
#> 1          Physical Sciences    139    711    274   1029
#> 2          Physical Sciences     45    258    166    908
#> 3          Physical Sciences     44    236    219    928
#> 4          Physical Sciences     63    251     96    507
#> 5  Social Sciences / Biology    157    910    252   1118
#> 6          Physical Sciences    114    589    460   2244
#> 7    Life Sciences / Biology    381   2027    489   2275
#> 8  Social Sciences / Biology      8     13     53     72
#> 9  Social Sciences / Biology      5      8     53     82
#> 10 Social Sciences / Biology      6      8     63     97
#> 11 Social Sciences / Biology      8     16     53     94
#> 12 Social Sciences / Biology      4     11     43     92
#> 13 Social Sciences / Biology     20     44     55     93
#> 14 Social Sciences / Biology      5     15     70    116
#> 15         Multidisciplinary     11     56     82    344
#> 16   Life Sciences / Biology    204   1085    430   1612

# \dontrun{

### load metafor package
require(metafor)

### calculate log odds ratios and corresponding sampling variances
dat <- escalc(measure="OR", ai=waward, n1i=wtotal, ci=maward, n2i=mtotal, data=dat)

### fit multilevel meta-analysis model
res <- rma.mv(yi, vi, random = ~ 1 | study/obs, data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 66; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed     factor 
#> sigma^2.1  0.0161  0.1268     21     no      study 
#> sigma^2.2  0.0038  0.0613     66     no  study/obs 
#> 
#> Test for Heterogeneity:
#> Q(df = 65) = 221.2850, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb    ci.ub 
#>  -0.1010  0.0417  -2.4196  0.0155  -0.1828  -0.0192  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average odds ratio (with 95
predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.90  0.83  0.98  0.68  1.21 
#> 

### test for a difference between fellowship and grant applications
res <- rma.mv(yi, vi, mods = ~ type, random = ~ 1 | study/obs, data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 66; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed     factor 
#> sigma^2.1  0.0045  0.0670     21     no      study 
#> sigma^2.2  0.0035  0.0596     66     no  study/obs 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 64) = 133.4811, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 11.2312, p-val = 0.0008
#> 
#> Model Results:
#> 
#>            estimate      se     zval    pval    ci.lb    ci.ub 
#> intrcpt     -0.2010  0.0429  -4.6816  <.0001  -0.2852  -0.1169  *** 
#> typeGrant    0.1890  0.0564   3.3513  0.0008   0.0785   0.2995  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, newmods=0:1, transf=exp, digits=2)

#> 
#>   pred ci.lb ci.ub pi.lb pi.ub 
#> 1 0.82  0.75  0.89  0.67  0.99 
#> 2 0.99  0.92  1.06  0.82  1.20 
#> 

# }