Studies on the Association between Handedness and Eye-Dominance

Results from 47 studies on the association between handedness and eye-dominance.

dat.bourassa1996

Format

The data frame contains the following columns:

study	numeric	study number

sample	numeric	sample number

author	character	(first) author

year	numeric	publication year

selection	character	selection of subjects on the basis of eyedness or handedness

investigator	character	investigator (psychologist, educationalist, or other)

hand_assess	character	method to assess handedness (questionnaire or performance based)

eye_assess	character	method to assess eyedness (see ‘Details’)

mage	numeric	mean age of sample

lh.le	numeric	number of left-handed left-eyed individuals

lh.re	numeric	number of left-handed right-eyed individuals

rh.le	numeric	number of right-handed left-eyed individuals

rh.re	numeric	number of right-handed right-eyed individuals

sex	character	sex of the sample (combined, male, or female)

Details

The 47 studies included in this meta-analysis examined the association between handedness and eye-dominance (ocular dominance or eyedness). Results are given in terms of \(2 \times 2\) tables, indicating the number of left-handed left-eyed, left-handed right-eyed, right-handed left-eyed, and right-handed right-eyed individuals. Note that some studies included multiple (independent) samples, so that the meta-analysis included 54 samples in total. Also, for some studies, the combined data of the males and females are further broken down into the two subgroups.

In some studies, there was indication that the selection of subjects was not random with respect to handedness and/or eyedness. While this should not influence the size of the association as measured with the odds ratio, this invalidates those studies for assessing the overall percentage of left-eyed and left-handed individuals.

Handedness was assessed in the individual studies either based on a questionnaire or inventory or based on task performance. Eyedness was assessed based on various methods: E.1 methods are based on task performance, while E.2.a denotes assessment based on a questionnaire. The performance based methods could be further broken down into: E.1.a.i (monocular procedure with object/instrument held in one hand), E.1.a.ii (monocular procedure with object/instrument held in both hands), E.1.b (binocular procedure), E.1.c (a combination of the previous methods), and E.1.d (some other method).

Source

Bourassa, D. C., McManus, I. C., & Bryden, M. P. (1996). Handedness and eye-dominance: A meta-analysis of their relationship. Laterality, 1(1), 5--34. https://doi.org/10.1080/713754206

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Examples

### copy data into 'dat' and examine data
dat <- dat.bourassa1996
head(dat, 10)

#>    study sample    author year selection investigator   hand_assess eye_assess mage lh.le lh.re rh.le
#> 1      1      1     Mills 1925        no        other questionnaire      E.1.d   NA    93    17   130
#> 2      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37   140    91   305
#> 3      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    74    57   158
#> 4      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    66    34   147
#> 5      3      3     Miles 1930        no psychologist questionnaire   E.1.a.ii   22    16    14    43
#> 6      4      4    Quinan 1931        no        other   performance      E.1.b   25   102    97   597
#> 7      5      5    Jasper 1932       yes psychologist questionnaire   E.1.a.ii   20    17    14    38
#> 8      6      6      Lund 1932        no psychologist   performance   E.1.a.ii   20    10     2    52
#> 9      7      7      Eyre 1933        no psychologist   performance   E.1.a.ii   15     7     2    15
#> 10     8      8 Updegraff 1933        no psychologist   performance   E.1.a.ii    4     4     2    15
#>    rh.re      sex
#> 1    760 combined
#> 2    697 combined
#> 3    427     male
#> 4    270   female
#> 5    114 combined
#> 6   1898 combined
#> 7     80 combined
#> 8    170 combined
#> 9    169 combined
#> 10    38 combined

# \dontrun{

### load metafor package
require(metafor)

### calculate log(OR) and corresponding sampling variance with 1/2 correction
dat <- escalc(measure="OR", ai=lh.le, bi=lh.re, ci=rh.le, di=rh.re, data=dat, add=1/2, to="all")
head(dat, 10)

#>    study sample    author year selection investigator   hand_assess eye_assess mage lh.le lh.re rh.le 
#> 1      1      1     Mills 1925        no        other questionnaire      E.1.d   NA    93    17   130 
#> 2      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37   140    91   305 
#> 3      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    74    57   158 
#> 4      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    66    34   147 
#> 5      3      3     Miles 1930        no psychologist questionnaire   E.1.a.ii   22    16    14    43 
#> 6      4      4    Quinan 1931        no        other   performance      E.1.b   25   102    97   597 
#> 7      5      5    Jasper 1932       yes psychologist questionnaire   E.1.a.ii   20    17    14    38 
#> 8      6      6      Lund 1932        no psychologist   performance   E.1.a.ii   20    10     2    52 
#> 9      7      7      Eyre 1933        no psychologist   performance   E.1.a.ii   15     7     2    15 
#> 10     8      8 Updegraff 1933        no psychologist   performance   E.1.a.ii    4     4     2    15 
#>    rh.re      sex     yi     vi 
#> 1    760 combined 3.4384 0.0768 
#> 2    697 combined 1.2544 0.0228 
#> 3    427     male 1.2512 0.0395 
#> 4    270   female 1.2627 0.0545 
#> 5    114 combined 1.0970 0.1613 
#> 6   1898 combined 1.2061 0.0222 
#> 7     80 combined 0.9257 0.1645 
#> 8    170 combined 2.6130 0.5202 
#> 9    169 combined 3.4906 0.6037 
#> 10    38 combined 1.4976 0.7127 

### overall association between handedness and eyedness
res <- rma(yi, vi, data=dat, subset=sex=="combined")
res

#> 
#> Random-Effects Model (k = 54; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.4340 (SE = 0.1208)
#> tau (square root of estimated tau^2 value):      0.6588
#> I^2 (total heterogeneity / total variability):   89.08%
#> H^2 (total variability / sampling variability):  9.16
#> 
#> Test for Heterogeneity:
#> Q(df = 53) = 355.7748, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub 
#>   1.3135  0.1098  11.9597  <.0001  1.0982  1.5287  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  3.72  3.00  4.61  1.00 13.77 
#> 

# }