dat.cannon2006 {metadat}	R Documentation

Studies on the Effectiveness of Intensive Versus Moderate Statin Therapy for Preventing Coronary Death or Myocardial Infarction

Description

Results from 4 trials examining the effectiveness of intensive (high dose) versus moderate (standard dose) statin therapy for preventing coronary death or myocardial infarction.

Usage


dat.cannon2006

Format

The data frame contains the following columns:

trial 	 character 	 trial name

pop 	 character 	 study population (post-ACS: post acute coronary syndrome; stable CAD: stable coronary artery disease)

nt 	 numeric 	 number of patients in the high dose group

nc 	 numeric 	 number of patients in the standard dose group

ep1t 	 numeric 	 number of events in the high dose group for end point 1: coronary death or non-fatal myocardial infarction

ep1c 	 numeric 	 number of events in the standard dose group for end point 1: coronary death or non-fatal myocardial infarction

ep2t 	 numeric 	 number of events in the high dose group for end point 2: coronary death or any cardiovascular event (MI, stroke, hospitalization for unstable angina, or revascularization)

ep2c 	 numeric 	 number of events in the standard dose group for end point 2: coronary death or any cardiovascular event (MI, stroke, hospitalization for unstable angina, or revascularization)

ep3t 	 numeric 	 number of events in the high dose group for end point 3: cardiovascular death

ep3c 	 numeric 	 number of events in the standard dose group for end point 3: cardiovascular death

ep4t 	 numeric 	 number of events in the high dose group for end point 4: non-cardiovascular death

ep4c 	 numeric 	 number of events in the standard dose group for end point 4: non-cardiovascular death

ep5t 	 numeric 	 number of events in the high dose group for end point 5: deaths (all-cause mortality)

ep5c 	 numeric 	 number of events in the standard dose group for end point 5: deaths (all-cause mortality)

ep6t 	 numeric 	 number of events in the high dose group for end point 6: stroke

ep6c 	 numeric 	 number of events in the standard dose group for end point 6: stroke

Details

The data were obtained from Figures 2, 3, 4, and 5 in Cannon et al. (2006). The authors used the Mantel-Haenszel method for combining the results from the 4 trials. This approach is implemented in the rma.mh function.

Concepts

medicine, cardiology, odds ratios, Mantel-Haenszel method

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Cannon, C. P., Steinberg, B. A., Murphy, S. A., Mega, J. L., & Braunwald, E. (2006). Meta-analysis of cardiovascular outcomes trials comparing intensive versus moderate statin therapy. Journal of the American College of Cardiology, 48(3), 438–445. https://doi.org/10.1016/j.jacc.2006.04.070

Examples


### copy data into 'dat' and examine data
dat <- dat.cannon2006
dat

## Not run: 

### load metafor package
library(metafor)

### meta-analysis of log odds ratios using the MH method for endpoint 1
res <- rma.mh(measure="OR", ai=ep1t, n1i=nt, ci=ep1c, n2i=nc, data=dat, slab=trial)
print(res, digits=2)

### forest plot
forest(res, xlim=c(-.8,.8), atransf=exp, at=log(c(2/3, 1, 3/2)),
       header=TRUE, top=2, cex=1.2, xlab="Odds Ratio")
mtext("(high dose better)", side=1, line=par("mgp")[1]-0.5, at=log(2/3), cex=1.2, font=3)
mtext("(standard dose better)", side=1, line=par("mgp")[1]-0.5, at=log(3/2), cex=1.2, font=3)


## End(Not run)

[Package metadat version 1.2-0 Index]