Studies on the Effects of Diuretics in Pregnancy

Results from 9 studies examining the effects of diuretics in pregnancy on various outcomes.

dat.collins1985b

Format

The data frame contains the following columns:

id	numeric	study number

author	character	study author(s)

year	numeric	publication year

pre.nti	numeric	number of women in treatment group followed up for pre-eclampsia outcome

pre.nci	numeric	number of women in control/placebo group followed up for pre-eclampsia outcome

pre.xti	numeric	number of women in treatment group with any form of pre-eclampsia

pre.xci	numeric	number of women in control/placebo group with any form of pre-eclampsia

oedema	numeric	dummy variable indicating whether oedema was a diagnostic criterion

fup.nti	numeric	number of women in treatment group followed up for mortality outcomes

fup.nci	numeric	number of women in control/placebo group followed up for mortality outcomes

ped.xti	numeric	number of perinatal deaths in treatment group

ped.xci	numeric	number of perinatal deaths in control/placebo group

stb.xti	numeric	number of stillbirths in treatment group

stb.xci	numeric	number of stillbirths in control/placebo group

ned.xti	numeric	number of neonatal deaths in treatment group

ned.xci	numeric	number of neonatal deaths in control/placebo group

Details

The 9 studies in this dataset examined the effects of diuretics in pregnancy on various outcomes, including the presence of any form of pre-eclampsia, perinatal death, stillbirth, and neonatal death.

Source

Collins, R., Yusuf, S., & Peto, R. (1985). Overview of randomised trials of diuretics in pregnancy. British Medical Journal, 290(6461), 17–23. https://doi.org/10.1136/bmj.290.6461.17

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, obstetrics, odds ratios, Peto's method

Examples

### copy data into 'dat' and examine data
dat <- dat.collins1985b
dat
#>   id                  author year pre.nti pre.nci pre.xti pre.xci oedema fup.nti fup.nci ped.xti
#> 1  1       Weseley & Douglas 1962     131     136      14      14      0     131     136       1
#> 2  2          Flowers et al. 1962     385     134      21      17      0     335     110       6
#> 3  3                 Menzies 1964      57      48      14      24      1      57      48       3
#> 4  4           Fallis et al. 1964      38      40       6      18      0      34      40       1
#> 5  5         Cuadros & Tatum 1964    1011     760      12      35      1    1011     760      14
#> 6  6        Landesman et al. 1965    1370    1336     138     175      0    1370    1336      24
#> 7  7            Kraus et al. 1966     506     524      15      20      0     506     524      14
#> 8  8    Tervila & Vartiainen 1971     108     103       6       2      0     108     103       0
#> 9  9 Campbell & MacGillivray 1975     153     102      65      40      0     153     102       0
#>   ped.xci stb.xti stb.xci ned.xti ned.xci
#> 1       4       1       2       0       2
#> 2       3       3       2       3       1
#> 3       2       1       1       2       1
#> 4       3       0       1       1       2
#> 5      13       6       5       8       8
#> 6      19      NA      NA      NA      NA
#> 7      16       6       9       8       7
#> 8       0       0       0       0       0
#> 9       0       0       0       0       0

### load metafor package
library(metafor)

### calculate (log) odds ratio and sampling variance
dat <- escalc(measure="OR", n1i=pre.nti, n2i=pre.nci, ai=pre.xti, ci=pre.xci, data=dat)
summary(dat, digits=2, transf=exp)
#> 
#>   id                  author year pre.nti pre.nci pre.xti pre.xci oedema fup.nti fup.nci ped.xti 
#> 1  1       Weseley & Douglas 1962     131     136      14      14      0     131     136       1 
#> 2  2          Flowers et al. 1962     385     134      21      17      0     335     110       6 
#> 3  3                 Menzies 1964      57      48      14      24      1      57      48       3 
#> 4  4           Fallis et al. 1964      38      40       6      18      0      34      40       1 
#> 5  5         Cuadros & Tatum 1964    1011     760      12      35      1    1011     760      14 
#> 6  6        Landesman et al. 1965    1370    1336     138     175      0    1370    1336      24 
#> 7  7            Kraus et al. 1966     506     524      15      20      0     506     524      14 
#> 8  8    Tervila & Vartiainen 1971     108     103       6       2      0     108     103       0 
#> 9  9 Campbell & MacGillivray 1975     153     102      65      40      0     153     102       0 
#>   ped.xci stb.xti stb.xci ned.xti ned.xci   yi ci.lb ci.ub 
#> 1       4       1       2       0       2 1.04  0.48  2.28 
#> 2       3       3       2       3       1 0.40  0.20  0.78 
#> 3       2       1       1       2       1 0.33  0.14  0.74 
#> 4       3       0       1       1       2 0.23  0.08  0.67 
#> 5      13       6       5       8       8 0.25  0.13  0.48 
#> 6      19      NA      NA      NA      NA 0.74  0.59  0.94 
#> 7      16       6       9       8       7 0.77  0.39  1.52 
#> 8       0       0       0       0       0 2.97  0.59 15.07 
#> 9       0       0       0       0       0 1.14  0.69  1.91 
#> 

### meta-analysis using Peto's method for any form of pre-eclampsia
rma.peto(n1i=pre.nti, n2i=pre.nci, ai=pre.xti, ci=pre.xci, data=dat, digits=2)
#> 
#> Equal-Effects Model (k = 9)
#> 
#> I^2 (total heterogeneity / total variability):  72.74%
#> H^2 (total variability / sampling variability): 3.67
#> 
#> Test for Heterogeneity: 
#> Q(df = 8) = 29.34, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.41  0.09  -4.65  <.01  -0.58  -0.24 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.66   0.56   0.79 
#> 

### meta-analysis including only studies where oedema was not a diagnostic criterion
rma.peto(n1i=pre.nti, n2i=pre.nci, ai=pre.xti, ci=pre.xci, data=dat, digits=2, subset=(oedema==0))
#> 
#> Equal-Effects Model (k = 7)
#> 
#> I^2 (total heterogeneity / total variability):  60.84%
#> H^2 (total variability / sampling variability): 2.55
#> 
#> Test for Heterogeneity: 
#> Q(df = 6) = 15.32, p-val = 0.02
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.28  0.09  -2.97  <.01  -0.47  -0.10 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.76   0.63   0.91 
#> 

### meta-analyses of mortality outcomes (perinatal deaths, stillbirths, and neonatal deaths)
rma.peto(n1i=fup.nti, n2i=fup.nci, ai=ped.xti, ci=ped.xci, data=dat, digits=2)
#> Warning: Some yi/vi values are NA.
#> 
#> Equal-Effects Model (k = 9)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 0.58
#> 
#> Test for Heterogeneity: 
#> Q(df = 6) = 3.49, p-val = 0.75
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.09  0.18  -0.50  0.62  -0.45   0.27 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.91   0.64   1.31 
#> 
rma.peto(n1i=fup.nti, n2i=fup.nci, ai=stb.xti, ci=stb.xci, data=dat, digits=2)
#> Warning: 1 study with NAs omitted from model fitting.
#> Warning: Some yi/vi values are NA.
#> 
#> Equal-Effects Model (k = 8)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 0.20
#> 
#> Test for Heterogeneity: 
#> Q(df = 5) = 0.99, p-val = 0.96
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.39  0.34  -1.16  0.25  -1.05   0.27 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.68   0.35   1.31 
#> 
rma.peto(n1i=fup.nti, n2i=fup.nci, ai=ned.xti, ci=ned.xci, data=dat, digits=2)
#> Warning: 1 study with NAs omitted from model fitting.
#> Warning: Some yi/vi values are NA.
#> 
#> Equal-Effects Model (k = 8)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 0.51
#> 
#> Test for Heterogeneity: 
#> Q(df = 5) = 2.54, p-val = 0.77
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.15  0.31  -0.47  0.64  -0.76   0.47 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.86   0.47   1.59 
#>