dat.debruin2009 {metadat}	R Documentation

Studies on Standard Care Quality and HAART-Adherence

Description

Results from 13 trials providing information about standard care quality and HAART-adherence in control groups.

Usage


dat.debruin2009

Format

The data frame contains the following columns:

author 	 character 	 (first) author of study

year 	 numeric 	 publication year

scq 	 numeric 	 standard care quality

ni 	 numeric 	 number of patients in the standard care group

xi 	 numeric 	 number of patients with an undetectable viral load in standard care group

mi 	 numeric 	 number of patients with a detectable viral load in standard care group

ethnicity 	 character 	 dominant ethnicity of the patients in the standard care group

patients 	 character 	 inclusion of patients continuing or starting (a new) treatment

select 	 character 	 baseline selection of patients with adherence problems or no selection

sens 	 character 	 sensitivity of viral load assessments (<400 vs. >=400 copies/ml)

Details

Highly active antiretroviral therapy (HAART) refers to a combination of multiple antiretroviral drugs that can effectively suppress the HIV virus. However, achieving viral suppression (to the point that the virus becomes essentially undetectable in a blood sample) requires high levels of adherence to an often complicated medication regimen. A number of trials have examined various interventions that aim to increase adherence levels. In each trial, patients receiving the intervention are compared to patients in a control group receiving standard care (often referred to as ‘care as usual’). However, the quality of standard care can vary substantially between these studies. de Bruin et al. (2009) assessed the quality of standard care provided (based on a quantification of the number of behavior change techniques applied) and examined to what extent the quality of standard care was related to the proportion of patients achieving effective viral suppression in the control groups.

Concepts

psychology, medicine, proportions, single-arm studies, meta-regression

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

de Bruin, M., Viechtbauer, W., Hospers, H. J., Schaalma, H. P., & Kok, G. (2009). Standard care quality determines treatment outcomes in control groups of HAART-adherence intervention studies: Implications for the interpretation and comparison of intervention effects. Health Psychology, 28(6), 668–674. https://doi.org/10.1037/a0015989

Examples


### copy data into 'dat' and examine data
dat <- dat.debruin2009
dat

## Not run: 

### load metafor package
library(metafor)

### calculate proportions and corresponding sampling variances
dat <- escalc(measure="PR", xi=xi, ni=ni, data=dat)
dat

### random-effects model
res <- rma(yi, vi, data=dat)
print(res, digits=2)

### mixed-effects meta-regression model with all predictors/covariates
res <- rma(yi, vi, mods = ~ scq + ethnicity + patients + select + sens, data=dat)
print(res, digits=3)

### mixed-effects meta-regression model with scq and ethnicity as predictors/covariates
res <- rma(yi, vi, mods = ~ scq + ethnicity, data=dat)
print(res, digits=3)


## End(Not run)

[Package metadat version 1.2-0 Index]