dat.dorn2007 {metadat}	R Documentation

Studies on Complementary and Alternative Medicine for Irritable Bowel Syndrome

Description

Results from 19 trials examining complementary and alternative medicine (CAM) for irritable bowel syndrome (IBS).

Usage


dat.dorn2007

Format

The data frame contains the following columns:

id 	 numeric 	 trial id number

study 	 character 	 (first) author

year 	 numeric 	 publication year

country 	 character 	 country where trial was conducted

ibs.crit 	 character 	 IBS diagnostic criteria (Manning, Rome I, Rome II, or Other)

days 	 numeric 	 number of treatment days

visits 	 numeric 	 number of practitioner visits

jada 	 numeric 	 Jadad score

x.a 	 numeric 	 number of responders in the active treatment group

n.a 	 numeric 	 number of participants in the active treatment group

x.p 	 numeric 	 number of responders in the placebo group

n.p 	 numeric 	 number of participants in the placebo group

Details

The dataset includes the results from 19 randomized clinical trials that examined the effectiveness of complementary and alternative medicine (CAM) for irritable bowel syndrome (IBS).

Concepts

medicine, alternative medicine, risk ratios

Note

The data were extracted from Table I in Dorn et al. (2009). Comparing the funnel plot in Figure 1 with the one obtained below indicates that the data for study 5 (Davis et al., 2006) in the table were not the ones that were used in the actual analyses.

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Dorn, S. D., Kaptchuk, T. J., Park, J. B., Nguyen, L. T., Canenguez, K., Nam, B. H., Woods, K. B., Conboy, L. A., Stason, W. B., & Lembo, A. J. (2007). A meta-analysis of the placebo response in complementary and alternative medicine trials of irritable bowel syndrome. Neurogastroenterology & Motility, 19(8), 630–637. https://doi.org/10.1111/j.1365-2982.2007.00937.x

Examples


### copy data into 'dat' and examine data
dat <- dat.dorn2007
dat

## Not run: 

### load metafor package
library(metafor)

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=x.a, n1i=n.a, ci=x.p, n2i=n.p, data=dat)

### random-effects model
res <- rma(yi, vi, data=dat, digits=2, method="DL")
res

### estimated average risk ratio
predict(res, transf=exp)

### funnel plot with study 5 highlighted in red
funnel(res, atransf=exp, at=log(c(.1, .2, .5, 1, 2, 5, 10)),
       ylim=c(0,1), steps=6, las=1, col=ifelse(id == 5, "red", "black"))

### change log risk ratio for study 5
dat$yi[5] <- -0.44

### results are now more in line with what is reported in the paper
### (although the CI in the paper is not wide enough)
res <- rma(yi, vi, data=dat, digits=2, method="DL")
predict(res, transf=exp)

### funnel plot with study 5 highlighted in red
funnel(res, atransf=exp, at=log(c(.1, .2, .5, 1, 2, 5, 10)),
       ylim=c(0,1), steps=6, las=1, col=ifelse(id == 5, "red", "black"))


## End(Not run)

[Package metadat version 1.2-0 Index]