dat.frank2008 {metadat}	R Documentation

Studies on the Association Between the CASP8 -652 6N del Promoter Polymorphism and Breast Cancer Risk

Description

Results from 4 case-control studies examining the association between the CASP8 -652 6N del promoter polymorphism and breast cancer risk.

Usage


dat.frank2008

Format

The data frame contains the following columns:

study 	 character 	 study identifier

bc.ins.ins 	 numeric 	 number of cases who have a homozygous insertion polymorphism

bc.ins.del 	 numeric 	 number of cases who have a heterozygous insertion/deletion polymorphism

bc.del.del 	 numeric 	 number of cases who have a homozygous deletion polymorphism

ct.ins.ins 	 numeric 	 number of controls who have a homozygous insertion polymorphism

ct.ins.del 	 numeric 	 number of controls who are heterozygous insertion/deletion polymorphism

ct.del.del 	 numeric 	 number of controls who have a homozygous deletion polymorphism

Details

The 4 studies included in this dataset are case-control studies that have examined the association between the CASP8 -652 6N del promoter polymorphism and breast cancer risk. Breast cancer cases and controls were genotyped and either had a homozygous insertion, a heterozygous insertion/deletion, or a homozygous deletion polymorphism.

Ziegler et al. (2011) used the same dataset to illustrate the use of meta-analytic methods to examine deviations from Hardy-Weinberg equilibrium across multiple studies. The relative excess heterozygosity (REH) is the proposed measure for such a meta-analysis, which can be computed by setting measure="REH".

Concepts

medicine, oncology, genetics, odds ratios

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Frank, B., Rigas, S. H., Bermejo, J. L., Wiestler, M., Wagner, K., Hemminki, K., Reed, M. W., Sutter, C., Wappenschmidt, B., Balasubramanian, S. P., Meindl, A., Kiechle, M., Bugert, P., Schmutzler, R. K., Bartram, C. R., Justenhoven, C., Ko, Y.-D., Brüning, T., Brauch, H., Hamann, U., Pharoah, P. P. D., Dunning, A. M., Pooley, K. A., Easton, D. F., Cox, A. & Burwinkel, B. (2008). The CASP8 -652 6N del promoter polymorphism and breast cancer risk: A multicenter study. Breast Cancer Research and Treatment, 111(1), 139-144. https://doi.org/10.1007/s10549-007-9752-z

References

Ziegler, A., Steen, K. V. & Wellek, S. (2011). Investigating Hardy-Weinberg equilibrium in case-control or cohort studies or meta-analysis. Breast Cancer Research and Treatment, 128(1), 197–201. https://doi.org/10.1007/s10549-010-1295-z

Examples


### copy data into 'dat' and examine data
dat <- dat.frank2008
dat

## Not run: 

### load metafor package
library(metafor)

### calculate log odds ratios comparing ins/del versus ins/ins
dat <- escalc(measure="OR", ai=bc.ins.del, bi=bc.ins.ins,
                            ci=ct.ins.del, di=ct.ins.ins, data=dat)

### fit random-effects model and get the pooled odds ratio (with 95% CI)
res <- rma(yi, vi, data=dat)
res
predict(res, transf=exp, digits=2)

### calculate log odds ratios comparing del/del versus ins/ins
dat <- escalc(measure="OR", ai=bc.del.del, bi=bc.ins.ins,
                            ci=ct.del.del, di=ct.ins.ins, data=dat)

### fit random-effects model and get the pooled odds ratio (with 95% CI)
res <- rma(yi, vi, data=dat)
res
predict(res, transf=exp, digits=2)

### calculate log odds ratios comparing ins/del+del/del versus ins/ins
dat <- escalc(measure="OR", ai=bc.ins.del+bc.del.del, bi=bc.ins.ins,
                            ci=ct.ins.del+ct.del.del, di=ct.ins.ins, data=dat)

### fit random-effects model and get the pooled odds ratio (with 95% CI)
res <- rma(yi, vi, data=dat)
res
predict(res, transf=exp, digits=2)

############################################################################

### compute the relative excess heterozygosity in the controls
dat <- escalc(measure="REH", ai=ct.ins.ins, bi=ct.ins.del, ci=ct.del.del,
              slab=study, data=dat)

### fit random-effects model and get the pooled REH value (with 90% CI)
res <- rma(yi, vi, data=dat, level=90)
res
predict(res, transf=exp, digits=2)

### draw forest plot
forest(res, atransf=exp, header=TRUE, xlim=c(-1.5,1.5), at=log(c(0.5,5/7,1,7/5,2)))
segments(log(5/7), -2, log(5/7), res$k+1, lty="dotted")
segments(log(7/5), -2, log(7/5), res$k+1, lty="dotted")


## End(Not run)

[Package metadat version 1.2-0 Index]