dat.hahn2001 {metadat}	R Documentation

Studies on the Effectiveness of Different Rehydration Solutions for the Prevention of Unscheduled Intravenous Infusion in Children with Diarrhoea

Description

Results from 12 trials examining the effectiveness of a reduced versus standard rehydration solution for the prevention of unscheduled intravenous infusion in children with diarrhoea.

Usage


dat.hahn2001

Format

The data frame contains the following columns:

study 	 character 	 trial name and year

ai 	 numeric 	 number of children requiring unscheduled intravenous infusion in the reduced rehydration solution group

n1i 	 numeric 	 number of children in the reduced rehydration solution group

ci 	 numeric 	 number of children requiring unscheduled intravenous infusion in the standard rehydration solution group

n2i 	 numeric 	 number of children in the standard rehydration solution group

Details

The dataset includes the results from 12 randomized clinical trials that examined the effectiveness of a reduced osmolarity oral rehydration solution (total osmolarity <250 mmol/l with reduced sodium) with a standard WHO oral rehydration solution (sodium 90 mmol/l, glucose 111mmol/l, total osmolarity 311 mmol/l) for the prevention of unscheduled intravenous infusion in children with diarrhoea.

Concepts

medicine, odds ratios, Mantel-Haenszel method

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Hahn, S., Kim, Y., & Garner, P. (2001). Reduced osmolarity oral rehydration solution for treating dehydration due to diarrhoea in children: Systematic review. British Medical Journal, 323(7304), 81–85. https://doi.org/10.1136/bmj.323.7304.81

Examples


### copy data into 'dat' and examine data
dat <- dat.hahn2001
dat

## Not run: 

### load metafor package
library(metafor)

### meta-analysis of (log) odds rations using the Mantel-Haenszel method
res <- rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, digits=2, slab=study)
res

### forest plot (also show studies that were excluded from the analysis)
options(na.action="na.pass")
forest(res, atransf=exp, at=log(c(.01, .1, 1, 10, 100)), header=TRUE)
options(na.action="na.omit")


## End(Not run)

[Package metadat version 1.2-0 Index]