Studies Comparing Objective and Subjective Olfactory Loss in COVID-19 Patients

Results from 35 studies measuring olfactory loss in COVID-19 patients using either objective or subjective measures.

dat.hannum2020

Format

The data frame contains the following columns:

authorName	character	(first) author of study

DOI	character	article DOI number

ni	numeric	number of Covid-19 positive patients in the study

xi	numeric	number of Covid-19 positive patients in the study with olfactory loss

percentOlfactoryLoss	numeric	percent of the sample with olfactory loss

objectivity	character	objective or subjective measure used

measured	character	outcome measure

testType	character	type of test used

country	character	country where patients were treated

patientType	character	type of patient information and location where being treated

Details

One of the symptoms of COVID-19 infection is olfactory loss (loss of smell) either recently acquired anosmia (complete loss of smell) or hyposmia (partial loss of smell). One challenge to reaching this symptom is the wide range of reported prevalence for this symptom ranging from 5 percent to 98 percent. In this dataset studies were grouped into one of two groups based on the type of method used to measure smell loss (either subjective measures, such as self-reported smell loss, or objective measures using rated stimuli).

Source

Ramirez VA , Hannum ME, Lipson SJ, Herriman RD, Toskala AK, Lin C, Joseph PV, Reed DR. 2020. COVID-19 Smell Loss Prevalence Tracker. Available from: https://vicente-ramirez.shinyapps.io/COVID19_Olfactory_Dashboard/ and https://github.com/vramirez4/OlfactoryLoss (accessed August 11, 2021)

References

Hannum, M. E., Ramirez, V. A., Lipson, S. J., Herriman, R. D., Toskala, A. K., Lin, C., Joseph, P. V., & Reed, D. R. (2020). Objective sensory testing methods reveal a higher prevalence of olfactory loss in COVID-19 positive patients compared to subjective methods: A systematic review and meta-analysis. Chemical Senses, 45(9), 865–874. https://doi.org/10.1093/chemse/bjaa064

Author

W. Kyle Hamilton whamilton@ucmerced.edu https://kylehamilton.com

Concepts

medicine, covid-19, proportions

Examples

# copy data into 'dat' and examine data
dat <- dat.hannum2020
dat
#>             authorName                              DOI   ni   xi percentOlfactoryLoss
#> 1       Vaira et al. 1                10.1002/hed.26204   72   44            61.111111
#> 2       Iravani et al.      10.1101/2020.05.07.20094516   16   13            81.250000
#> 3       Vaira et al. 2                10.1002/hed.26228   33   17            51.515152
#> 4         Merza et al.        10.1016/j.dsx.2020.04.047   15    2            13.333333
#> 5       Menni et al. 1      10.1101/2020.04.05.20048421  579  344            59.412781
#> 6      Levinson et al.      10.1101/2020.04.11.20055483   42   15            35.714286
#> 7       Haehner et al.      10.1101/2020.04.27.20081356   34   22            64.705882
#> 8         Speth et al.         10.1177/0194599820929185  103   63            61.165049
#> 9      De Maria et al.               10.1002/jmv.25995.   95   48            50.526316
#> 10      Menni et al. 2        10.1038/s41591-020-0916-2 7178 4668            65.032042
#> 11        Yan et al. 1               10.1111/alr.22579.   59   40            67.796610
#> 12        Luers et al.              10.1093/cid/ciaa525   72   53            73.611111
#> 13       Roland et al.               10.1111/alr.22602.  145   95            65.517241
#> 14 Boscolo-Rizzo et al       10.1007/s00405-020-06066-9   54   34            62.962963
#> 15          Liu et al.           10.3390/ijerph17093311  321   42            13.084112
#> 16      Paderno et al.               10.1111/alr.22610.  508  284            55.905512
#> 17          Lee et al.        10.3346/jkms.2020.35.e174 3191  488            15.293012
#> 18      Lechien et al.       10.1007/s00405-020-05965-1  417  357            85.611511
#> 19      Gelardi et al.          10.23750/abm.v91i2.9524   72   42            58.333333
#> 20   Giacomelli et al.              10.1093/cid/ciaa330   59   14            23.728814
#> 21        Shoer et al. 10.1101/2020.05.18.20105569.this  498  136            27.309237
#> 22      Vaira et al. 3                10.1002/hed.26269  345  241            69.855072
#> 23  Bocksberger et al.       10.1007/s00106-020-00891-4   63   26            41.269841
#> 24                 Mao      10.1101/2020.02.22.20026500  214   11             5.140187
#> 25             Spinato           10.1001/jama.2020.6771  202  130            64.356436
#> 26  Beltran-Corbellini       10.1007/s00405-020-05999-5   79   25            31.645570
#> 27     Trubiano et al.              10.1093/cid/ciaa655   28    7            25.000000
#> 28        Moein et al.               10.1111/alr.22587.   60   59            98.333333
#> 29      Hornuss et al.      10.1101/2020.04.28.20083311   45   38            84.444444
#> 30        Yan et al. 2               10.1111/alr.22592.  128   75            58.593750
#> 31 Klopfenstein et al.     10.1016/j.medmal.2020.04.006  114   54            47.368421
#> 32        Gudbjartsson            10.1056/NEJMoa2006100 1113  119            10.691824
#> 33           Wee et al       10.1007/s00405-020-05999-5  154   35            22.727273
#> 34        Dawson et al    10.1101/2020.05.13.20101006.t   42   18            42.857143
#> 35    Ji Yun Noh et al       10.1016/j.jinf.2020.05.035  199   52            26.130653
#>                                                                                                                         comments
#> 1  In the abstact of the paper it says 60 cases with varying degree of hyposmia and 2 patients with anosmia -- Results showed 44
#> 2                                                                                       Preprint, only small part of whole paper
#> 3                                                                                             Changed to 17 cases, peer reviewed
#> 4                       Changed cases to 2; 4 patients had "smell and taste disorders", of which 2 had both and 2 had only taste
#> 5                                                                                                                     looks good
#> 6                                                                                                Changed cases to 14, looks good
#> 7                                                                                     Changes case numbers to 22 and total to 34
#> 8                                                                          Looks good, measured olfactory dysfunction over phone
#> 9                                                                                   Studied "loss of smell or taste", looks good
#> 10                                                                    Smartphone app tracking, Studied "Loss of smell and taste"
#> 11                                                                                                                    looks good
#> 12                                                                                                                    looks good
#> 13                                                                                                           Changed cases to 95
#> 14                                                  Of those who were tested, I found 34/54 had an altered sense of taste/smnell
#> 15                                                                                              Studied "loss of smell or taste"
#> 16                                                                                                                    looks good
#> 17                                                                                                  studied "anosmia or ageusia"
#> 18                                                                                                                    looks good
#> 19                                                                                                   I could not find this paper
#> 20                                                                                                           Changed to 14 cases
#> 21                                                                                                          Changed cases to 136
#> 22                                                                        Looks good, Measures Mild-servere hyponosmia + Anosmia
#> 23                                                                                                   Check on this; some english
#> 24                                                                                                        Good study; subjective
#> 25                                                                                                                    Subjective
#> 26                                                                                                                    Subjective
#> 27                                                                                                                    looks good
#> 28                                                                                                                          good
#> 29                                                                                                                          Good
#> 30                               68 ambulatory + 7 admitted of all covid positive; are we sure this does not overlap with Yan 1?
#> 31                                                                                                                          Good
#> 32                                                                                                                          Good
#> 33                                                                                                                          <NA>
#> 34                                                                                                                          <NA>
#> 35                                                                                                                          <NA>
#>    objectivity                    measured
#> 1    Objective Taste and Smell, Smell only
#> 2    Objective                  Smell only
#> 3    Objective                  Smell only
#> 4   Subjective                  Smell only
#> 5   Subjective             Taste and Smell
#> 6   Subjective                  Smell only
#> 7   Subjective                  Smell only
#> 8   Subjective                  Smell only
#> 9   Subjective             Taste and Smell
#> 10  Subjective             Taste and Smell
#> 11  Subjective                  Smell only
#> 12  Subjective                  Smell only
#> 13  Subjective              Taste or Smell
#> 14  Subjective              Taste or Smell
#> 15  Subjective              Taste or Smell
#> 16  Subjective                  Smell only
#> 17  Subjective              Taste or Smell
#> 18  Subjective                  Smell only
#> 19  Subjective smell only, taste and smell
#> 20  Subjective Taste and Smell, Smell only
#> 21  Subjective              Taste or smell
#> 22   Objective                       Smell
#> 23  Subjective          Taste and/or Smell
#> 24  Subjective                  Smell only
#> 25  Subjective              Taste or Smell
#> 26  Subjective                  Smell only
#> 27  Subjective Taste and Smell, Smell only
#> 28   Objective                       Smell
#> 29   Objective                       Smell
#> 30  Subjective                       Smell
#> 31  Subjective                       Smell
#> 32  Subjective              Taste or Smell
#> 33  Subjective              Smell or Taste
#> 34  Subjective                       Smell
#> 35  Subjective                  Smell only
#>                                                                                                                                                 testType
#> 1                                                                            Connecticut Chemosensory Clinical Research Center orthonasal olfaction test
#> 2                                                             5 odor smell panel - no indication of validation - used test-retest to measure reliability
#> 3                                          Connecticut Chemosensory Clinical Research Center orthonasal olfaction test, self administered olfactory test
#> 4                                                                                                                             Unknown, Hospital Reported
#> 5                                                                                                                                          Self Reported
#> 6                                                                                                                                          Self Reported
#> 7                                                                                                                                          Self Reported
#> 8                                                                                                                                          Self Reported
#> 9                                                                                                                                          Self Reported
#> 10                                                                                                                                         Self Reported
#> 11                                                                                                                                         Self Reported
#> 12                                                                                                                                         Self Reported
#> 13                                                                                                                                         Self Reported
#> 14                                                                                                                                         Self Reported
#> 15                                                                                                                            Unknown, Hospital Reported
#> 16                                                                                                                                         Self Reported
#> 17                                                                                                                                         Self Reported
#> 18                                                                                                     Self Reported, Survey based on NHANES and sQOD-NS
#> 19                                                                                                                          Self Reported, Not specified
#> 20                                                                                                                                         Self Reported
#> 21                                                                                                                                  Self Reported Survey
#> 22 Connecticut Chemosensory Clinical Research Center orthonasal olfaction test, home discrimination test, validated test from previoud Vaira et al study
#> 23                                                                                   Sniffin Sticks Screening 12 Test and Taste Strips for some patients
#> 24                                                                                                                            Self Reported, EHR Records
#> 25                                                                                                                                               SNOT-22
#> 26                                                                                                                                         Self Reported
#> 27                                                                                                                                         Self Reported
#> 28                                                                                                  University of Pennsylvania Smell Identification Test
#> 29                                                                                                                                        Sniffin Sticks
#> 30                                                                                                                                         Self Reported
#> 31                                                                                                                                         Self Reported
#> 32                                                                                                                                         Self Reported
#> 33                                                                                                                            Self Reported, Unspecified
#> 34                                                                                                                                         Self Reported
#> 35                                                                                                                                         Self Reported
#>                          country
#> 1                          Italy
#> 2                         Sweden
#> 3                          Italy
#> 4         Duhok, Iraqi Kurdistan
#> 5                      UK and US
#> 6                         Israel
#> 7                        Germany
#> 8                    Switzerland
#> 9                          Italy
#> 10                     UK and US
#> 11                            US
#> 12                       Germany
#> 13              US/Online Global
#> 14                         Italy
#> 15                        Taiwan
#> 16                         Italy
#> 17                         Korea
#> 18 Belgium, France, Spain, Italy
#> 19                         Italy
#> 20                         Italy
#> 21                        Israel
#> 22                         Italy
#> 23                       Germany
#> 24                         China
#> 25                         Italy
#> 26                         Spain
#> 27                     Australia
#> 28                          Iran
#> 29                       Germany
#> 30                            US
#> 31                        France
#> 32                       Iceland
#> 33                     Singapore
#> 34                            US
#> 35                   South Korea
#>                                                                                     patientType
#> 1                                                                     18+, nasal swab confirmed
#> 2                                                                       18+, 16 COVID-Confirmed
#> 3                                                   18+, nasal swab confirmed, home-quarantined
#> 4                                                                     All ages, COVID-Confirmed
#> 5                                                                        16-90, COVID-Confirmed
#> 6                                                       15+, Hospitalized, nasal swab confirmed
#> 7                                   Patient showing symptoms, only COVID positive patients used
#> 8                                All ages, COVID-Confirmed, Hospitalized or Treated at hospital
#> 9                                                   All ages, COVID-Confirmed, home-quarantined
#> 10                                16-90, COVID-Confirmed,unspecified hospitalization/quarantine
#> 11                                                            COVID-Confirmed, home-quarantined
#> 12                                                            COVID-Confirmed, home-quarantined
#> 13                                       COVID-19 Confirmed, online presumably home-quarantined
#> 14 Home-quarantined, COVID-positive included, Household contacts of patients from Spinato paper
#> 15                                                                                      Unknown
#> 16                                                            Hospitalized and home-quarantined
#> 17                                                                             Home-quarantined
#> 18                                                              Home-quarantined, not specified
#> 19                                                                                Not specified
#> 20                                                                                 Hospitalized
#> 21                                                              Home-quarantined, not specified
#> 22                                   Home-quarantined hospital staff, and hospitalized patients
#> 23                                                                                 hospitalized
#> 24                                                                                 Hospitalized
#> 25                                                                             Home-quarantined
#> 26                                                                                Hospitalized?
#> 27                                                            Unspecified, Clinically evaluated
#> 28                                                                                 Hospitalized
#> 29                                                                                 Hospitalized
#> 30                                                            Hospitalized and home-quarantined
#> 31                                                            Hospitalized and home-quarantined
#> 32                                                                                      Unknown
#> 33                                                                                 Hospitalized
#> 34                                                                             Home-quarantined
#> 35                                        Unspecified, Likely Hospitalized and Home-quarantined

# load metafor package
library(metafor)

# compute effect size
dat <- escalc(measure="PR", xi=xi, ni=ni, data=dat)

# split data into objective and subjective datasets
dat_split <- split(dat, dat$objectivity)
dat_objective <- dat_split[["Objective"]]
dat_subjective <- dat_split[["Subjective"]]

# random-effects model all studies
res_all <- rma(yi, vi, data=dat)
print(res_all, digits=2)
#> 
#> Random-Effects Model (k = 35; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.06 (SE = 0.01)
#> tau (square root of estimated tau^2 value):      0.24
#> I^2 (total heterogeneity / total variability):   99.28%
#> H^2 (total variability / sampling variability):  138.10
#> 
#> Test for Heterogeneity:
#> Q(df = 34) = 7878.14, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub      
#>     0.49  0.04  11.88  <.01   0.41   0.57  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# random-effects model objective
res_objective <- rma(yi, vi, data=dat_objective)
print(res_objective, digits=2)
#> 
#> Random-Effects Model (k = 6; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.03 (SE = 0.02)
#> tau (square root of estimated tau^2 value):      0.16
#> I^2 (total heterogeneity / total variability):   94.83%
#> H^2 (total variability / sampling variability):  19.35
#> 
#> Test for Heterogeneity:
#> Q(df = 5) = 132.69, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub      
#>     0.75  0.07  10.70  <.01   0.61   0.89  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# random-effects model subjective
res_subjective <- rma(yi, vi, data=dat_subjective)
print(res_subjective, digits=2)
#> 
#> Random-Effects Model (k = 29; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.05 (SE = 0.01)
#> tau (square root of estimated tau^2 value):      0.22
#> I^2 (total heterogeneity / total variability):   99.23%
#> H^2 (total variability / sampling variability):  129.15
#> 
#> Test for Heterogeneity:
#> Q(df = 28) = 6391.95, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub      
#>     0.44  0.04  10.51  <.01   0.36   0.52  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>