Studies on the Effectiveness of Nicotine Replacement Therapy for Smoking Cessation

Results from 133 studies examining the effectiveness of nicotine replacement therapy (NRT) for smoking cessation at 6+ months of follow-up.

dat.hartmannboyce2018

Format

The data frame contains the following columns:

study	numeric	study identifier

x.nrt	numeric	number of participants in the NRT group who were abstinent at the follow-up

n.nrt	numeric	number of participants in the NRT group

x.ctrl	numeric	number of participants in the control group who were abstinent at the follow-up

n.ctrl	numeric	number of participants in the control group

treatment	character	type of NRT provided in the treatment group

Details

The dataset includes the results from 133 studies examining the effectiveness of nicotine replacement therapy (NRT) for smoking cessation. The results given in this dataset pertain to abstinence at 6+ months of follow-up. NRT was provided to participants in the treatment groups in various forms as indicated by the treatment variable (e.g., gum, patch, inhalator). Note that the dataset includes 136 rows, since a few studies included multiple treatments.

Source

Hartmann‐Boyce, J., Chepkin, S. C., Ye, W., Bullen, C. & Lancaster, T. (2018). Nicotine replacement therapy versus control for smoking cessation. Cochrane Database of Systematic Reviews, 5, CD000146. https://doi.org//10.1002/14651858.CD000146.pub5

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, smoking, risk ratios, Mantel-Haenszel method

Examples

### copy data into 'dat' and examine data
dat <- dat.hartmannboyce2018
head(dat, 10)
#>                   study x.nrt n.nrt x.ctrl n.ctrl treatment
#> 1        Ahluwalia 2006    53   378     42    377       gum
#> 2         Areechon 1988    56    99     37    101       gum
#> 3          Blondal 1989    30    92     22     90       gum
#> 4  Br Thor Society 1983    39   410    111   1208       gum
#> 5         Campbell 1987    13   424      9    412       gum
#> 6         Campbell 1991    21   107     21    105       gum
#> 7           Clavel 1985    24   205      6    222       gum
#> 8  Clavel-Chapelon 1992    47   481     42    515       gum
#> 9           Cooper 2005    17   146     15    147       gum
#> 10      Fagerstrom 1982    30    50     23     50       gum

### load metafor package
library(metafor)

### turn treatment into a factor with the desired ordering
dat$treatment <- factor(dat$treatment, levels=unique(dat$treatment))

### meta-analysis per treatment using the M-H method
lapply(split(dat, dat$treatment), function(x)
       rma.mh(measure="RR", ai=x.nrt,  n1i=n.nrt,
                            ci=x.ctrl, n2i=n.ctrl, data=x, digits=2))
#> $gum
#> 
#> Equal-Effects Model (k = 56)
#> 
#> I^2 (total heterogeneity / total variability):  39.59%
#> H^2 (total variability / sampling variability): 1.66
#> 
#> Test for Heterogeneity: 
#> Q(df = 55) = 91.05, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>     0.40  0.03  11.56  <.01   0.33   0.47 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.49   1.40   1.60 
#> 
#> 
#> $patch
#> 
#> Equal-Effects Model (k = 51)
#> 
#> I^2 (total heterogeneity / total variability):  23.59%
#> H^2 (total variability / sampling variability): 1.31
#> 
#> Test for Heterogeneity: 
#> Q(df = 50) = 65.44, p-val = 0.07
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>     0.49  0.03  14.23  <.01   0.43   0.56 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.64   1.53   1.75 
#> 
#> 
#> $inhalator
#> 
#> Equal-Effects Model (k = 4)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 0.64
#> 
#> Test for Heterogeneity: 
#> Q(df = 3) = 1.93, p-val = 0.59
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.64  0.17  3.73  <.01   0.30   0.98 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.90   1.36   2.67 
#> 
#> 
#> $`intranasal spray`
#> 
#> Equal-Effects Model (k = 4)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 0.54
#> 
#> Test for Heterogeneity: 
#> Q(df = 3) = 1.63, p-val = 0.65
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.70  0.15  4.53  <.01   0.40   1.00 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     2.02   1.49   2.73 
#> 
#> 
#> $`tablets/lozenges`
#> 
#> Equal-Effects Model (k = 8)
#> 
#> I^2 (total heterogeneity / total variability):  71.31%
#> H^2 (total variability / sampling variability): 3.48
#> 
#> Test for Heterogeneity: 
#> Q(df = 7) = 24.39, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.42  0.07  5.97  <.01   0.28   0.55 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.52   1.32   1.74 
#> 
#> 
#> $`oral spray`
#> 
#> Equal-Effects Model (k = 1)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 1.00
#> 
#> Test for Heterogeneity: 
#> Q(df = 0) = 0.00, p-val = 1.00
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.91  0.35  2.57  0.01   0.21   1.60 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     2.48   1.24   4.94 
#> 
#> 
#> $`choice of product`
#> 
#> Equal-Effects Model (k = 7)
#> 
#> I^2 (total heterogeneity / total variability):  42.34%
#> H^2 (total variability / sampling variability): 1.73
#> 
#> Test for Heterogeneity: 
#> Q(df = 6) = 10.41, p-val = 0.11
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.32  0.05  6.34  <.01   0.22   0.42 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.37   1.25   1.52 
#> 
#> 
#> $`patch and inhalator`
#> 
#> Equal-Effects Model (k = 1)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 1.00
#> 
#> Test for Heterogeneity: 
#> Q(df = 0) = 0.00, p-val = 1.00
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.07  0.32  0.21  0.83  -0.55   0.69 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.07   0.57   1.99 
#> 
#> 
#> $`patch and lozenge`
#> 
#> Equal-Effects Model (k = 1)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 1.00
#> 
#> Test for Heterogeneity: 
#> Q(df = 0) = 0.00, p-val = 1.00
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.60  0.30  1.98  0.05   0.01   1.20 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.83   1.01   3.31 
#> 
#> 
#> $`patch and gum`
#> 
#> Equal-Effects Model (k = 2)
#> 
#> I^2 (total heterogeneity / total variability):  50.04%
#> H^2 (total variability / sampling variability): 2.00
#> 
#> Test for Heterogeneity: 
#> Q(df = 1) = 2.00, p-val = 0.16
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.14  0.30  0.46  0.64  -0.45   0.72 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.15   0.64   2.06 
#> 
#> 
#> $`patch, gum, and lozenge`
#> 
#> Equal-Effects Model (k = 1)
#> 
#> I^2 (total heterogeneity / total variability):  0.00%
#> H^2 (total variability / sampling variability): 1.00
#> 
#> Test for Heterogeneity: 
#> Q(df = 0) = 0.00, p-val = 1.00
#> 
#> Model Results (log scale):
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     2.71  1.03  2.63  <.01   0.69   4.72 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb   ci.ub 
#>    15.00   2.00  112.54 
#> 
#> 

### all combined
rma.mh(measure="RR", ai=x.nrt,  n1i=n.nrt,
                     ci=x.ctrl, n2i=n.ctrl, data=dat, digits=2)
#> 
#> Equal-Effects Model (k = 136)
#> 
#> I^2 (total heterogeneity / total variability):  38.69%
#> H^2 (total variability / sampling variability): 1.63
#> 
#> Test for Heterogeneity: 
#> Q(df = 135) = 220.20, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>     0.44  0.02  21.25  <.01   0.40   0.48 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     1.55   1.49   1.61 
#>