Studies on Prophylactic Use of Lidocaine After a Heart Attack

Results from 6 studies evaluating mortality from prophylactic use of lidocaine in acute myocardial infarction.

dat.hine1989

Format

The data frame contains the following columns:

study	numeric	study number

source	character	source of data

n1i	numeric	number of patients in lidocaine group

n2i	numeric	number of patients in control group

ai	numeric	number of deaths in lidocaine group

ci	numeric	number of deaths in control group

Details

Hine et al. (1989) conducted a meta-analysis of death rates in randomized controlled trials in which prophylactic lidocaine was administered to patients with confirmed or suspected acute myocardial infarction. The dataset describes the mortality at the end of the assigned treatment period for control and intravenous lidocaine treatment groups for six studies. The question of interest is whether there is a detrimental effect of lidocaine. Because the studies were conducted to compare rates of arrhythmias following a heart attack, the studies, taken individually, are too small to detect important differences in mortality rates.

The data in this dataset were obtained from Table I in Normand (1999, p. 322).

Source

Normand, S. T. (1999). Meta-analysis: Formulating, evaluating, combining, and reporting. Statistics in Medicine, 18(3), 321–359. https://doi.org/10.1002/(sici)1097-0258(19990215)18:3<321::aid-sim28>3.0.co;2-p

References

Hine, L. K., Laird, N., Hewitt, P., & Chalmers, T. C. (1989). Meta-analytic evidence against prophylactic use of lidocaine in acute myocardial infarction. Archives of Internal Medicine, 149(12), 2694–2698. https://doi.org/10.1001/archinte.1989.00390120056011

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, cardiology, risk differences

Examples

### copy data into 'dat' and examine data
dat <- dat.hine1989
dat
#>   study         source n1i n2i ai ci
#> 1     1  Chopra et al.  39  43  2  1
#> 2     2       Mogensen  44  44  4  4
#> 3     3    Pitt et al. 107 110  6  4
#> 4     4   Darby et al. 103 100  7  5
#> 5     5 Bennett et al. 110 106  7  3
#> 6     6 O'Brien et al. 154 146 11  4

# \dontrun{

### load metafor package
library(metafor)

### calculate risk differences and corresponding sampling variances
dat <- escalc(measure="RD", n1i=n1i, n2i=n2i, ai=ai, ci=ci, data=dat)
dat
#> 
#>   study         source n1i n2i ai ci     yi     vi 
#> 1     1  Chopra et al.  39  43  2  1 0.0280 0.0018 
#> 2     2       Mogensen  44  44  4  4 0.0000 0.0038 
#> 3     3    Pitt et al. 107 110  6  4 0.0197 0.0008 
#> 4     4   Darby et al. 103 100  7  5 0.0180 0.0011 
#> 5     5 Bennett et al. 110 106  7  3 0.0353 0.0008 
#> 6     6 O'Brien et al. 154 146 11  4 0.0440 0.0006 
#> 

### meta-analysis of risk differences using a random-effects model
res <- rma(yi, vi, data=dat)
res
#> 
#> Random-Effects Model (k = 6; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0 (SE = 0.0006)
#> tau (square root of estimated tau^2 value):      0
#> I^2 (total heterogeneity / total variability):   0.00%
#> H^2 (total variability / sampling variability):  1.00
#> 
#> Test for Heterogeneity:
#> Q(df = 5) = 0.8597, p-val = 0.9731
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub   ​ 
#>   0.0294  0.0131  2.2531  0.0243  0.0038  0.0551  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }