Studies on the Effectiveness of CBT for Reducing Recidivism

Results from 58 studies on the effectiveness of cognitive-behavioral therapy (CBT) for reducing recidivism in juvenile and adult offenders.

dat.landenberger2005

Format

The data frame contains the following columns:

study	character	(first) author and year

pubtype	character	publication type (book chapter, journal article, report, or thesis)

country	character	country where study was carried out (Canada, New Zealand, UK, or USA)

design	character	study design (matched groups, nonequivalent groups, or randomized trial)

program	character	purpose of setting up the CBT program (for demonstration, practice, or research purposes)

setting	character	treatment setting (community or prison)

designprob	character	indication of study design problems (no, favors the control group, or favors the treatment group)

n.ctrl.rec	numeric	number of recidivists in the control group

n.ctrl.non	numeric	number of non-recidivists in the control group

n.cbt.rec	numeric	number of recidivists in the CBT group

n.cbt.non	numeric	number of non-recidivists in the CBT group

interval	numeric	recidivism interval (in months)

group	numeric	study group (adults or juveniles)

age	numeric	mean age of the study group

male	numeric	percentage of males in the study group

minority	numeric	percentage of minorities in the study group

length	numeric	treatment length (in weeks)

sessions	numeric	number of CBT sessions per week

hrs_week	numeric	treatment hours per week

hrs_total	numeric	total hours of treatment

cbt.cogskills	character	CBT component: cognitive skills (yes, no)

cbt.cogrestruct	character	CBT component: cognitive restructuring (yes, no)

cbt.intpprbsolv	character	CBT component: interpersonal problem solving (yes, no)

cbt.socskills	character	CBT component: social skills (yes, no)

cbt.angerctrl	character	CBT component: anger control (yes, no)

cbt.victimimpact	character	CBT component: victim impact (yes, no)

cbt.subabuse	character	CBT component: substance abuse (yes, no)

cbt.behavmod	character	CBT component: behavior modification (yes, no)

cbt.relapseprev	character	CBT component: relapse prevention (yes, no)

cbt.moralrsng	character	CBT component: moral reasoning (yes, no)

cbt.roletaking	character	CBT component: role taking (yes, no)

cbt.other	character	CBT component: other (yes, no)

Details

Landenberger and Lipsey (2005) conducted a meta-analysis of 58 experimental and quasi-experimental studies of the effects of cognitive-behavioral therapy (CBT) on the recidivism rates of adult and juvenile offenders (see also Lipsey et al., 2007). The present dataset includes the results of these studies and a range of potential moderator variables to identify factors associated with variation in treatment effects.

Source

Personal communication.

References

Landenberger, N. A., & Lipsey, M. W. (2005). The positive effects of cognitive-behavioral programs for offenders: A meta-analysis of factors associated with effective treatment. Journal of Experimental Criminology, 1, 451–476. https://doi.org/10.1007/s11292-005-3541-7

Lipsey, M. W., Landenberger, N. A., & Wilson, S. J. (2007). Effects of cognitive-behavioral programs for criminal offenders. Campbell Systematic Reviews, 3(1), 1–27. https://doi.org/10.4073/csr.2007.6

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

psychology, criminology, odds ratios, meta-regression

Examples

### copy data into 'dat' and examine data
dat <- dat.landenberger2005
head(dat)
#>              study pubtype country design  program   setting designprob n.ctrl.rec n.ctrl.non
#> 1    Guerra (1990) journal     USA    rct research    prison         no         23         29
#> 2    Leeman (1993) journal     USA    rct     demo    prison         no         15         22
#> 3  Kownacki (1995) chapter     USA    rct research community         no          4          6
#> 4    Larson (1989) chapter     USA    rct research community         no         11          2
#> 5      Finn (1998)  report     USA    rct practice    prison         no         22         60
#> 6 Goldstein (1989) chapter     USA    rct     demo community         no         14         18
#>   n.cbt.rec n.cbt.non interval     group age male minority length sessions hrs_week hrs_total
#> 1        10        19       12 juveniles  17   50       60     12        1      1.0        12
#> 2         3        17       12 juveniles  16  100       34     26        1      6.0       169
#> 3         2         9        6    adults  28  100       NA      7        1      2.0        14
#> 4         8         5       15    adults  19  100       94     63        1      1.0        63
#> 5        16        66       12    adults  30  100       74     17        5     15.0       240
#> 6         5        28        6 juveniles  NA   NA       NA     13        2      1.5        44
#>   cbt.cogskills cbt.cogrestruct cbt.intpprbsolv cbt.socskills cbt.angerctrl cbt.victimimpact
#> 1           yes             yes             yes            no            no               no
#> 2           yes             yes              no           yes           yes               no
#> 3           yes              no             yes           yes            no               no
#> 4           yes              no             yes           yes            no               no
#> 5           yes             yes             yes            no           yes              yes
#> 6           yes              no              no           yes           yes               no
#>   cbt.subabuse cbt.behavmod cbt.relapseprev cbt.moralrsng cbt.roletaking cbt.other
#> 1           no           no              no            no            yes        no
#> 2           no           no              no           yes            yes        no
#> 3           no           no              no            no            yes        no
#> 4           no           no              no            no            yes       yes
#> 5           no           no              no           yes             no       yes
#> 6           no           no              no           yes            yes       yes

### load metafor package
library(metafor)

### calculate log odds ratios (for non-recidivism in CBT vs. control groups) and sampling variances
dat <- escalc(measure="OR", ai=n.cbt.non, bi=n.cbt.rec, ci=n.ctrl.non, di=n.ctrl.rec, data=dat)

### fit random-effects model
res <- rma(yi, vi, data=dat)
res
#> 
#> Random-Effects Model (k = 58; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.1046 (SE = 0.0352)
#> tau (square root of estimated tau^2 value):      0.3234
#> I^2 (total heterogeneity / total variability):   70.62%
#> H^2 (total variability / sampling variability):  3.40
#> 
#> Test for Heterogeneity:
#> Q(df = 57) = 213.6898, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub      
#>   0.4226  0.0605  6.9880  <.0001  0.3041  0.5411  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average OR and corresponding 95% CI/PI
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.53  1.36  1.72  0.80  2.91 
#> 

### examine if number of treatment sessions per week is a potential moderator
res <- rma(yi, vi, mods = ~ sessions, data=dat)
res
#> 
#> Mixed-Effects Model (k = 58; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0761 (SE = 0.0285)
#> tau (square root of estimated tau^2 value):             0.2758
#> I^2 (residual heterogeneity / unaccounted variability): 62.88%
#> H^2 (unaccounted variability / sampling variability):   2.69
#> R^2 (amount of heterogeneity accounted for):            27.27%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 56) = 146.6400, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 6.0695, p-val = 0.0138
#> 
#> Model Results:
#> 
#>           estimate      se    zval    pval   ci.lb   ci.ub    
#> intrcpt     0.2022  0.1015  1.9931  0.0463  0.0034  0.4010  * 
#> sessions    0.0695  0.0282  2.4636  0.0138  0.0142  0.1248  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### predicted ORs for 1, 2, 5, or 10 sessions per week
predict(res, newmods=c(1,2,5,10), transf=exp, digits=2)
#> 
#>   pred ci.lb ci.ub pi.lb pi.ub 
#> 1 1.31  1.12  1.53  0.75  2.30 
#> 2 1.41  1.25  1.59  0.81  2.45 
#> 3 1.73  1.49  2.02  0.99  3.04 
#> 4 2.45  1.64  3.66  1.25  4.80 
#>