Studies on Acupoint P6 Stimulation for Preventing Nausea

Results from studies examining the effectiveness of wrist acupuncture point P6 stimulation for preventing postoperative nausea.

dat.lee2004

Format

The data frame contains the following columns:

id	numeric	trial id number

study	character	first author

year	numeric	study year

ai	numeric	number of patients experiencing nausea in the treatment group

n1i	numeric	total number of patients in treatment group

ci	numeric	number of patients experiencing nausea in the sham group

n2i	numeric	total number of patients in the sham group

Details

Postoperative nausea and vomiting are common complications following surgery and anaesthesia. As an alternative to drug therapy, acupuncture has been studied as a potential treatment in several trials. The dataset contains the results from 16 clinical trials examining the effectiveness of wrist acupuncture point P6 stimulation for preventing postoperative nausea.

Source

Lee, A., & Done, M. L. (2004). Stimulation of the wrist acupuncture point P6 for preventing postoperative nausea and vomiting. Cochrane Database of Systematic Reviews, 3, CD003281. https://doi.org/10.1002/14651858.CD003281.pub2

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, alternative medicine, risk ratios

Examples

### copy data into 'dat' and examine data
dat <- dat.lee2004
dat
#>    id        study year ai n1i ci n2i
#> 1   1      Agarwal 2000 18 100 20 100
#> 2   2      Agarwal 2002  5  50 18  50
#> 3   3     Alkaissi 1999  9  20  7  20
#> 4   4     Alkaissi 2002 32 135 31 139
#> 5   5        Allen 1994  9  23 10  23
#> 6   6 Andrzejowski 1996 11  18 12  18
#> 7   7       Duggal 1998 69 122 80 122
#> 8   8       Dundee 1986  3  25 12  25
#> 9   9 Ferrera-Love 1996  1  30  1  30
#> 10 10       Gieron 1993 11  30 19  30
#> 11 11       Harmon 1999  7  44 16  39
#> 12 12       Harmon 2000  4  47  6  47
#> 13 13           Ho 1996  1  30 13  30
#> 14 14         Rusy 2002 24  40 71  80
#> 15 15         Wang 2002 16  50 53  88
#> 16 16       Zarate 2001 28 110 25 111

# \dontrun{

### load metafor package
library(metafor)

### meta-analysis based on log risk ratios
res <- rma(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat)
res
#> 
#> Random-Effects Model (k = 16; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0554 (SE = 0.0487)
#> tau (square root of estimated tau^2 value):      0.2353
#> I^2 (total heterogeneity / total variability):   46.04%
#> H^2 (total variability / sampling variability):  1.85
#> 
#> Test for Heterogeneity:
#> Q(df = 15) = 30.2238, p-val = 0.0111
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb    ci.ub    ​ 
#>  -0.3145  0.0977  -3.2200  0.0013  -0.5059  -0.1231  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.73  0.60  0.88  0.44  1.20 
#> 

# }