The Effect of Red on Perceived Attractiveness

Results from studies in which participants rated the attractiveness of photos that featured red or a control color. See OSF project at https://osf.io/xy47p/.

dat.lehmann2018

Format

The data frame contains the following columns:

Short_Title	character	Shortened citation formatted Author name(s), year of publication - Experiment number. All cells in the column are unique for use as labels in the meta-analysis.

Full_Citation	character	Full citation in APA format.

Short_Citation	character	Shortened citation of different format, exactly as it would appear in an in-text citation.

Year	numeric	Year study published (whether in journal or published online).

Study	character	Experiment number. If only one experiment presented in a paper, then ‘Exp 1’, otherwise numbered according to numbering within paper.

Peer_Reviewed	character	Whether the experiment was published in a peer-reviewed journal or not. ‘Yes’ = peer-reviewed journal, ‘No’ can mean in press, online publication, or other. Column for moderator analysis.

Source_Type	character	Location where experiment is available, including journal articles, conference proceedings, online-only, and other options. More specific than whether peer-reviewed or not.

Preregistered	character	Whether experiment was pre-registered or not.

Moderator_Group	character	In some studies, a moderator was intentionally investigated that was meant to reduce the red-romance effect. Data for studies where the red-romance effect is expected to be moderated are marked ‘Yes’ in this column. All others are blank.

Gender	character	Gender of rater (male or female). In all cases, gender of stimuli will be opposite.

Color_Contrast	character	The color used as the contrast against red. In some cases, not every contrast color was listed. We chose to examine only contrasts that were present in the original studies, when possible. This column contains only the contrasts we examined in this meta-analysis.

Color_Form	character	Location of color in photo. Background = background or border color manipulated; Face = facial redness manipulated; Shirt, Dress, Item = color of specified object manipulated; Dot = a dot of color on shirt manipulated.

Photo_Type	character	Amount of body visible in photo. Head Shot = head only; Bust = head, shoulders, sometimes torso; Full Body = entire body visible.

DV_Type	character	Scale used for DV. ‘Perceived attractiveness’ = the perceived attractiveness scale used in the original studies; alternate scales are differentiated.

DV_Items	numeric	Number of items in DV scale.

DV_Scale	character	Full length of DV scale, if clear.

DV_ScaleBottom	numeric	Lower anchor of DV scale.

DV_ScaleTop	numeric	Upper anchor of DV scale.

Location	character	Country where study took place, if clear. ‘Worldwide’ in some cases of online participation without IP filtering of participants.

Continent	character	Continent where study took place, for the sake of creating larger categories for analysis.

Participants	character	Basic notes about participants. Students = high school, undergraduate, or graduate students; online = participants were gathered online; adult = no other common identifying factor given. Put into fewer categories for ease of analysis.

Participant_Notes	character	A finer grained description of participant characteristics.

Design	character	Whether study was a between- or within-subjects design.

Eth_Majority	character	Basic notes about participant ethnicity for ease of analysis. This represents the ethnic majority within the sample.

Eth_Majority_Detail	character	A finer grained description of participant characteristics, including in some cases participant counts when the ethnic majority was close to another category.

Eth_Stim	character	Ethnicity of the people pictured in the stimulus materials.

Eth_Match	character	Whether the ethnic majority of the participant pool matched the ethnicity of stimulus photos.

Red_Age	numeric	Mean age of participants in red group. If not given for specific group, then mean age overall.

Control_Age	numeric	Mean age of participants in control group. If not given for specific group, then mean age overall.

Color_Red	character	Specific values of red color, if given. ‘No data’ if not given or unclear.

Color_Control	character	Specific values of control color, if given. ‘No data’ if not given or unclear.

Red_Original	character	Whether the red color used in the study is within 5 units of the LCh values for red used in the original study.

Color_Match	character	Whether the control color used in the study is within 5 units of the red color on the L and C parameters. In cases where the control color used was white, it was not possible for the L and C parameters to match.

Presentation_Control	character	Whether the color of the stimulus viewed by each participant was consistent, as in participants viewing everything on paper or the same computer, versus uncontrolled presentation of the stimulus, as in viewing stimulus on different computers.

Stimuli_Presentation	character	Method for presenting stimuli. ‘Paper’ = stimuli printed on paper, shown in-person; ‘Screen’ = stimuli shown on-screen, not carefully controlled; ‘Screen Control’ = stimuli shown on-screen, but screen carefully color-matched.

Red_N	numeric	Number of participants in red group.

Red_M	numeric	Mean rating of DV in red group.

Red_SD	numeric	Standard deviation of DV in red group.

Control_N	numeric	Number of participants in control group.

Control_M	numeric	Mean rating of DV in control group.

Control_SD	numeric	Standard deviation of DV in control group.

SD_diff	numeric	Calculated for within-subjects studies, standard deviation of difference scores.

RM_r	numeric	Calculated for within-subjects studies, correlation between participant ratings of red and control attractiveness.

Control_Attractiveness	numeric	Attractiveness of stimuli in control condition, calculated as (Control_M - DV_ScaleBottom) / DV_ScaleTop, in order to compare attractiveness ratings across different scales.

Notes	character	Any additional notes on the study.

Total.SampleSize	numeric	Total unique participants in the study.

pooled	numeric	Pooled standard deviation for within-subjects studies.

yi	numeric	Standardized mean difference.

vi	numeric	Corresponding sampling variance.

Details

This is data from a meta-analysis of studies that test the red-romance hypothesis, which is that the color red enhances heterosexual attraction in romantic contexts. Analyzing male participants only, the meta-analysis should show a small, statistically significant effect (d = 0.26 [0.12, 0.40], p = .0004, N = 2,961). Analyzing female participants only should show a very small effect (d = 0.13 [0.01, 0.25], p = .03, N = 2,739). The analyses in the published meta-analysis found clear evidence of upward bias in the estimate for female participants and equivocal evidence for male participants. Moderator analyses suggest effect sizes may have declined over time (both genders), may be largest when an original shade of red is used (men only), and may be smaller in pre-registered studies (women only).

Source

Lehmann, G. K., Elliot, A. J., & Calin-Jageman, R. J. (2018). Meta-analysis of the effect of red on perceived attractiveness. Evolutionary Psychology, 16(4). https://doi.org/10.1177/1474704918802412 https://osf.io/xy47p/

Author

Robert Calin-Jageman, rcalinjageman@dom.edu, https://calin-jageman.net

Examples

### copy data into 'dat' and examine data
dat <- dat.lehmann2018
head(dat)
#>                                   Short_Title
#> 1                Roberts et al., 2010 - Exp 2
#> 2                Roberts et al., 2010 - Exp 2
#> 3  Wartenberg et al., 2011 - Exp 1 - In Group
#> 4 Gilston & Privitera, 2016 - Exp 1 - Healthy
#> 5                Roberts et al., 2010 - Exp 1
#> 6                Roberts et al., 2010 - Exp 1
#>                                                                                                                                                                                                                          Full_Citation
#> 1 Roberts, S. C., Owen, R. C., & Havlicek, J. (2010). Distinguishing between Perceiver and Wearer Effects in Clothing Color-Associated Attributions. Evolutionary Psychology, 8(3), 350-364. http://doi.org/10.1177/147470491000800304
#> 2 Roberts, S. C., Owen, R. C., & Havlicek, J. (2010). Distinguishing between Perceiver and Wearer Effects in Clothing Color-Associated Attributions. Evolutionary Psychology, 8(3), 350-364. http://doi.org/10.1177/147470491000800304
#> 3                  Wartenberg, W., Hoepfner, T., Potthast, P., & Mirau, A. (2011). If you wear red on a date, you will please your mate. Proceedings of Empiriepraktikumskongress, 6th, Aug. 7, pp 26-27. University of Jena, Germany.
#> 4                         Gilston, A., & Privitera, G. J. (2015). A 'Healthy' Color: Information About Healthy Eating Attenuates the 'Red Effect.' Global Journal of Health Science, 8(1), 56-61. https://doi.org/10.5539/gjhs.v8n1p56
#> 5 Roberts, S. C., Owen, R. C., & Havlicek, J. (2010). Distinguishing between Perceiver and Wearer Effects in Clothing Color-Associated Attributions. Evolutionary Psychology, 8(3), 350-364. http://doi.org/10.1177/147470491000800304
#> 6 Roberts, S. C., Owen, R. C., & Havlicek, J. (2010). Distinguishing between Perceiver and Wearer Effects in Clothing Color-Associated Attributions. Evolutionary Psychology, 8(3), 350-364. http://doi.org/10.1177/147470491000800304
#>     Year PRPublication            Source_Type      Preregistered Moderator_Group  Context  Gender
#> 1 2010.1           Yes                Journal Not Pre-Registered              No Romantic Females
#> 2 2010.1           Yes                Journal Not Pre-Registered              No Romantic   Males
#> 3 2011.0            No Conference Proceedings Not Pre-Registered              No Romantic Females
#> 4 2016.0           Yes                Journal Not Pre-Registered              No Romantic   Males
#> 5 2010.1           Yes                Journal Not Pre-Registered              No Romantic Females
#> 6 2010.1           Yes                Journal Not Pre-Registered              No Romantic   Males
#>     Color_Contrast Color_Form Photo_Type             Photo_Similarity
#> 1 Blue/Green/White   Clothing       Bust Different between conditions
#> 2 Blue/Green/White   Clothing       Bust Different between conditions
#> 3             Blue Background       Bust      Same between conditions
#> 4            White   Clothing  Head Shot      Same between conditions
#> 5 Blue/Green/White   Clothing       Bust Different between conditions
#> 6 Blue/Green/White   Clothing       Bust Different between conditions
#>                                DV_Type DV_Items DV_Scale DV_ScaleBottom DV_ScaleTop Location
#> 1 Single item rating of attractiveness        1     1-10              1          10  England
#> 2 Single item rating of attractiveness        1     1-10              1          10  England
#> 3             Perceived attractiveness        4      1-7              1           7  Germany
#> 4             Perceived attractiveness        1      1-7              1           7      USA
#> 5 Single item rating of attractiveness        1     1-10              1          10  England
#> 6 Single item rating of attractiveness        1     1-10              1          10  England
#>       Continent Participants Participant_Notes          Design Eth_Majority
#> 1        Europe     Students        Undergrads Within Subjects        White
#> 2        Europe     Students        Undergrads Within Subjects        White
#> 3        Europe     Students        Undergrads Within Subjects        White
#> 4 North America     Students         Undergrad Within Subjects         <NA>
#> 5        Europe     Students        Undergrads Within Subjects        White
#> 6        Europe     Students        Undergrads Within Subjects        White
#>                 Eth_Majority_Detail Eth_Stim Eth_Match Red_Age Control_Age Color_Red Color_Control
#> 1 Noted in manuscript all caucasian    White   Matched      NA          NA   No Data       No Data
#> 2 Noted in manuscript all caucasian    White   Matched      NA          NA   No Data       No Data
#> 3                             White    White   Matched   21.47       21.47   No Data       No Data
#> 4                              <NA>    White      <NA>   19.85       19.85   No Data       No Data
#> 5 Noted in manuscript all caucasian    White   Matched      NA          NA   No Data       No Data
#> 6 Noted in manuscript all caucasian    White   Matched      NA          NA   No Data       No Data
#>   Red_Original Color_Match Presentation_Control Stimuli_Presentation Red_N    Red_M   Red_SD
#> 1           No          No                  Yes               Screen    15 3.660000 0.599760
#> 2           No          No                  Yes               Screen    15 3.653300 0.448600
#> 3           No          No                  Yes               Screen    39 3.000000 1.180000
#> 4           No          No                  Yes               Screen    54 5.650000 1.510000
#> 5           No          No                  Yes               Screen    30 4.496552 1.227654
#> 6           No          No                  Yes               Screen    30 4.596552 1.095445
#>   Control_N Control_M Control_SD SD_diff  RM_r Control_Attractiveness
#> 1        15  3.446667  0.6085667      NA 0.660              0.2718519
#> 2        15  3.455567  0.4820896      NA 0.920              0.2728407
#> 3        39  2.760000  1.1300000    0.63 0.856              0.2933333
#> 4        54  3.090000  1.0900000      NA 0.893              0.3483333
#> 5        30  4.263218  1.1838188      NA 0.660              0.3625798
#> 6        30  4.404598  0.9151259      NA 0.920              0.3782886
#>                                                                          Notes Total.SampleSize
#> 1                        Age range includes both male and female participants.               15
#> 2                        Age range includes both male and female participants.               15
#> 3 Translation of summary provided by Elliot; within subjects info still needed               39
#> 4                                                                         <NA>               54
#> 5                        Age range includes both male and female participants.               30
#> 6                        Age range includes both male and female participants.               30
#>      pooled        yi          vi
#> 1 0.6041794 0.3337773 0.049046910
#> 2 0.4656460 0.4014099 0.016037663
#> 3 1.1552705 0.2036116 0.007916124
#> 4 1.3168523 1.9163675 0.037967265
#> 5 1.2059356 0.1884325 0.023258447
#> 6 1.0093204 0.1852130 0.005905064

# \dontrun{

### load metafor package
library(metafor)

### meta-analyses for male and female participants
red_romance_malep   <- dat[dat$Gender == "Males", ]
red_romance_femalep <- dat[dat$Gender == "Females", ]

res_malep <- rma(yi, vi, data=red_romance_malep, test="knha")
res_malep
#> 
#> Random-Effects Model (k = 45; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.1383 (SE = 0.0422)
#> tau (square root of estimated tau^2 value):      0.3718
#> I^2 (total heterogeneity / total variability):   88.94%
#> H^2 (total variability / sampling variability):  9.04
#> 
#> Test for Heterogeneity:
#> Q(df = 44) = 172.4651, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    tval  df    pval   ci.lb   ci.ub     ​ 
#>   0.2610  0.0680  3.8382  44  0.0004  0.1240  0.3980  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
res_femalep <- rma(yi, vi, data=red_romance_femalep, test="knha")
res_femalep
#> 
#> Random-Effects Model (k = 36; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0483 (SE = 0.0235)
#> tau (square root of estimated tau^2 value):      0.2197
#> I^2 (total heterogeneity / total variability):   52.92%
#> H^2 (total variability / sampling variability):  2.12
#> 
#> Test for Heterogeneity:
#> Q(df = 35) = 73.0481, p-val = 0.0002
#> 
#> Model Results:
#> 
#> estimate      se    tval  df    pval   ci.lb   ci.ub   ​ 
#>   0.1295  0.0580  2.2319  35  0.0321  0.0117  0.2473  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }