dat.li2007 {metadat}	R Documentation

Studies on the Effectiveness of Intravenous Magnesium in Acute Myocardial Infarction

Description

Results from 22 trials examining the effectiveness of intravenous magnesium in the prevention of death following acute myocardial infarction.

Usage


dat.li2007

Format

The data frame contains the following columns:

id 	 numeric 	 trial id number

study 	 character 	 first author or trial name

year 	 numeric 	 publication year

ai 	 numeric 	 number of deaths in the magnesium group

n1i 	 numeric 	 number of patients in the magnesium group

ci 	 numeric 	 number of deaths in the control group

n2i 	 numeric 	 number of patients in the control group

Details

The dataset includes the results from 22 randomized clinical trials that examined the effectiveness of intravenous magnesium in the prevention of death following acute myocardial infarction. It is similar to the dataset dat.egger2001, with some slight differences in the included trials and data used.

Concepts

medicine, cardiology, odds ratios, publication bias

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Li, J., Zhang, Q., Zhang, M., & Egger, M. (2007). Intravenous magnesium for acute myocardial infarction. Cochrane Database of Systematic Reviews, 2, CD002755. https://doi.org/10.1002/14651858.CD002755.pub2

See Also

dat.egger2001

Examples


### copy data into 'dat' and examine data
dat <- dat.li2007
dat

## Not run: 

### load metafor package
library(metafor)

### meta-analysis of all trials except ISIS-4
res <- rma(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, method="EE", subset=-14)
print(res, digits=2)
predict(res, transf=exp, digits=2)

### meta-analysis of all trials including ISIS-4
res <- rma(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, method="EE")
print(res, digits=2)
predict(res, transf=exp, digits=2)

### contour-enhanced funnel plot centered at 0
funnel(res, refline=0, level=c(90, 95, 99), shade=c("white", "gray", "darkgray"))


## End(Not run)

[Package metadat version 1.2-0 Index]