Example Data for the Meta-Analysis of Significance Values

A collection of datasets for illustrating the meta-analysis of significance values (i.e., methods for combining p-values from tests of significance).

dat.metap

Format

A list with the following elements:

beckerp

A numeric vector with 5 hypothetical \(p\)-values

cholest

A data frame with 34 observations on the following 5 variables:

ntreat	numeric	number of patients in the treated group

ncontrol	numeric	number of patients in the control group

dtreat	numeric	number of deaths in the treated group

dcontrol	numeric	number of deaths in the control group

p	numeric	one-sided \(p\)-values

edgington

A vector of with 7 hypothetical \(p\)-values

mourning

A data frame with 9 observations on the following 3 variables:

stance	character	character variable with levels No stand, Opponent, Supporter

grade	character	character variable with levels G11-12, G7-8, G9-10

p	numeric	\(p\)-values

naep

A data frame with 34 observations on the following 2 variables:

state	character	character variable with two-letter US state names

p	numeric	\(p\)-values

rosenthal

A data frame with 5 observations on the following 3 variables:

t	numeric	t-statistics

df	numeric	degrees of freedom

p	numeric	one-sided \(p\)-values

teachexpect

A vector of 19 \(p\)-values

validity

A data frame with 20 observations on the following 3 variables:

n	numeric	sample sizes

r	numeric	correlation coefficients

p	numeric	one-sided \(p\)-values

zhang

A data frame with 22 observations on the following 11 variables:

study	character	study names

smd	numeric	standardized mean differences

lo	numeric	lower confidence interval limits

hi	numeric	upper confidence interval limits

ntreat	numeric	treated group sample sizes

ncont	numeric	control group sample sizes

n	numeric	total sample sizes

phase	factor	phase the patients were in: acute, healing, healed

sd	numeric	the calculated standard deviations

z	numeric	the calculated z-values

p	numeric	one-sided \(p\)-values

Details

beckerp

Hypothetical \(p\)-values from Becker (1994).

cholest

Trials of interventions for cholesterol lowering from Sutton et al. (2000).

edgington

Hypothetical \(p\)-values from Edgington (1972).

mourning

Results from a study of mourning practices of Israeli youth following the assassination of Itzakh Rabin from Benjamini and Hochberg (2000).

naep

Results of mathematical achievement scores from the National Assessment of Educational Progress from Benjamini and Hochberg (2000).

rosenthal

Hypothetical example from Rosenthal (1978).

teachexpect

\(p\)-values from studies of the effect of manipulating teacher expectancy on student IQ from Becker (1994).

validity

Data from studies of validity of student ratings of their instructors from Becker (1994) including correlations and sample sizes as well as \(p\)-values.

zhang

Data from trials of exercise training for patients with cardiovascular disease from Zhang et al. (2016).

Author

Michael Dewey

Note

The \(p\)-values in cholest have been re-calculated from other data given in the book and so are of higher accuracy than the ones given in the book which are only to two decimal places.

References

Becker, B. J. (1994). Combining significance levels. In H. Cooper & L. V. Hedges (Eds.), The handbook of research synthesis (pp. 215–230). New York: Russell Sage Foundation.

Benjamini, Y., & Hochberg, Y. (2000). On the adaptive control of the false discovery rate in multiple testing with independent statistics. Journal of Educational and Behavioral Statistics, 25(1), 60–83. https://doi.org/10.3102/10769986025001060

Edgington, E. S. (1972). An additive method for combining probability values from independent experiments. Journal of Psychology, 80(2), 351-363. https://doi.org/10.1080/00223980.1972.9924813

Rosenthal, R. (1978). Combining results of independent studies. Psychological Bulletin, 85(1), 185–193. https://doi.org/10.1037/0033-2909.85.1.185

Sutton, A. J., Abrams, K. R., Jones, D. R., Sheldon, T. A., & Song, F. (2000). Methods for meta-analysis in medical research. Chichester, UK: Wiley.

Zhang, Y.-M., Lu, Y., Tang, Y., Yang, D., Wu, H.-F., Bian, Z.-P., Xu, J.-D., Gu, C.-R., Wang, L.-S., & Chen, X.-J. (2016). The effects of different initiation time of exercise training on left ventricular remodeling and cardiopulmonary rehabilitation in patients with left ventricular dysfunction after myocardial infarction. Disability and Rehabilitation, 38(3), 268–276. https://doi.org/10.3109/09638288.2015.1036174

Concepts

combining p-values

Examples

dat.metap
#> $beckerp
#> [1] 0.016 0.067 0.250 0.405 0.871
#> 
#> $cholest
#>    ntreat ncontrol dtreat dcontrol           p
#> 1     204      202     28       51 0.998016873
#> 2     285      147     70       38 0.621867354
#> 3     156      119     37       40 0.964218674
#> 4      88       30      2        3 0.958251446
#> 5      30       33      0        3 0.897783763
#> 6     279      276     61       82 0.982183843
#> 7     206      206     41       55 0.947379481
#> 8     123      129     20       24 0.685277683
#> 9    1018     1015    111      113 0.565458229
#> 10    427      143     81       27 0.505816538
#> 11    244      253     31       51 0.986394049
#> 12     50       50     17       12 0.139483080
#> 13     47       48     23       20 0.240786632
#> 14     30       60      0        4 0.852802572
#> 15   5552     2789   1025      723 1.000000000
#> 16    424      422    174      178 0.631808415
#> 17    199      194     28       31 0.700823720
#> 18    350      367     42       48 0.667009267
#> 19     79       78      4        5 0.635555111
#> 20   1149     1129     37       48 0.901171382
#> 21    221      237     39       28 0.040320723
#> 22     54       26      8        1 0.108710301
#> 23     71       72      5        7 0.710030246
#> 24   4541     4516    269      248 0.187977610
#> 25    421      417     49       62 0.914767596
#> 26     94       94      0        1 0.750663988
#> 27    311      317     19       12 0.094383576
#> 28   1906     1900     68       71 0.609229717
#> 29   2051     2030     44       43 0.476479349
#> 30   6582     1663     33        3 0.057485568
#> 31   5331     5296    236      181 0.003805613
#> 32     48       49      0        1 0.747811170
#> 33     94       52      1        0 0.375407797
#> 34     23       29      1        2 0.613915542
#> 
#> $edgington
#> [1] 0.20 0.35 0.35 0.40 0.40 0.40 0.40
#> 
#> $mourning
#>      stance  grade      p
#> 1  Opponent G11-12 0.9600
#> 2 Supporter   G7-8 0.8094
#> 3  No stand G11-12 0.7240
#> 4  No stand   G7-8 0.5870
#> 5  Opponent   G7-8 0.4989
#> 6 Supporter G11-12 0.4241
#> 7  Opponent  G9-10 0.0133
#> 8  No stand  G9-10 0.0098
#> 9 Supporter  G9-10 0.0074
#> 
#> $naep
#>    state       p
#> 1     GA 0.85628
#> 2     AR 0.60282
#> 3     AL 0.44008
#> 4     NJ 0.41998
#> 5     NE 0.38640
#> 6     ND 0.36890
#> 7     DE 0.31162
#> 8     MI 0.23522
#> 9     LA 0.20964
#> 10    IN 0.19388
#> 11    WI 0.15872
#> 12    VA 0.14374
#> 13    WV 0.10026
#> 14    MD 0.08226
#> 15    CA 0.07912
#> 16    OH 0.06590
#> 17    NY 0.05802
#> 18    PA 0.05572
#> 19    FL 0.05490
#> 20    WY 0.04678
#> 21    NM 0.04650
#> 22    CT 0.04104
#> 23    OK 0.02036
#> 24    KY 0.00964
#> 25    AZ 0.00904
#> 26    ID 0.00748
#> 27    TX 0.00404
#> 28    CO 0.00282
#> 29    IA 0.00200
#> 30    NH 0.00180
#> 31    NC 0.00002
#> 32    HI 0.00002
#> 33    MN 0.00002
#> 34    RI 0.00000
#> 
#> $rosenthal
#>       t df          p
#> 1  1.19 40 0.12053081
#> 2  2.39 60 0.01000296
#> 3 -0.60 10 0.71907241
#> 4  1.52 30 0.06949071
#> 5  0.98 20 0.16939644
#> 
#> $teachexpect
#>  [1] 0.405 0.208 0.799 0.002 0.243 0.720 0.577 0.926 0.051 0.001 0.040 0.211 0.528 0.216 0.871 0.640
#> [17] 0.016 0.227 0.656
#> 
#> $validity
#>      n     r        p
#> 1   10  0.68 0.015223
#> 2   20  0.56 0.005117
#> 3   13  0.23 0.224837
#> 4   22  0.64 0.000669
#> 5   28  0.49 0.004063
#> 6   12 -0.04 0.549106
#> 7   12  0.49 0.052925
#> 8   36  0.33 0.024674
#> 9   19  0.58 0.004618
#> 10  12  0.18 0.287803
#> 11  36 -0.11 0.738475
#> 12  75  0.27 0.009563
#> 13  33  0.26 0.071971
#> 14 121  0.40 0.000003
#> 15  37  0.49 0.001040
#> 16  14  0.51 0.031221
#> 17  40  0.40 0.005274
#> 18  16  0.34 0.098791
#> 19  14  0.42 0.067441
#> 20  20  0.16 0.250210
#> 
#> $zhang
#>            study   smd    lo   hi ntreat ncont   phase  n        sd          z            p
#> 1   Giallauria13  0.67  0.07 1.26     25    21   acute 46 0.3035714  2.2070588 1.365498e-02
#> 2   Giallauria12  0.11 -0.44 0.67     24    26   acute 50 0.2831633  0.3884685 3.488347e-01
#> 3   Giallauria11  1.05  0.56 1.53     37    38   acute 75 0.2474490  4.2432990 1.101288e-05
#> 4        Chung10  0.05 -0.37 0.47     42    45   acute 87 0.2142857  0.2333333 4.077513e-01
#> 5   Giallauria09  1.16  0.61 1.70     30    30   acute 60 0.2780612  4.1717431 1.511392e-05
#> 6        Zheng08  0.70  0.18 1.22     30    30   acute 60 0.2653061  2.6384615 4.164157e-03
#> 7   Giallauria08  0.54  0.03 1.05     30    31   acute 61 0.2602041  2.0752941 1.897964e-02
#> 8        Brehm09  0.34 -0.36 1.03     25    12 healing 37 0.3545918  0.9588489 1.688174e-01
#> 9  Giallauria06a  0.04 -0.58 0.66     20    20 healing 40 0.3163265  0.1264516 4.496872e-01
#> 10 Giallauria06b  0.48 -0.12 1.08     22    22 healing 44 0.3061224  1.5680000 5.844057e-02
#> 11      Mimura05 -0.15 -0.86 0.57     15    15 healing 30 0.3647959 -0.4111888 6.595330e-01
#> 12          Yu04  0.50 -0.24 1.24     15    14 healing 29 0.3775510  1.3243243 9.269768e-02
#> 13        Kubo04 -0.22 -0.81 0.38     24    20 healing 44 0.3035714 -0.7247059 7.656838e-01
#> 14     Koizumi03  0.60 -0.15 1.35     14    15 healing 29 0.3826531  1.5680000 5.844057e-02
#> 15   Giannuzzi97  0.90  0.43 1.37     39    38 healing 77 0.2397959  3.7531915 8.729869e-05
#> 16         Lee08  0.16 -0.47 0.79     20    19  healed 39 0.3214286  0.4977778 3.093203e-01
#> 17      Heidal00  0.12 -0.53 0.77     19    18  healed 37 0.3316327  0.3618462 3.587335e-01
#> 18      Dubach97 -0.10 -0.89 0.68     12    13  healed 25 0.4005102 -0.2496815 5.985832e-01
#> 19   Gainnuzzi93  0.07 -0.33 0.47     49    46  healed 95 0.2040816  0.3430000 3.657992e-01
#> 20       Jette91 -0.58 -1.46 0.29     31    31  healed 62 0.4464286 -1.2992000 9.030623e-01
#> 21     Jugdutt88 -0.31 -0.99 0.36     13    24  healed 37 0.3443878 -0.9001481 8.159793e-01
#> 22  Grodzinski87 -0.30 -0.70 0.10     53    46  healed 99 0.2040816 -1.4700000 9.292191e-01
#>