Assessing the Function of House Sparrows' Bib Size Using a Flexible Meta-Analysis Method

A meta-analysis on the association between the size of a male's bib and their social status in house sparrows (Passer domesticus).

dat.nakagawa2007

Format

The data frame contains the following columns:

StudyID	character	identity of primary study

Place	character	location of study population

Correlation	numeric	correlation coefficient

SampleSize	integer	sample size of population

Details

Each study measures the association between a sparrows bib size and its social status. Effects are quantified as correlation coefficients.

Source

Nakagawa, S., Ockendon, N., Gillespie, D. O. S, Hatchwell, B. J., & Burke, T. (2007). Assessing the function of house sparrows' bib size using a flexible meta-analysis method. Behavioral Ecology, 18(5), 831--840. https://doi.org/10.1093/beheco/arm050

Author

Daniel Noble, daniel.noble@anu.edu.au

Examples

### copy data into 'dat' and examine data
dat <- dat.nakagawa2007
dat

#>      StudyID    Place Correlation SampleSize
#> 1  Sparrow01  Demark1       0.570         13
#> 2  Sparrow02  Demark2       0.520         10
#> 3  Sparrow03  Demark3       0.890         14
#> 4  Sparrow04 Hungary1       0.880         10
#> 5  Sparrow05 Hungary1       0.483         19
#> 6  Sparrow06  Norway1       0.320          9
#> 7  Sparrow07  Norway1       0.040          6
#> 8  Sparrow08  Norway2       0.330         11
#> 9  Sparrow09  Norway2       0.540          9
#> 10 Sparrow10  Norway3       0.330          9
#> 11 Sparrow11   Spain1       0.488         41
#> 12 Sparrow12     USA1       0.530         25
#> 13 Sparrow13     USA2       0.147         20
#> 14 Sparrow14     USA3       0.370         22
#> 15 Sparrow15     USA4       0.100         28

# \dontrun{

### load metafor package
library(metafor)

### calculate Zr
dat <- escalc(measure="ZCOR", ri=Correlation, ni=SampleSize, data=dat)

### fit meta-analytic model
res <- rma.mv(yi, vi, random = list(~ 1 | StudyID), data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 15; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed   factor 
#> sigma^2    0.0512  0.2263     15     no  StudyID 
#> 
#> Test for Heterogeneity:
#> Q(df = 14) = 22.7605, p-val = 0.0643
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.5205  0.0964  5.4020  <.0001  0.3317  0.7094  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }