Studies on Anti-Infective-Treated Central Venous Catheters for Prevention of Catheter-Related Bloodstream Infections

Results from 18 studies comparing the risk of catheter-related bloodstream infection when using anti-infective-treated versus standard catheters in the acute care setting.

dat.nielweise2007

Format

The data frame contains the following columns:

study	numeric	study number

author	character	(first) author

year	numeric	publication year

ai	numeric	number of CRBSIs in patients receiving an anti-infective catheter

n1i	numeric	number of patients receiving an anti-infective catheter

ci	numeric	number of CRBSIs in patients receiving a standard catheter

n2i	numeric	number of patients receiving a standard catheter

Details

The use of a central venous catheter may lead to a catheter-related bloodstream infection (CRBSI), which in turn increases the risk of morbidity and mortality. Anti-infective-treated catheters have been developed that are meant to reduce the risk of CRBSIs. Niel-Weise et al. (2007) conducted a meta-analysis of studies comparing infection risk when using anti-infective-treated versus standard catheters in the acute care setting. The results from 18 such studies are included in this dataset.

The dataset was used in the article by Stijnen et al. (2010) to illustrate various generalized linear mixed-effects models for the meta-analysis of proportions and odds ratios (see ‘References’).

Source

Niel-Weise, B. S., Stijnen, T., & van den Broek, P. J. (2007). Anti-infective-treated central venous catheters: A systematic review of randomized controlled trials. Intensive Care Medicine, 33(12), 2058–2068. https://doi.org/10.1007/s00134-007-0897-3

References

Stijnen, T., Hamza, T. H., & Ozdemir, P. (2010). Random effects meta-analysis of event outcome in the framework of the generalized linear mixed model with applications in sparse data. Statistics in Medicine, 29(29), 3046–3067. https://doi.org/10.1002/sim.4040

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, odds ratios, generalized linear models

Examples

### copy data into 'dat' and examine data
dat <- dat.nielweise2007
dat
#>    study         author year ai n1i ci n2i
#> 1      1           Bach 1996  0 116  3 117
#> 2      2         George 1997  1  44  3  35
#> 3      3           Maki 1997  2 208  9 195
#> 4      4           Raad 1997  0 130  7 136
#> 5      5          Heard 1998  5 151  6 157
#> 6      6         Collin 1999  1  98  4 139
#> 7      7         Hannan 1999  1 174  3 177
#> 8      8          Marik 1999  1  74  2  39
#> 9      9         Pierce 2000  1  97 19 103
#> 10    10          Sheng 2000  1 113  2 122
#> 11    11 Chatzinikolaou 2003  0  66  7  64
#> 12    12         Corral 2003  0  70  1  58
#> 13    13   Brun-Buisson 2004  3 188  5 175
#> 14    14           Leon 2004  6 187 11 180
#> 15    15          Yucel 2004  0 118  0 105
#> 16    16        Moretti 2005  0 252  1 262
#> 17    17           Rupp 2005  1 345  3 362
#> 18    18           Osma 2006  4  64  1  69

# \dontrun{

### load metafor package
library(metafor)

### standard (inverse-variance) random-effects model
res <- rma(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, drop00=TRUE)
#> Warning: Studies with NAs omitted from model fitting.
print(res, digits=3)
#> 
#> Random-Effects Model (k = 17; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.036 (SE = 0.299)
#> tau (square root of estimated tau^2 value):      0.189
#> I^2 (total heterogeneity / total variability):   3.46%
#> H^2 (total variability / sampling variability):  1.04
#> 
#> Test for Heterogeneity:
#> Q(df = 16) = 15.812, p-val = 0.466
#> 
#> Model Results:
#> 
#> estimate     se    zval   pval   ci.lb   ci.ub      
#>   -0.980  0.243  -4.027  <.001  -1.458  -0.503  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.38  0.23  0.60  0.21  0.69 
#> 

### random-effects conditional logistic model
res <- rma.glmm(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, model="CM.EL")
#> Warning: Studies with NAs omitted from model fitting.
#> Warning: Some yi/vi values are NA.
print(res, digits=3)
#> 
#> Random-Effects Model (k = 17; tau^2 estimator: ML)
#> Model Type: Conditional Model with Exact Likelihood
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.693 (SE = 0.640)
#> tau (square root of estimated tau^2 value):      0.833
#> I^2 (total heterogeneity / total variability):   41.133%
#> H^2 (total variability / sampling variability):  1.699
#> 
#> Tests for Heterogeneity:
#> Wld(df = 16) = 11.879, p-val = 0.752
#> LRT(df = 16) = 28.609, p-val = 0.027
#> 
#> Model Results:
#> 
#> estimate     se    zval   pval   ci.lb   ci.ub      
#>   -1.353  0.351  -3.854  <.001  -2.041  -0.665  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.26  0.13  0.51  0.04  1.52 
#> 

# }