Studies on the Effectiveness of Locoregional Treatment plus Chemotherapy for Head and Neck Squamous-Cell Carcinoma

Results from studies examining mortality risk in patients with nonmetastatic head and neck squamous-cell carcinoma receiving either locoregional treatment plus chemotherapy versus locoregional treatment alone.

dat.pignon2000

Format

The data frame contains the following columns:

id	numeric	study id number

trial	character	trial abbreviation

OmE	numeric	observed minus expected number of deaths in the locoregional treatment plus chemotherapy group

V	numeric	corresponding variance

grp	numeric	timing of chemotherapy: 1 = adjuvant, 2 = neoadjuvant, 3 = concomitant

Details

The purpose of this meta-analysis was to examine the mortality risk in patients with nonmetastatic head and neck squamous-cell carcinoma receiving either locoregional treatment plus chemotherapy versus locoregional treatment alone. For 65 trials, the dataset provides the observed minus expected number of deaths and corresponding variances in the locoregional treatment plus chemotherapy group. Based on these values, we can estimate the log hazard ratios with OmE/V and the corresponding sampling variance with 1/V.

The trials were also divided according to the timing of the chomotherapy: (1) adjuvant, after the locoregional treatment, (2) neoadjuvant, before the locoregional treatment, and (3) concomitant, chemotherapy given concomitantly or alternating with radiotherapy.

Source

Pignon, J. P., Bourhis, J., Domenge, C., & Designe, L. (2000). Chemotherapy added to locoregional treatment for head and neck squamous-cell carcinoma: Three meta-analyses of updated individual data. Lancet, 355(9208), 949--955. https://doi.org/10.1016/S0140-6736(00)90011-4

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

Examples

### copy data into 'dat' and examine data
dat <- dat.pignon2000
head(dat)

#>    id     trial  OmE    V grp
#> 1  44  Pitie-74 -1.8 14.5   1
#> 2  45 GETTECadj 11.8 57.1   1
#> 3  46  Int 0034 -7.4 80.6   1
#> 4 102    JHCFUS -5.6  8.5   1
#> 5  97   TMH R-4  0.4  6.4   1
#> 6 103    KKD-86  1.8  8.7   1

# \dontrun{

### load metafor package
require(metafor)

### calculate log hazard ratios and sampling variances
dat$yi <- with(dat, OmE/V)
dat$vi <- with(dat, 1/V)
head(dat)

#>    id     trial  OmE    V grp          yi         vi
#> 1  44  Pitie-74 -1.8 14.5   1 -0.12413793 0.06896552
#> 2  45 GETTECadj 11.8 57.1   1  0.20665499 0.01751313
#> 3  46  Int 0034 -7.4 80.6   1 -0.09181141 0.01240695
#> 4 102    JHCFUS -5.6  8.5   1 -0.65882353 0.11764706
#> 5  97   TMH R-4  0.4  6.4   1  0.06250000 0.15625000
#> 6 103    KKD-86  1.8  8.7   1  0.20689655 0.11494253

### meta-analysis based on all 65 trials
res <- rma(yi, vi, data=dat, method="FE", digits=2)
res

#> 
#> Fixed-Effects Model (k = 65)
#> 
#> I^2 (total heterogeneity / total variability):   48.25%
#> H^2 (total variability / sampling variability):  1.93
#> 
#> Test for Heterogeneity:
#> Q(df = 64) = 123.67, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.11  0.02  -4.67  <.01  -0.16  -0.06  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.90  0.85  0.94 
#> 

### only adjuvant trials
res <- rma(yi, vi, data=dat, method="FE", subset=grp==1, digits=2)
res

#> 
#> Fixed-Effects Model (k = 8)
#> 
#> I^2 (total heterogeneity / total variability):   10.56%
#> H^2 (total variability / sampling variability):  1.12
#> 
#> Test for Heterogeneity:
#> Q(df = 7) = 7.83, p-val = 0.35
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.02  0.07  -0.34  0.74  -0.16   0.11    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.98  0.85  1.12 
#> 

### only neoadjuvant trials
res <- rma(yi, vi, data=dat, method="FE", subset=grp==2, digits=2)
res

#> 
#> Fixed-Effects Model (k = 31)
#> 
#> I^2 (total heterogeneity / total variability):   4.96%
#> H^2 (total variability / sampling variability):  1.05
#> 
#> Test for Heterogeneity:
#> Q(df = 30) = 31.57, p-val = 0.39
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.06  0.03  -1.64  0.10  -0.12   0.01    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.95  0.88  1.01 
#> 

### only concomitant trials
res <- rma(yi, vi, data=dat, method="FE", subset=grp==3, digits=2)
res

#> 
#> Fixed-Effects Model (k = 26)
#> 
#> I^2 (total heterogeneity / total variability):   66.13%
#> H^2 (total variability / sampling variability):  2.95
#> 
#> Test for Heterogeneity:
#> Q(df = 25) = 73.81, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.21  0.04  -5.43  <.01  -0.28  -0.13  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.81  0.76  0.88 
#> 

# }