Studies on MYC-N as a Prognostic Marker for Neuroblastoma

Results from 81 studies examining overall and disease-free survival in neuroblastoma patients with amplified versus normal MYC-N protein levels.

dat.riley2003

Format

The data frame contains the following columns:

study	numeric	study number

yi	numeric	log hazard ratio of the outcome in those with amplified versus normal MYC-N protein levels

vi	numeric	sampling variance of the log hazard ratio

sei	numeric	standard error of the log hazard ratio

outcome	character	outcome (OS = overall survival; DFS = disease-free survival)

Details

The meta-analysis by Riley et al. (2003) examined a variety of prognostic markers for overall and disease-free survival in patients with neuroblastoma. One of the markers examined was amplified levels of the MYC-N protein, with is associated with poorer outcomes.

The dataset given here was extracted from Riley (2011) and has been used in several other publications (e.g., Riley et al., 2004, 2007). The dataset provides the (log) hazard ratios (and corresponding standard errors) with respect to these two outcomes in 81 studies, with positive values indicating a greater risk of death (for OS) or disease recurrence/death (for DFS) for patients with high MYC-N levels compared to those with normal/low levels. Note that information on both outcomes could only be extracted from 17 studies (39 studies only provided sufficient information to extract the OS estimate, while 25 studies only allowed for extraction of the DFS estimate).

Source

Riley, R. D., Sutton, A. J., Abrams, K. R., & Lambert, P. C. (2004). Sensitivity analyses allowed more appropriate and reliable meta-analysis conclusions for multiple outcomes when missing data was present. Journal of Clinical Epidemiology, 57(9), 911–924. https://doi.org/10.1016/j.jclinepi.2004.01.018

Riley, R. D., Abrams, K. R., Lambert, P. C., Sutton, A. J., & Thompson, J. R. (2007). An evaluation of bivariate random-effects meta-analysis for the joint synthesis of two correlated outcomes. Statistics in Medicine, 26(1), 78–97. https://doi.org/10.1002/sim.2524

Riley, R. D. (2011). Erratum: An evaluation of bivariate random-effects meta-analysis for the joint synthesis of two correlated outcomes. Statistics in Medicine, 30(4), 400. https://doi.org/10.1002/sim.4100

References

Riley, R. D., Burchill, S. A., Abrams, K. R., Heney, D., Lambert, P. C., Jones, D. R., Sutton, A. J., Young, B., Wailoo, A. J., & Lewis, I. J. (2003). A systematic review and evaluation of the use of tumour markers in paediatric oncology: Ewing's sarcoma and neuroblastoma. Health Technology Assessment, 7(5), 1–162. https://doi.org/10.3310/hta7050

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, oncology, hazard ratios

Examples

### copy data into 'dat' and examine data
dat <- dat.riley2003
dat
#> 
#>    study    yi   vi  sei outcome 
#> 1      1 -0.11 0.45 0.67     DFS 
#> 2      1 -0.14 0.66 0.81      OS 
#> 3      2  0.30 0.07 0.26     DFS 
#> 4      2  0.67 0.08 0.29      OS 
#> 5      3  0.41 0.77 0.88     DFS 
#> 6      3  0.43 0.66 0.81      OS 
#> 7      4  0.47 0.29 0.54     DFS 
#> 8      4  2.08 0.45 0.67      OS 
#> 9      5  0.76 0.24 0.49     DFS 
#> 10     5  0.70 0.31 0.56      OS 
#> 11     6  1.06 0.29 0.54     DFS 
#> 12     6  1.32 0.26 0.51      OS 
#> 13     7  1.46 0.17 0.41     DFS 
#> 14     7  1.51 0.23 0.48      OS 
#> 15     8  1.64 0.41 0.64     DFS 
#> 16     8  0.71 0.40 0.63      OS 
#> 17     9  1.64 0.41 0.64     DFS 
#> 18     9  2.59 1.08 1.04      OS 
#> 19    10  1.64 0.26 0.51     DFS 
#> 20    10  1.54 0.27 0.52      OS 
#> 21    11  1.70 0.15 0.39     DFS 
#> 22    11  1.38 0.14 0.37      OS 
#> 23    12  1.85 0.44 0.66     DFS 
#> 24    12  2.90 1.21 1.10      OS 
#> 25    13  1.90 0.21 0.46     DFS 
#> 26    13  1.83 0.22 0.47      OS 
#> 27    14  1.90 0.77 0.88     DFS 
#> 28    14  1.82 0.50 0.71      OS 
#> 29    15  2.19 0.18 0.42     DFS 
#> 30    15  2.99 0.26 0.51      OS 
#> 31    16  2.95 1.17 1.08     DFS 
#> 32    16  2.75 1.21 1.10      OS 
#> 33    17  5.70 2.99 1.73     DFS 
#> 34    17  5.70 2.99 1.73      OS 
#> 35    18  0.25 0.08 0.29     DFS 
#> 36    19  0.29 0.35 0.59     DFS 
#> 37    20  0.52 0.17 0.41     DFS 
#> 38    21  0.55 0.14 0.38     DFS 
#> 39    22  0.84 0.07 0.26     DFS 
#> 40    23  0.93 0.10 0.32     DFS 
#> 41    24  1.18 0.32 0.57     DFS 
#> 42    25  1.34 0.26 0.51     DFS 
#> 43    26  1.43 0.14 0.37     DFS 
#> 44    27  1.44 1.37 1.17     DFS 
#> 45    28  1.45 0.32 0.57     DFS 
#> 46    29  1.52 0.12 0.35     DFS 
#> 47    30  1.60 0.24 0.49     DFS 
#> 48    31  1.62 0.18 0.42     DFS 
#> 49    32  1.77 0.21 0.46     DFS 
#> 50    33  1.90 0.34 0.58     DFS 
#> 51    34  1.92 0.12 0.34     DFS 
#> 52    35  2.04 0.38 0.62     DFS 
#> 53    36  2.19 0.12 0.35     DFS 
#> 54    37  2.37 1.00 1.00     DFS 
#> 55    38  2.39 0.53 0.73     DFS 
#> 56    39  2.50 0.58 0.76     DFS 
#> 57    40  2.56 0.30 0.55     DFS 
#> 58    41  2.98 0.34 0.58     DFS 
#> 59    42  3.29 0.25 0.50     DFS 
#> 60    43 -0.84 0.72 0.85      OS 
#> 61    44  0.05 0.16 0.40      OS 
#> 62    45  0.73 0.50 0.71      OS 
#> 63    46  0.76 0.04 0.20      OS 
#> 64    47  0.91 0.44 0.66      OS 
#> 65    48  0.93 0.07 0.27      OS 
#> 66    49  0.96 0.22 0.47      OS 
#> 67    50  1.05 0.74 0.86      OS 
#> 68    51  1.16 1.39 1.18      OS 
#> 69    52  1.22 0.05 0.22      OS 
#> 70    53  1.26 0.24 0.49      OS 
#> 71    54  1.26 0.14 0.38      OS 
#> 72    55  1.27 1.64 1.28      OS 
#> 73    56  1.31 0.67 0.82      OS 
#> 74    57  1.52 0.21 0.46      OS 
#> 75    58  1.54 0.30 0.55      OS 
#> 76    59  1.55 0.49 0.70      OS 
#> 77    60  1.63 0.69 0.83      OS 
#> 78    61  1.67 1.28 1.13      OS 
#> 79    62  1.72 0.45 0.67      OS 
#> 80    63  1.74 0.20 0.45      OS 
#> 81    64  1.75 0.52 0.72      OS 
#> 82    65  1.75 0.41 0.64      OS 
#> 83    66  1.87 0.32 0.57      OS 
#> 84    67  2.07 0.48 0.69      OS 
#> 85    68  2.13 0.69 0.83      OS 
#> 86    69  2.19 0.01 0.12      OS 
#> 87    70  2.25 0.76 0.87      OS 
#> 88    71  2.31 0.25 0.50      OS 
#> 89    72  2.33 0.77 0.88      OS 
#> 90    73  2.36 0.32 0.57      OS 
#> 91    74  2.37 0.52 0.72      OS 
#> 92    75  2.63 0.56 0.75      OS 
#> 93    76  2.66 0.46 0.68      OS 
#> 94    77  2.77 1.21 1.10      OS 
#> 95    78  2.80 0.27 0.52      OS 
#> 96    79  3.33 0.50 0.71      OS 
#> 97    80  3.54 0.83 0.91      OS 
#> 98    81  5.04 1.21 1.10      OS 
#> 

### load metafor package
library(metafor)

### random-effects model analysis for outcome DFS
res <- rma(yi, sei=sei, data=dat, subset=(outcome == "DFS"), method="DL")
res
#> 
#> Random-Effects Model (k = 42; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.4212 (SE = 0.1561)
#> tau (square root of estimated tau^2 value):      0.6490
#> I^2 (total heterogeneity / total variability):   65.98%
#> H^2 (total variability / sampling variability):  2.94
#> 
#> Test for Heterogeneity:
#> Q(df = 41) = 120.5237, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub      
#>   1.4843  0.1303  11.3876  <.0001  1.2288  1.7398  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  4.41  3.42  5.70  1.21 16.15 
#> 

### random-effects model analysis for outcome OS
res <- rma(yi, sei=sei, data=dat, subset=(outcome == "OS"), method="DL")
res
#> 
#> Random-Effects Model (k = 56; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.4223 (SE = 0.1889)
#> tau (square root of estimated tau^2 value):      0.6499
#> I^2 (total heterogeneity / total variability):   65.36%
#> H^2 (total variability / sampling variability):  2.89
#> 
#> Test for Heterogeneity:
#> Q(df = 55) = 158.7926, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub      
#>   1.6356  0.1219  13.4193  <.0001  1.3967  1.8745  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  5.13  4.04  6.52  1.40 18.76 
#>