Studies on Nedocromil Sodium for Preventing Exercise-Induced Bronchoconstriction

Results from 17 trials, 11 studies in children and 6 studies in adults, reporting the maximum fall in the forced expiratory volume in 1 second (FEV_1) over the course of follow-up, expressed as a percentage.

dat.spooner2002

Format

The data frame contains the following columns:

author	character	first author

year	character	year of publication

Ne	integer	number of participants in nedocromil sodium group

Me	numeric	maximum fall in the FEV_1 (nedocromil sodium)

Se	numeric	standard deviation (nedocromil sodium)

Nc	integer	number of participants in placebo group

Mc	numeric	maximum fall in the FEV_1 (placebo)

Sc	numeric	standard deviation (placebo)

agegroup	factor	age group (children or adults)

order	numeric	order of studies in Table 1.1 in the Cochrane review

Details

Spooner et al. (2002) conducted a Cochrane review comparing nedocromil sodium (experimental treatment) with placebo (control) for preventing exercise-induced bronchoconstriction. Primary outcome was the maximum fall in the forced expiratory volume in 1 second (FEV_1) over the course of follow-up, expressed as a percentage. This outcome is available for 17 studies, 11 studies in children and 6 studies in adults. For each study, the mean value, standard deviation, and sample size are reported for both the experimental and control group. The authors conducted a random-effects meta-analysis with the mean difference as effect measure, i.e.\ mean value in the nedocromil sodium group minus mean value in the placebo group.

This data set is used as an example in Schwarzer et al. (2015).

Source

Spooner, C., Saunders, L. D., & Rowe, B. H. (2002). Nedocromil sodium for preventing exercise‐induced bronchoconstriction. Cochrane Database of Systematic Reviews, 1, CD001183. https://doi.org/10.1002/14651858.CD001183

References

Schwarzer, G., Carpenter, J. R., & Rücker, G. (2015). Meta-analysis with R. Cham, Switzerland: Springer.

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

medicine, pulmonology, raw mean differences, subgroup analysis

Examples

### Show first five studies
head(dat.spooner2002, 5)
#>    author year Ne    Me    Se Nc    Mc    Sc agegroup order
#> 1   Boner 1988 13 13.54 13.85 13 20.77 21.46 Children     1
#> 2   Boner 1989 20 15.70 13.10 20 22.70 16.47 Children     7
#> 3  Chudry 1987 12 21.30 13.10 12 39.70 12.90 Children     9
#> 4   Comis 1993 12 14.50 12.20 12 31.30 15.10 Children    10
#> 5 Debelic 1986 12 29.83 15.95 12 48.08 15.08 Children    11

### Load meta package
suppressPackageStartupMessages(library(meta))

### Use settings from RevMan5
oldset <- settings.meta("RevMan5")

### Conduct random effects meta-analysis with age subgroups
mc1 <- metacont(Ne, Me, Se, Nc, Mc, Sc,
                data = dat.spooner2002, studlab = paste(author, year),
                subgroup = agegroup, print.subgroup.name = FALSE,
                label.e = "Nedocromil sodium", label.c = "Placebo",
                common = FALSE)
mc1
#> Number of studies: k = 17
#> Number of observations: o = 480 (o.e = 240, o.c = 240)
#> 
#>                          MD           95%-CI      z  p-value
#> Random effects model -15.64 [-18.14, -13.15] -12.30 < 0.0001
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 2.437 [0.000, 40.900]; tau = 1.5612 [0.0000, 6.3953]
#>  I^2 = 9% [0%, 45%]; H = 1.05 [1.00, 1.35]
#> 
#> Test of heterogeneity:
#>      Q d.f. p-value
#>  17.57   16  0.3496
#> 
#> Results for subgroups (random effects model):
#>            k     MD           95%-CI tau^2    tau     Q I^2
#> Children  11 -14.81 [-18.16, -11.45] 3.589 1.8944 11.25 11%
#> Adults     6 -16.89 [-20.81, -12.97] 3.127 1.7683  5.71 12%
#> 
#> Test for subgroup differences (random effects model):
#>                   Q d.f. p-value
#> Between groups 0.63    1  0.4287
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q

### Use previous settings
settings.meta(oldset)