Studies on the Relationship between Sexual Signal Expression and Individual Quality

Results from 41 studies examining the relationship between measures of individual quality and the expression of structurally coloured sexual signals.

dat.white2020

Format

The object is a data frame which contains the following columns:

study_id	character	study-level ID

obs	character	observation-level ID

exp_obs	character	whether the study is observational or experimental

control	numeric	whether the study did (1) or did not (0) include a non-sexual control trait

class	character	class of the study organisms

genus	character	class of the study organisms

species	character	species of the study organisms

sex	character	sex of the study organisms

iridescent	numeric	whether the colour signals were iridescent (1) or not (0)

col_var	character	the colour variable quantified

col_component	character	whether the colour variable is chromatic or achromatic

quality_measure	character	the measure of individual quality used

region	character	the body region from which colour was sampled

n	numeric	study sample size

r	numeric	Pearson's correlation coefficient

Details

The 186 rows in this dataset come from 41 experimental and observational studies reporting on the correlation between measures of individual quality (age, body condition, immune function, parasite resistance) and the expression of structurally coloured sexual signals across 28 species. The purpose of this meta-analysis was to test whether structural colour signals show heightened condition-dependent expression, as predicted by evolutionary models of 'honest' signalling.

Source

White, T. E. (2020). Structural colours reflect individual quality: A meta-analysis. Biology Letters, 16(4), 20200001. https://doi.org/10.1098/rsbl.2020.0001

Author

Thomas E. White, thomas.white@sydney.edu.au

Examples

### copy data into 'dat' and examine data
dat <- dat.white2020
head(dat, 10)

#>    study_id  obs exp_obs control   class        genus                    species    sex iridescent
#> 1      p105 e001     exp       0 insecta pseudomantis Pseudomantis_albofimbriata female          0
#> 2       p11 e002     obs       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 3       p11 e003     obs       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 4       p11 e004     obs       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 5       p11 e005     obs       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 6       p11 e006     obs       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 7       p11 e007     obs       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 8       p11 e008     exp       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 9       p11 e009     exp       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#> 10      p11 e010     exp       0    aves   Cyanocitta        Cyanocitta_stelleri   male          0
#>       col_var col_component quality_measure  region  n      r
#> 1  brightness    achromatic       condition abdomen 50  0.437
#> 2         hue     chromatic       condition    wing 22  0.410
#> 3      chroma     chromatic       condition    wing 22 -0.330
#> 4  brightness    achromatic       condition    wing 22 -0.130
#> 5         hue     chromatic        parasite    wing 59 -0.080
#> 6      chroma     chromatic        parasite    wing 59 -0.310
#> 7  brightness    achromatic        parasite    wing 59 -0.120
#> 8         hue     chromatic        parasite    wing  9  0.150
#> 9      chroma     chromatic        parasite    wing  9 -0.090
#> 10 brightness    achromatic        parasite    wing 18  0.270

# \dontrun{

### load metafor package
require(metafor)

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=r, ni=n, data=dat)

### fit multilevel meta-analytic model
res <- rma.mv(yi, vi, random = list(~ 1 | study_id, ~ 1 | obs), data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 186; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed    factor 
#> sigma^2.1  0.0153  0.1237     41     no  study_id 
#> sigma^2.2  0.0649  0.2548    186     no       obs 
#> 
#> Test for Heterogeneity:
#> Q(df = 185) = 759.2939, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.1573  0.0329  4.7742  <.0001  0.0927  0.2218  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }