dat.woods2010 {metadat}	R Documentation

Studies on Treatments for Chronic Obstructive Pulmonary Disease

Description

Results from 3 trials examining the mortality risk of three treatments and placebo in patients with chronic obstructive pulmonary disease.

Usage


dat.woods2010

Format

The data frame contains the following columns:

author 	 character 	 first author / study name

treatment 	 character 	 treatment

r 	 integer 	 number of deaths

N 	 integer 	 number of patients

Details

Count mortality statistics in randomised controlled trials of treatments for chronic obstructive pulmonary disease (Woods et al., 2010, Table 1).

Concepts

medicine, odds ratios, network meta-analysis

Author(s)

Guido Schwarzer, sc@imbi.uni-freiburg.de, https://github.com/guido-s/

Source

Woods, B. S., Hawkins, N., & Scott, D. A. (2010). Network meta-analysis on the log-hazard scale, combining count and hazard ratio statistics accounting for multi-arm trials: A tutorial. BMC Medical Research Methodology, 10, 54. https://doi.org/10.1186/1471-2288-10-54

Examples


### Show full data set
dat.woods2010

## Not run: 

### Load netmeta package
suppressPackageStartupMessages(library(netmeta))

### Print odds ratios and confidence limits with two digits
settings.meta(digits = 2)

### Change appearance of confidence intervals
cilayout("(", "-")

### Transform data from long arm-based format to contrast-based
### format. Argument 'sm' has to be used for odds ratio as summary
### measure; by default the risk ratio is used in the metabin function
### called internally.
pw <- pairwise(treatment, event = r, n = N,
  studlab = author, data = dat.woods2010, sm = "OR")
pw

### Conduct network meta-analysis
net <- netmeta(pw)
net

### Show forest plot
forest(net, ref = "Placebo", drop = TRUE,
  leftlabs = "Contrast to Placebo")


## End(Not run)

[Package metadat version 1.2-0 Index]