dat.aloe2013 {metadat}	R Documentation

Studies on the Association Between Supervision Quality and Various Outcomes in Social, Mental Health, and Child Welfare Workers

Description

Results from 5 studies examining the association between various measures of supervision quality and various work-related outcomes in social, mental health, and child welfare workers.

Usage


dat.aloe2013

Format

The data frame contains the following columns:

study 	 character 	 study author(s) and year

n 	 integer 	 sample size

tval 	 numeric 	 t-statistic for the test of the association/predictor

preds 	 integer 	 number of predictors included in the regression model

R2 	 numeric 	 the coefficient of determination (i.e., R-squared value) of the regression model

Details

The dataset is based on studies that used regression models to examine the association between some measure of perceived supervision quality (e.g., the quality of the relationship with one's supervisor) and some work-related outcome (e.g., job satisfaction) in social, mental health, and child welfare workers. The dataset was extracted from Aloe and Thompson (2013), which in turn is a subset of the studies included in the meta-analysis by Mor Barak et al. (2009).

The dataset can be used to illustrate the meta-analysis of regression models, using measures such as the (semi-)partial correlation coefficient. For this, the t-statistic from the regression model for the association (i.e., predictor) of interest was extracted from each regression model (tval), as well as the sample size (n), the number of predictors included in the regression model (preds), and the coefficient of determination (i.e., R-squared value) of the regression model (R2). Based on this information, the (semi-)partial correlation coefficient can be computed for each study, as well as its corresponding sampling variance. These values can then be meta-analyzed using standard methods.

Concepts

social work, (semi-)partial correlations, meta-regression

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Aloe, A. M., & Thompson, C. G. (2013). The synthesis of partial effect sizes. Journal of the Society for Social Work and Research, 4(4), 390–405. https://doi.org/10.5243/jsswr.2013.24

References

Mor Barak, M. E., Travis, D. J., Pyun, H., & Xie, B. (2009). The impact of supervision on worker outcomes: A meta-analysis. Social Service Review, 83(1), 3–32. https://doi.org/10.1086/599028

Examples


### copy data into 'dat' and examine data
dat <- dat.aloe2013
dat

## Not run: 

### load metafor package
suppressPackageStartupMessages(library(metafor))

### compute the partial correlation coefficients and corresponding sampling variances
dat <- escalc(measure="PCOR", ti=tval, ni=n, mi=preds, data=dat)
dat

### random-effects model
res <- rma(yi, vi, data=dat)
res

### mixed-effects meta-regression model examining the relationship between the partial
### correlation coefficients and the number of predictors included in the models
res <- rma(yi, vi, mods = ~ preds, data=dat)
res

### compute the r-to-z transformed partial correlation coefficients and their variances
dat <- escalc(measure="ZPCOR", ti=tval, ni=n, mi=preds, data=dat)
dat

### random-effects model
res <- rma(yi, vi, data=dat)
res

### back-transformation to the partial correlation scale
predict(res, transf=transf.ztor)

### compute the semi-partial correlation coefficients and their variances
dat <- escalc(measure="SPCOR", ti=tval, ni=n, mi=preds, r2i=R2, data=dat)
dat

### random-effects model
res <- rma(yi, vi, data=dat)
res


## End(Not run)

[Package metadat version 1.2-0 Index]