dat.anand1999 {metadat}	R Documentation

Studies on the Effectiveness of Oral Anticoagulants in Patients with Coronary Artery Disease

Description

Results from 34 trials examining the effectiveness of oral anticoagulants in patients with coronary artery disease.

Usage


dat.anand1999

Format

The data frame contains the following columns:

study 	 character 	 author(s) or trial name

year 	 numeric 	 publication year

intensity 	 character 	 intensity of anticoagulation (low, medium, or high)

asp.t 	 numeric 	 concomitant use of aspirin in the treatment group (0 = no, 1 = yes)

asp.c 	 numeric 	 concomitant use of aspirin in the control group (0 = no, 1 = yes)

ai 	 numeric 	 number of deaths in the treatment group

n1i 	 numeric 	 number of patients in the treatment group

ci 	 numeric 	 number of deaths in the control group

n2i 	 numeric 	 number of patients in the control group

Details

The dataset includes the results from 34 randomized clinical trials that examined the effectiveness of oral anticoagulants in patients with coronary artery disease. The results given here are focused on the total mortality in the treatment versus control groups.

Concepts

medicine, cardiology, odds ratios, Mantel-Haenszel method

Note

Strictly speaking, there are only 31 trials, since Breddin et al. (1980) and ATACS (1990) are multiarm trials.

According to a correction, dat.anand1999$ci[29] should be 1. But then dat.anand1999$ci[21] would also have to be 1 (if these data indeed refer to the same control group). This appears contradictory, so this correction was not made.

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Anand, S. S., & Yusuf, S. (1999). Oral anticoagulant therapy in patients with coronary artery disease: A meta-analysis. Journal of the American Medical Association, 282(21), 2058–2067. https://doi.org/10.1001/jama.282.21.2058

Examples


### copy data into 'dat' and examine data
dat <- dat.anand1999
dat

## Not run: 

### load metafor package
library(metafor)

### High-Intensity OA vs Control
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity=="high" & asp.t==0 & asp.c==0), digits=2)

### High- or Moderate-Intensity OA vs Aspirin
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity %in% c("high","moderate") & asp.t==0 & asp.c==1), digits=2)

### Moderate-Intensity OA vs Control
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity=="moderate" & asp.t==0 & asp.c==0), digits=2)

### High- or Moderate-Intensity OA and Aspirin vs Aspirin
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity %in% c("high","moderate") & asp.t==1 & asp.c==1), digits=2)

### Low-Intensity OA and Aspirin vs Aspirin
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat,
       subset=(intensity=="low" & asp.t==1 & asp.c==1), digits=2)


## End(Not run)

[Package metadat version 1.2-0 Index]