Studies on the Effectiveness of Writing-to-Learn Interventions

Results from 48 studies on the effectiveness of school-based writing-to-learn interventions on academic achievement.

dat.bangertdrowns2004

Format

The data frame contains the following columns:

id	numeric	study number

author	character	study author(s)

year	numeric	publication year

grade	numeric	grade level (1 = elementary; 2 = middle; 3 = high-school; 4 = college)

length	numeric	treatment length (in weeks)

minutes	numeric	minutes per assignment

wic	numeric	writing tasks were completed in class (0 = no; 1 = yes)

feedback	numeric	feedback on writing was provided (0 = no; 1 = yes)

info	numeric	writing contained informational components (0 = no; 1 = yes)

pers	numeric	writing contained personal components (0 = no; 1 = yes)

imag	numeric	writing contained imaginative components (0 = no; 1 = yes)

meta	numeric	prompts for metacognitive reflection (0 = no; 1 = yes)

subject	character	subject matter

ni	numeric	total sample size of the study

yi	numeric	standardized mean difference

vi	numeric	corresponding sampling variance

Details

In each of the studies included in this meta-analysis, an experimental group (i.e., a group of students that received instruction with increased emphasis on writing tasks) was compared against a control group (i.e., a group of students that received conventional instruction) with respect to some content-related measure of academic achievement (e.g., final grade, an exam/quiz/test score). The outcome measure for this meta-analysis was the standardized mean difference (with positive values indicating a higher mean level of academic achievement in the intervention group).

The standardized mean differences given here are bias-corrected and therefore differ slightly from the values reported in the article. Also, since only the total sample size is given in the article, the sampling variances were computed under the assumption that \(n_{i1} = n_{i2} = n_i / 2\).

Source

Bangert-Drowns, R. L., Hurley, M. M., & Wilkinson, B. (2004). The effects of school-based writing-to-learn interventions on academic achievement: A meta-analysis. Review of Educational Research, 74(1), 29–58. https://doi.org/10.3102/00346543074001029

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

education, standardized mean differences, meta-regression

Examples

### copy data into 'dat' and examine data
dat <- dat.bangertdrowns2004
dat[1:10,-13]
#> 
#>    id      author year grade length minutes wic feedback info pers imag meta  ni     yi    vi 
#> 1   1    Ashworth 1992     4     15      NA   1        1    1    1    0    1  60  0.650 0.070 
#> 2   2       Ayers 1993     2     10      NA   1       NA    1    1    1    0  34 -0.750 0.126 
#> 3   3      Baisch 1990     2      2      NA   1        0    1    1    0    1  95 -0.210 0.042 
#> 4   4       Baker 1994     4      9      10   1        1    1    0    0    0 209 -0.040 0.019 
#> 5   5      Bauman 1992     1     14      10   1        1    1    1    0    1 182  0.230 0.022 
#> 6   6      Becker 1996     4      1      20   1        0    0    1    0    0 462  0.030 0.009 
#> 7   7 Bell & Bell 1985     3      4      NA   1        1    1    1    0    1  38  0.260 0.106 
#> 8   8     Brodney 1994     1     15      NA   1        1    1    1    0    1 542  0.060 0.007 
#> 9   9      Burton 1986     4      4      NA   0        1    1    0    0    0  99  0.060 0.040 
#> 10 10   Davis, BH 1990     1      9      10   1        0    1    1    0    0  77  0.120 0.052 
#> 

### load metafor package
library(metafor)

### fit random-effects model
res <- rma(yi, vi, data=dat)
res
#> 
#> Random-Effects Model (k = 48; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0499 (SE = 0.0197)
#> tau (square root of estimated tau^2 value):      0.2235
#> I^2 (total heterogeneity / total variability):   58.37%
#> H^2 (total variability / sampling variability):  2.40
#> 
#> Test for Heterogeneity:
#> Q(df = 47) = 107.1061, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub      
#>   0.2219  0.0460  4.8209  <.0001  0.1317  0.3122  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### some examples of mixed-effects meta-regression models
res <- rma(yi, vi, mods = ~ factor(grade), data=dat)
res
#> 
#> Mixed-Effects Model (k = 48; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0539 (SE = 0.0216)
#> tau (square root of estimated tau^2 value):             0.2322
#> I^2 (residual heterogeneity / unaccounted variability): 59.15%
#> H^2 (unaccounted variability / sampling variability):   2.45
#> R^2 (amount of heterogeneity accounted for):            0.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 44) = 102.0036, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:4):
#> QM(df = 3) = 5.9748, p-val = 0.1128
#> 
#> Model Results:
#> 
#>                 estimate      se     zval    pval    ci.lb    ci.ub     
#> intrcpt           0.2639  0.0898   2.9393  0.0033   0.0879   0.4399  ** 
#> factor(grade)2   -0.3727  0.1705  -2.1856  0.0288  -0.7069  -0.0385   * 
#> factor(grade)3    0.0248  0.1364   0.1821  0.8555  -0.2425   0.2922     
#> factor(grade)4   -0.0155  0.1160  -0.1338  0.8935  -0.2429   0.2118     
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
res <- rma(yi, vi, mods = ~ length, data=dat)
#> Warning: Studies with NAs omitted from model fitting.
res
#> 
#> Mixed-Effects Model (k = 46; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0441 (SE = 0.0188)
#> tau (square root of estimated tau^2 value):             0.2100
#> I^2 (residual heterogeneity / unaccounted variability): 55.26%
#> H^2 (unaccounted variability / sampling variability):   2.24
#> R^2 (amount of heterogeneity accounted for):            5.08%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 44) = 96.2810, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 4.2266, p-val = 0.0398
#> 
#> Model Results:
#> 
#>          estimate      se    zval    pval    ci.lb   ci.ub    
#> intrcpt    0.0692  0.0825  0.8384  0.4018  -0.0925  0.2309    
#> length     0.0149  0.0073  2.0559  0.0398   0.0007  0.0292  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
res <- rma(yi, vi, mods = ~ info + pers + imag + meta, data=dat)
#> Warning: Studies with NAs omitted from model fitting.
res
#> 
#> Mixed-Effects Model (k = 46; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0412 (SE = 0.0192)
#> tau (square root of estimated tau^2 value):             0.2030
#> I^2 (residual heterogeneity / unaccounted variability): 51.71%
#> H^2 (unaccounted variability / sampling variability):   2.07
#> R^2 (amount of heterogeneity accounted for):            18.34%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 41) = 82.7977, p-val = 0.0001
#> 
#> Test of Moderators (coefficients 2:5):
#> QM(df = 4) = 10.0061, p-val = 0.0403
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb   ci.ub     
#> intrcpt    0.3440  0.2179   1.5784  0.1145  -0.0831  0.7711     
#> info      -0.1988  0.2106  -0.9438  0.3453  -0.6115  0.2140     
#> pers      -0.3223  0.1782  -1.8094  0.0704  -0.6715  0.0268   . 
#> imag       0.2540  0.2023   1.2557  0.2092  -0.1424  0.6504     
#> meta       0.4817  0.1678   2.8708  0.0041   0.1528  0.8106  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>