dat.besson2016 {metadat}	R Documentation

Dataset on How Maternal Diet Impacts Copying Styles in Rodents

Description

Results from 46 studies synthesising maternal nutritional effects on coping styles in rodents.

Usage


dat.besson2016

Format

The data frame contains the following columns:

comp_ID 	 character 	 effect-size unique identifier

study_ID 	 character 	 study unique identifier

dam_ID 	 character 	 dam unique identifier (group of dams subjected to the same treatment)

animal_ID 	 character 	 offspring unique identifier (group of offspring from the same dam group subjected to the same treatment)

Reference 	 character 	 author’s names and date

species 	 character 	 species [rats or mice]

strain 	 character 	 strain

manip_type 	 character 	 maternal nutritional manipulation type [protein or calorie]

manip_direction 	 character 	 direction of maternal nutritional manipulation [- = restriction, + = overfeeding]

nom_manip_val 	 character 	 degree of maternal nutritional manipulation as described in the original publications [% = percentage of caloric or protein restriction, # = increase in caloric intake]

exp 	 character 	 percentage of caloric or protein maternal restriction or increase in caloric intake of the experimental group

control 	 character 	 percentage of caloric or protein maternal restriction or increase in caloric intake for the control group

manip_parameter 	 character 	 protein content, percentage fat or intake

vitmin_eql 	 character 	 were vitamins equalized across maternal diets? [yes or no]

adlib_con 	 character 	 were maternal control groups fed ad libitum? [yes or no]

adlib_exp 	 character 	 were maternal experimental groups fed ad libitum? [yes or no]

diet_con 	 character 	 name of maternal control diet?

diet_exp 	 character 	 name of maternal experimental diet?

dam_diet_start_dPC 	 numeric 	 start of the dam diet [in days post-conception]

dam_diet_end_dPC 	 numeric 	 end of the dam diet [in days post-conception]

diet_label 	 character 	 period of maternal diet manipulation [pregestation = pre-gestation, pre = pregnancy, lact = lactation, or pre+lact = pregnancy and lactation]

age_mating 	 numeric 	 dam age at mating if known

n_con_dam 	 integer 	 sample size of the control dam groups

n_exp_dam 	 integer 	 sample size of the experimental dam groups

multi_use_con 	 character 	 were control groups used multiple time? [yes or no]

dam_housing 	 character 	 how were dams housed? [pair, group, or single]

temperature 	 numeric 	 temperature during the experiment [°C]

photoperiod 	 integer 	 photoperiod during the experiment [number of hours of light]

litter_size 	 integer 	 size of the litter [number of pups per dam]

litter_size_equalized 	 character 	 has litter size been equalized? [yes or no]

crossfostered 	 character 	 have pups been cross-fostered? [yes or no]

sex 	 character 	 sex of the offspring that were tested [m = male, f = female, both = mixed sex]

housing 	 character 	 offspring housing during the test period [dam, pair, single, or group]

bodymass_mean_contr 	 numeric 	 mean body mass of control offspring close to or during the testing period [g]

bodymass_SE_contr 	 numeric 	 S.E. for body mass of control offspring close to or during the testing period

bodymass_mean_exp 	 numeric 	 mean body mass of experimental offspring close to or during the testing period [g]

bodymass_SE_exp 	 numeric 	 S.E. for body mass of experimental offspring close to or during the testing period

bm_N_contr 	 integer 	 sample size for body mass of control offspring close to or during the testing period

bm_N_exp 	 integer 	 sample size for body mass of experimental offspring close to or during the testing period

bm_dPP 	 integer 	 age of offspring when body mass was measured [in days post-parturition]

offspring_diet 	 character 	 offspring diet after weaning [type of control diet]

offspring_con_adlib 	 character 	 were control offspring fed ad libitum after weaning? [yes or no]

offspring_diet_level 	 character 	 name of offspring diet after weaning

offspring_diet_end_dPP 	 integer 	 end of the offspring diet [in days post-parturition]

post_diet_adlib 	 character 	 were experimental offspring fed ad libitum after weaning? [yes or no]

response_age_dPP 	 numeric 	 offspring age when behavioural testing started [in days post-parturition]

authors_behaviour_classification 	 character 	 author's classification of offspring behaviour [anxiety, exploration, or activity]

our_behaviour_classification 	 character 	 our classification of offspring behaviour [anxiety, exploration, or activity]

response_test 	 character 	 type of test used [elevated T-maze (ETM), open field, etc.] to measure offspring behaviour

time_trial 	 integer 	 duration of the testing [min]

measure 	 character 	 measures taken during testing [total distance moved, time spent in open arm, etc.]

unit 	 character 	 unit of the behavioural measure taken [min, s, m, number (#), etc.]

high_better 	 character 	 for activity and exploration, a higher number is assumed to be better (i.e., animals were more active), but the opposite was assumed for anxiety (i.e., they were more anxious) [yes or no]

night.day 	 character 	 time of day when behaviours were measured [night or day]

comparison 	 character 	 for a given control-treatment group comparison, animal group codes as used in the original article [e.g., LP, HP]. This field allows identification of exactly which data (i.e., comparison of which pairs of groups) were extracted from the original paper, and is not used in our analyses. For our analyses the groups were re-coded as control/experimental.

exp_mean 	 numeric 	 mean of the offspring behaviour measured for the experimental group

exp_se 	 numeric 	 S.E. of the offspring behaviour measured for the experimental group

exp_n 	 integer 	 sample size for the offspring experimental group

con_mean 	 numeric 	 mean of offspring behaviour measured for the control group

con_se 	 numeric 	 S.E. of the offspring behaviour measured for the control group

con_n 	 integer 	 sample size for the offspring control group

con_ID 	 character 	 identifier for shared control groups within experiment

percentage 	 character 	 is the offspring behaviour measure a percentage? [yes or no]

Data_source 	 character 	 figure or table number in the original paper from which the data were extracted

measure_comments 	 character 	 any comments on the offspring behaviour measures

SE_imputed 	 character 	 was S.E. imputed for the offspring behaviour measure? [yes or no]

Comments 	 character 	 any comments on the data

Details

Data from experiments where dams were subject to caloric or protein restriction or were overfed around gestation were included. Offspring activity, exploration, or anxiety were measured outcomes variables from maternal experimental treatments. Multilevel meta-analysis and meta-regression models were used to analyze the meta-analytic data.

Concepts

ecology, evolution, standardized mean differences

Author(s)

Daniel Noble, daniel.noble@anu.edu.au

Source

Besson, A. A., Lagisz, M., Senior, A. M., Hector, K. L., & Nakagawa, S. (2016). Effect of maternal diet on offspring coping styles in rodents: A systematic review and meta-analysis. Biological Reviews, 91(4), 1065–1080. https://doi.org/10.1111/brv.12210

Examples


### copy data into 'dat' and examine data
dat <- dat.besson2016
head(dat)

## Not run: 

### load metafor
library(metafor)

### compute SD from SE
dat$sd_c <- with(dat, con_se * sqrt(con_n))
dat$sd_e <- with(dat, exp_se * sqrt(exp_n))

### compute standardized mean differences and corresponding sampling variances
dat <- escalc(measure="SMD", m1i=exp_mean, m2i=con_mean, sd1i=sd_e, sd2i=sd_c,
              n1i=exp_n, n2i=con_n, data=dat, add.measure=TRUE)

### fit model
mod1 <- rma.mv(yi ~ 1, V = vi, random = list(~ 1 | study_ID, ~ 1 | comp_ID), data = dat)
mod1


## End(Not run)

[Package metadat version 1.2-0 Index]