Dataset on How Maternal Diet Impacts Copying Styles in Rodents

Results from 46 studies synthesising maternal nutritional effects on coping styles in rodents.

dat.besson2016

Format

The data frame contains the following columns:

comp_ID	character	effect-size unique identifier

study_ID	character	study unique identifier

dam_ID	character	dam unique identifier (group of dams subjected to the same treatment)

animal_ID	character	offspring unique identifier (group of offspring from the same dam group subjected to the same treatment)

Reference	character	author’s names and date

species	character	species [rats or mice]

strain	character	strain

manip_type	character	maternal nutritional manipulation type [protein or calorie]

manip_direction	character	direction of maternal nutritional manipulation [- = restriction, + = overfeeding]

nom_manip_val	character	degree of maternal nutritional manipulation as described in the original publications [% = percentage of caloric or protein restriction, # = increase in caloric intake]

exp	character	percentage of caloric or protein maternal restriction or increase in caloric intake of the experimental group

control	character	percentage of caloric or protein maternal restriction or increase in caloric intake for the control group

manip_parameter	character	protein content, percentage fat or intake

vitmin_eql	character	were vitamins equalized across maternal diets? [yes or no]

adlib_con	character	were maternal control groups fed ad libitum? [yes or no]

adlib_exp	character	were maternal experimental groups fed ad libitum? [yes or no]

diet_con	character	name of maternal control diet?

diet_exp	character	name of maternal experimental diet?

dam_diet_start_dPC	numeric	start of the dam diet [in days post-conception]

dam_diet_end_dPC	numeric	end of the dam diet [in days post-conception]

diet_label	character	period of maternal diet manipulation [pregestation = pre-gestation, pre = pregnancy, lact = lactation, or pre+lact = pregnancy and lactation]

age_mating	numeric	dam age at mating if known

n_con_dam	integer	sample size of the control dam groups

n_exp_dam	integer	sample size of the experimental dam groups

multi_use_con	character	were control groups used multiple time? [yes or no]

dam_housing	character	how were dams housed? [pair, group, or single]

temperature	numeric	temperature during the experiment [°C]

photoperiod	integer	photoperiod during the experiment [number of hours of light]

litter_size	integer	size of the litter [number of pups per dam]

litter_size_equalized	character	has litter size been equalized? [yes or no]

crossfostered	character	have pups been cross-fostered? [yes or no]

sex	character	sex of the offspring that were tested [m = male, f = female, both = mixed sex]

housing	character	offspring housing during the test period [dam, pair, single, or group]

bodymass_mean_contr	numeric	mean body mass of control offspring close to or during the testing period [g]

bodymass_SE_contr	numeric	S.E. for body mass of control offspring close to or during the testing period

bodymass_mean_exp	numeric	mean body mass of experimental offspring close to or during the testing period [g]

bodymass_SE_exp	numeric	S.E. for body mass of experimental offspring close to or during the testing period

bm_N_contr	integer	sample size for body mass of control offspring close to or during the testing period

bm_N_exp	integer	sample size for body mass of experimental offspring close to or during the testing period

bm_dPP	integer	age of offspring when body mass was measured [in days post-parturition]

offspring_diet	character	offspring diet after weaning [type of control diet]

offspring_con_adlib	character	were control offspring fed ad libitum after weaning? [yes or no]

offspring_diet_level	character	name of offspring diet after weaning

offspring_diet_end_dPP	integer	end of the offspring diet [in days post-parturition]

post_diet_adlib	character	were experimental offspring fed ad libitum after weaning? [yes or no]

response_age_dPP	numeric	offspring age when behavioural testing started [in days post-parturition]

authors_behaviour_classification	character	author's classification of offspring behaviour [anxiety, exploration, or activity]

our_behaviour_classification	character	our classification of offspring behaviour [anxiety, exploration, or activity]

response_test	character	type of test used [elevated T-maze (ETM), open field, etc.] to measure offspring behaviour

time_trial	integer	duration of the testing [min]

measure	character	measures taken during testing [total distance moved, time spent in open arm, etc.]

unit	character	unit of the behavioural measure taken [min, s, m, number (#), etc.]

high_better	character	for activity and exploration, a higher number is assumed to be better (i.e., animals were more active), but the opposite was assumed for anxiety (i.e., they were more anxious) [yes or no]

night.day	character	time of day when behaviours were measured [night or day]

comparison	character	for a given control-treatment group comparison, animal group codes as used in the original article [e.g., LP, HP]. This field allows identification of exactly which data (i.e., comparison of which pairs of groups) were extracted from the original paper, and is not used in our analyses. For our analyses the groups were re-coded as control/experimental.

exp_mean	numeric	mean of the offspring behaviour measured for the experimental group

exp_se	numeric	S.E. of the offspring behaviour measured for the experimental group

exp_n	integer	sample size for the offspring experimental group

con_mean	numeric	mean of offspring behaviour measured for the control group

con_se	numeric	S.E. of the offspring behaviour measured for the control group

con_n	integer	sample size for the offspring control group

con_ID	character	identifier for shared control groups within experiment

percentage	character	is the offspring behaviour measure a percentage? [yes or no]

Data_source	character	figure or table number in the original paper from which the data were extracted

measure_comments	character	any comments on the offspring behaviour measures

SE_imputed	character	was S.E. imputed for the offspring behaviour measure? [yes or no]

Comments	character	any comments on the data

Details

Data from experiments where dams were subject to caloric or protein restriction or were overfed around gestation were included. Offspring activity, exploration, or anxiety were measured outcomes variables from maternal experimental treatments. Multilevel meta-analysis and meta-regression models were used to analyze the meta-analytic data.

Source

Besson, A. A., Lagisz, M., Senior, A. M., Hector, K. L., & Nakagawa, S. (2016). Effect of maternal diet on offspring coping styles in rodents: A systematic review and meta-analysis. Biological Reviews, 91(4), 1065–1080. https://doi.org/10.1111/brv.12210

Author

Daniel Noble, daniel.noble@anu.edu.au

Concepts

ecology, evolution, standardized mean differences

Examples

### copy data into 'dat' and examine data
dat <- dat.besson2016
head(dat)
#>   comp_ID study_ID dam_ID animal_ID                   Reference species strain manip_type
#> 1 comp_01     s_01   D_01      A_01 Reyes-Castro et al. (2012a)    rats Wistar    protein
#> 2 comp_02     s_01   D_02      A_02 Reyes-Castro et al. (2012a)    rats Wistar    protein
#> 3 comp_03     s_01   D_03      A_03 Reyes-Castro et al. (2012a)    rats Wistar    protein
#> 4 comp_04     s_01   D_01      A_01 Reyes-Castro et al. (2012a)    rats Wistar    protein
#> 5 comp_05     s_01   D_02      A_02 Reyes-Castro et al. (2012a)    rats Wistar    protein
#> 6 comp_06     s_01   D_03      A_03 Reyes-Castro et al. (2012a)    rats Wistar    protein
#>   manip_direction nom_manip_val exp control manip_parameter vitmin_eql adlib_con adlib_exp
#> 1               -           50% 10%     20% protein content        yes       yes       yes
#> 2               -           50% 10%     20% protein content        yes       yes       yes
#> 3               -           50% 10%     20% protein content        yes       yes       yes
#> 4               -           50% 10%     20% protein content        yes       yes       yes
#> 5               -           50% 10%     20% protein content        yes       yes       yes
#> 6               -           50% 10%     20% protein content        yes       yes       yes
#>                diet_con                     diet_exp dam_diet_start_dPC dam_diet_end_dPC diet_label
#> 1  Zeiger_Rodent_RQ22-5 Zeiger_Rodent_RQ22-5_LP_isoc                  1               22        pre
#> 2  Zeiger_Rodent_RQ22-6 Zeiger_Rodent_RQ22-5_LP_isoc                 22               43       lact
#> 3  Zeiger_Rodent_RQ22-7 Zeiger_Rodent_RQ22-5_LP_isoc                  1               43   pre+lact
#> 4  Zeiger_Rodent_RQ22-8 Zeiger_Rodent_RQ22-5_LP_isoc                  1               22        pre
#> 5  Zeiger_Rodent_RQ22-9 Zeiger_Rodent_RQ22-5_LP_isoc                 22               43       lact
#> 6 Zeiger_Rodent_RQ22-10 Zeiger_Rodent_RQ22-5_LP_isoc                  1               43   pre+lact
#>   age_mating n_con_dam n_exp_dam multi_use_con dam_housing temperature photoperiod litter_size
#> 1        122        20        20           yes      single          22          12          10
#> 2        122        20        20           yes      single          22          12          10
#> 3        122        20        20           yes      single          22          12          10
#> 4        122        20        20           yes      single          22          12          10
#> 5        122        20        20           yes      single          22          12          10
#> 6        122        20        20           yes      single          22          12          10
#>   litter_size_equalized crossfostered sex housing bodymass_mean_contr bodymass_SE_contr
#> 1                   yes            no   f   group                 209                 4
#> 2                   yes            no   f   group                 209                 4
#> 3                   yes            no   f   group                 209                 4
#> 4                   yes            no   f   group                 209                 4
#> 5                   yes            no   f   group                 209                 4
#> 6                   yes            no   f   group                 209                 4
#>   bodymass_mean_exp bodymass_SE_exp bm_N_contr bm_N_exp bm_dPP     offspring_diet offspring_con_adlib
#> 1               200               7         10       10     90 20% casein control                 yes
#> 2               190               6         10       10     90 20% casein control                 yes
#> 3               199              10         10       10     90 20% casein control                 yes
#> 4               200               7         10       10     90 20% casein control                 yes
#> 5               190               6         10       10     90 20% casein control                 yes
#> 6               199              10         10       10     90 20% casein control                 yes
#>    offspring_diet_level offspring_diet_end_dPP post_diet_adlib response_age_dPP
#> 1  Zeiger_Rodent_RQ22-5                      1             yes               90
#> 2  Zeiger_Rodent_RQ22-6                     21             yes               90
#> 3  Zeiger_Rodent_RQ22-7                     21             yes               90
#> 4  Zeiger_Rodent_RQ22-8                      1             yes               90
#> 5  Zeiger_Rodent_RQ22-9                     21             yes               90
#> 6 Zeiger_Rodent_RQ22-10                     21             yes               90
#>   authors_behaviour_classification our_behaviour_classification response_test time_trial
#> 1                          ANXIETY                     ACTIVITY           EPM          5
#> 2                          ANXIETY                     ACTIVITY           EPM          5
#> 3                          ANXIETY                     ACTIVITY           EPM          5
#> 4                          ANXIETY                      ANXIETY           EPM          5
#> 5                          ANXIETY                      ANXIETY           EPM          5
#> 6                          ANXIETY                      ANXIETY           EPM          5
#>            measure unit high_better night.day comparison exp_mean exp_se exp_n con_mean con_se con_n
#> 1     T_dist_moved    m         yes     night      cc-rc    13.90   1.28     8    11.97   1.28     8
#> 2     T_dist_moved    m         yes     night      cc-cr    13.90   0.96     8    11.97   1.28     8
#> 3     T_dist_moved    m         yes     night      cc-rr    10.69   0.32     8    11.97   1.28     8
#> 4 Open_arm_entries    #         yes     night      cc-rc     3.59   0.70     8     5.00   0.89     8
#> 5 Open_arm_entries    #         yes     night      cc-cr     4.29   0.70     8     5.00   0.89     8
#> 6 Open_arm_entries    #         yes     night      cc-rr     3.20   0.48     8     5.00   0.89     8
#>   con_ID percentage Data_source measure_comments SE_imputed
#> 1 con_01         no      Fig. 1             <NA>         no
#> 2 con_01         no      Fig. 1             <NA>         no
#> 3 con_01         no      Fig. 1             <NA>         no
#> 4 con_02         no      Fig. 1             <NA>         no
#> 5 con_02         no      Fig. 1             <NA>         no
#> 6 con_02         no      Fig. 1             <NA>         no
#>                                                   Comments
#> 1 details of diet can be found in Zambrano et al. (2005). 
#> 2 details of diet can be found in Zambrano et al. (2005). 
#> 3 details of diet can be found in Zambrano et al. (2005). 
#> 4 details of diet can be found in Zambrano et al. (2005). 
#> 5 details of diet can be found in Zambrano et al. (2005). 
#> 6 details of diet can be found in Zambrano et al. (2005). 

# \dontrun{

### load metafor
library(metafor)

### compute SD from SE
dat$sd_c <- with(dat, con_se * sqrt(con_n))
dat$sd_e <- with(dat, exp_se * sqrt(exp_n))

### compute standardized mean differences and corresponding sampling variances
dat <- escalc(measure="SMD", m1i=exp_mean, m2i=con_mean, sd1i=sd_e, sd2i=sd_c,
              n1i=exp_n, n2i=con_n, data=dat, add.measure=TRUE)

### fit model
mod1 <- rma.mv(yi ~ 1, V = vi, random = list(~ 1 | study_ID, ~ 1 | comp_ID), data = dat)
#> Warning: Rows with NAs omitted from model fitting.
mod1
#> 
#> Multivariate Meta-Analysis Model (k = 389; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed    factor 
#> sigma^2.1  0.1179  0.3433     46     no  study_ID 
#> sigma^2.2  0.3698  0.6081    389     no   comp_ID 
#> 
#> Test for Heterogeneity:
#> Q(df = 388) = 1262.0967, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb   ci.ub    
#>  -0.1067  0.0712  -1.4994  0.1338  -0.2462  0.0328    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }