dat.bourassa1996 {metadat}	R Documentation

Studies on the Association between Handedness and Eye-Dominance

Description

Results from 47 studies on the association between handedness and eye-dominance.

Usage


dat.bourassa1996

Format

The data frame contains the following columns:

study 	 numeric 	 study number

sample 	 numeric 	 sample number

author 	 character 	 (first) author

year 	 numeric 	 publication year

selection 	 character 	 selection of subjects on the basis of eyedness or handedness

investigator 	 character 	 investigator (psychologist, educationalist, or other)

hand_assess 	 character 	 method to assess handedness (questionnaire or performance based)

eye_assess 	 character 	 method to assess eyedness (see ‘Details’)

mage 	 numeric 	 mean age of sample

lh.le 	 numeric 	 number of left-handed left-eyed individuals

lh.re 	 numeric 	 number of left-handed right-eyed individuals

rh.le 	 numeric 	 number of right-handed left-eyed individuals

rh.re 	 numeric 	 number of right-handed right-eyed individuals

sex 	 character 	 sex of the sample (combined, male, or female)

Details

The 47 studies included in this meta-analysis examined the association between handedness and eye-dominance (ocular dominance or eyedness). Results are given in terms of \(2 \times 2\) tables, indicating the number of left-handed left-eyed, left-handed right-eyed, right-handed left-eyed, and right-handed right-eyed individuals. Note that some studies included multiple (independent) samples, so that the meta-analysis included 54 samples in total. Also, for some studies, the combined data of the males and females are further broken down into the two subgroups.

In some studies, there was indication that the selection of subjects was not random with respect to handedness and/or eyedness. While this should not influence the size of the association as measured with the odds ratio, this invalidates those studies for assessing the overall percentage of left-eyed and left-handed individuals.

Handedness was assessed in the individual studies either based on a questionnaire or inventory or based on task performance. Eyedness was assessed based on various methods: E.1 methods are based on task performance, while E.2.a denotes assessment based on a questionnaire. The performance based methods could be further broken down into: E.1.a.i (monocular procedure with object/instrument held in one hand), E.1.a.ii (monocular procedure with object/instrument held in both hands), E.1.b (binocular procedure), E.1.c (a combination of the previous methods), and E.1.d (some other method).

Concepts

psychology, odds ratios

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Bourassa, D. C., McManus, I. C., & Bryden, M. P. (1996). Handedness and eye-dominance: A meta-analysis of their relationship. Laterality, 1(1), 5–34. https://doi.org/10.1080/713754206

Examples


### copy data into 'dat' and examine data
dat <- dat.bourassa1996
head(dat, 10)

## Not run: 

### load metafor package
library(metafor)

### calculate log(OR) and corresponding sampling variance with 1/2 correction
dat <- escalc(measure="OR", ai=lh.le, bi=lh.re, ci=rh.le, di=rh.re, data=dat, add=1/2, to="all")
head(dat, 10)

### overall association between handedness and eyedness
res <- rma(yi, vi, data=dat, subset=sex=="combined")
res
predict(res, transf=exp, digits=2)


## End(Not run)

[Package metadat version 1.2-0 Index]