Studies on the Association between Handedness and Eye-Dominance

Results from 47 studies on the association between handedness and eye-dominance.

dat.bourassa1996

Format

The data frame contains the following columns:

study	numeric	study number

sample	numeric	sample number

author	character	(first) author

year	numeric	publication year

selection	character	selection of subjects on the basis of eyedness or handedness

investigator	character	investigator (psychologist, educationalist, or other)

hand_assess	character	method to assess handedness (questionnaire or performance based)

eye_assess	character	method to assess eyedness (see ‘Details’)

mage	numeric	mean age of sample

lh.le	numeric	number of left-handed left-eyed individuals

lh.re	numeric	number of left-handed right-eyed individuals

rh.le	numeric	number of right-handed left-eyed individuals

rh.re	numeric	number of right-handed right-eyed individuals

sex	character	sex of the sample (combined, male, or female)

Details

The 47 studies included in this meta-analysis examined the association between handedness and eye-dominance (ocular dominance or eyedness). Results are given in terms of \(2 \times 2\) tables, indicating the number of left-handed left-eyed, left-handed right-eyed, right-handed left-eyed, and right-handed right-eyed individuals in each sample. Note that some studies included multiple (independent) samples, so that the meta-analysis included 54 samples in total. Also, for some studies, the combined data of the males and females are further broken down into the two subgroups.

In some studies, there was indication that the selection of subjects was not random with respect to handedness and/or eyedness. While this should not influence the size of the association as measured with the odds ratio, this invalidates those studies for assessing the overall percentage of left-eyed and left-handed individuals.

Handedness was assessed in the individual studies either based on a questionnaire or based on task performance. Eyedness was assessed based on various methods: E.1 methods are based on task performance, while E.2.a denotes assessment based on a questionnaire. The performance based methods could be further broken down into: E.1.a.i (monocular procedure with object/instrument held in one hand), E.1.a.ii (monocular procedure with object/instrument held in both hands), E.1.b (binocular procedure), E.1.c (a combination of the previous methods), and E.1.d (some other method).

Source

Bourassa, D. C., McManus, I. C., & Bryden, M. P. (1996). Handedness and eye-dominance: A meta-analysis of their relationship. Laterality, 1(1), 5–34. https://doi.org/10.1080/713754206

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

psychology, odds ratios, multilevel models

Examples

### copy data into 'dat' and examine data
dat <- dat.bourassa1996
head(dat, 10)
#>    study sample    author year selection investigator   hand_assess eye_assess mage lh.le lh.re rh.le
#> 1      1      1     Mills 1925        no        other questionnaire      E.1.d   NA    93    17   130
#> 2      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37   140    91   305
#> 3      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    74    57   158
#> 4      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    66    34   147
#> 5      3      3     Miles 1930        no psychologist questionnaire   E.1.a.ii   22    16    14    43
#> 6      4      4    Quinan 1931        no        other   performance      E.1.b   25   102    97   597
#> 7      5      5    Jasper 1932       yes psychologist questionnaire   E.1.a.ii   20    17    14    38
#> 8      6      6      Lund 1932        no psychologist   performance   E.1.a.ii   20    10     2    52
#> 9      7      7      Eyre 1933        no psychologist   performance   E.1.a.ii   15     7     2    15
#> 10     8      8 Updegraff 1933        no psychologist   performance   E.1.a.ii    4     4     2    15
#>    rh.re      sex
#> 1    760 combined
#> 2    697 combined
#> 3    427     male
#> 4    270   female
#> 5    114 combined
#> 6   1898 combined
#> 7     80 combined
#> 8    170 combined
#> 9    169 combined
#> 10    38 combined

### load metafor package
library(metafor)

### calculate log(OR) and corresponding sampling variance with 1/2 correction
dat <- escalc(measure="OR", ai=lh.le, bi=lh.re, ci=rh.le, di=rh.re, data=dat, add=1/2, to="all")
head(dat, 10)
#> 
#>    study sample    author year selection investigator   hand_assess eye_assess mage lh.le lh.re rh.le 
#> 1      1      1     Mills 1925        no        other questionnaire      E.1.d   NA    93    17   130 
#> 2      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37   140    91   305 
#> 3      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    74    57   158 
#> 4      2      2    Downey 1927       yes psychologist questionnaire      E.1.b   37    66    34   147 
#> 5      3      3     Miles 1930        no psychologist questionnaire   E.1.a.ii   22    16    14    43 
#> 6      4      4    Quinan 1931        no        other   performance      E.1.b   25   102    97   597 
#> 7      5      5    Jasper 1932       yes psychologist questionnaire   E.1.a.ii   20    17    14    38 
#> 8      6      6      Lund 1932        no psychologist   performance   E.1.a.ii   20    10     2    52 
#> 9      7      7      Eyre 1933        no psychologist   performance   E.1.a.ii   15     7     2    15 
#> 10     8      8 Updegraff 1933        no psychologist   performance   E.1.a.ii    4     4     2    15 
#>    rh.re      sex     yi     vi 
#> 1    760 combined 3.4384 0.0768 
#> 2    697 combined 1.2544 0.0228 
#> 3    427     male 1.2512 0.0395 
#> 4    270   female 1.2627 0.0545 
#> 5    114 combined 1.0970 0.1613 
#> 6   1898 combined 1.2061 0.0222 
#> 7     80 combined 0.9257 0.1645 
#> 8    170 combined 2.6130 0.5202 
#> 9    169 combined 3.4906 0.6037 
#> 10    38 combined 1.4976 0.7127 
#> 

### overall association between handedness and eyedness
res <- rma(yi, vi, data=dat, subset=sex=="combined")
res
#> 
#> Random-Effects Model (k = 54; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.4340 (SE = 0.1208)
#> tau (square root of estimated tau^2 value):      0.6588
#> I^2 (total heterogeneity / total variability):   89.08%
#> H^2 (total variability / sampling variability):  9.16
#> 
#> Test for Heterogeneity:
#> Q(df = 53) = 355.7748, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub      
#>   1.3135  0.1098  11.9597  <.0001  1.0982  1.5287  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  3.72  3.00  4.61  1.00 13.77 
#> 

### multilevel model to account for heterogeneity at the study and sample levels
res <- rma.mv(yi, vi, random = ~ 1 | study/sample, data=dat, subset=sex=="combined")
res
#> 
#> Multivariate Meta-Analysis Model (k = 54; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed        factor 
#> sigma^2.1  0.2097  0.4579     47     no         study 
#> sigma^2.2  0.1985  0.4455     54     no  study/sample 
#> 
#> Test for Heterogeneity:
#> Q(df = 53) = 355.7748, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub      
#>   1.2712  0.1133  11.2169  <.0001  1.0491  1.4934  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  3.57  2.86  4.45  1.00 12.72 
#> 

### restructure the dataset to keep only the male/female data when it is reported
### separately and the combined data when this is the only data reported
dat <- lapply(split(dat, dat$sample), function(x) {
   if (nrow(x) == 3L) {
      x[-which(x$sex == "combined"),]
   } else {
      x
   }
})
dat <- do.call(rbind, dat)
rownames(dat) <- NULL
dat
#> 
#>    study sample       author year selection   investigator   hand_assess eye_assess mage lh.le lh.re 
#> 1      1      1        Mills 1925        no          other questionnaire      E.1.d   NA    93    17 
#> 2      2      2       Downey 1927       yes   psychologist questionnaire      E.1.b   37    74    57 
#> 3      2      2       Downey 1927       yes   psychologist questionnaire      E.1.b   37    66    34 
#> 4      3      3        Miles 1930        no   psychologist questionnaire   E.1.a.ii   22    16    14 
#> 5      4      4       Quinan 1931        no          other   performance      E.1.b   25   102    97 
#> 6      5      5       Jasper 1932       yes   psychologist questionnaire   E.1.a.ii   20    17    14 
#> 7      6      6         Lund 1932        no   psychologist   performance   E.1.a.ii   20    10     2 
#> 8      7      7         Eyre 1933        no   psychologist   performance   E.1.a.ii   15     7     2 
#> 9      8      8    Updegraff 1933        no   psychologist   performance   E.1.a.ii    4     4     2 
#> 10     9      9        Gates 1936        no educationalist   performance      E.1.c    5     2     3 
#> 11    10     10        Witty 1936        no educationalist questionnaire      E.1.c   10     3     4 
#> 12    11     11         Fink 1938        no          other   performance      E.1.d   NA     4     3 
#> 13    12     12      Castner 1939        no   psychologist   performance   E.1.a.ii   10     2     1 
#> 14    13     13     Schonell 1941        no educationalist   performance   E.1.a.ii   10     3     2 
#> 15    14     14       Lavery 1944        no          other questionnaire      E.1.d   31    27     0 
#> 16    15     15     Hildreth 1945        no   psychologist   performance   E.1.a.ii    9     9    11 
#> 17    16     16    Stevenson 1953        no educationalist   performance      E.1.d    5     5     2 
#> 18    17     17      Merrell 1957        no          other   performance      E.1.c   22    20    13 
#> 19    18     18        Meuhl 1963        no educationalist   performance   E.1.a.ii    5     6     8 
#> 20    19     19       Harmon 1966       yes          other   performance      E.1.d   NA     8     5 
#> 21    20     20   Rengstorff 1967        no          other questionnaire   E.1.a.ii   24   241   211 
#> 22    21     21     Stephens 1967       yes educationalist   performance      E.1.c    6    11     6 
#> 23    21     21     Stephens 1967       yes educationalist   performance      E.1.c    6     8     4 
#> 24    21     22     Stephens 1967        no   psychologist   performance   E.1.a.ii   27     2     2 
#> 25    21     22     Stephens 1967        no   psychologist   performance   E.1.a.ii   27     0     0 
#> 26    22     23       Dawson 1972        no   psychologist   performance   E.1.a.ii   27     4     2 
#> 27    22     23       Dawson 1972        no   psychologist   performance   E.1.a.ii   27     4     0 
#> 28    23     24          Gur 1974       yes   psychologist questionnaire      E.1.b   22     9     9 
#> 29    23     24          Gur 1974       yes   psychologist questionnaire      E.1.b   22     4     6 
#> 30    24     25    Chaurasia 1976        no          other questionnaire      E.1.b   25     8    14 
#> 31    24     25    Chaurasia 1976        no          other questionnaire      E.1.b   25     2     6 
#> 32    25     26      Hardyck 1976        no   psychologist   performance   E.1.a.ii    9   429   311 
#> 33    26     27       Dawson 1977        no   psychologist   performance   E.1.a.ii   17    10    10 
#> 34    27     28          Gur 1977        no   psychologist questionnaire      E.1.c   35    10     6 
#> 35    27     28          Gur 1977        no   psychologist questionnaire      E.1.c   35     1     5 
#> 36    28     29     Van-Camp 1977        no   psychologist   performance      E.1.c    7    18    20 
#> 37    29     30      Birkett 1979       yes   psychologist questionnaire      E.1.b   23    11    10 
#> 38    29     30      Birkett 1979       yes   psychologist questionnaire      E.1.b   23    27     4 
#> 39    30     31    Hovsepian 1980        no   psychologist questionnaire    E.1.a.i   22    19     6 
#> 40    31     32        Porac 1980        no   psychologist   performance   E.1.a.ii   10    19    12 
#> 41    32     33        Porac 1981        no   psychologist questionnaire      E.2.a   21   247   125 
#> 42    32     33        Porac 1981        no   psychologist questionnaire      E.2.a   21   164    73 
#> 43    33     34       Hebben 1981        no   psychologist questionnaire      E.1.c    8     4     5 
#> 44    33     34       Hebben 1981        no   psychologist questionnaire      E.1.c    8     5     3 
#> 45    34     35       Noonan 1981       yes   psychologist   performance      E.1.c   22    26    34 
#> 46    35     36     Van-Camp 1981        no   psychologist questionnaire   E.1.a.ii   75     1     1 
#> 47    35     36     Van-Camp 1981        no   psychologist questionnaire   E.1.a.ii   75     3     1 
#> 48    36     37        Combs 1983       yes   psychologist questionnaire      E.1.b   29    18    19 
#> 49    37     38        Rymar 1983        no          other   performance      E.1.c   10    31     8 
#> 50    37     39        Rymar 1983        no          other questionnaire      E.1.b   29    11    21 
#> 51    38     40      Nachson 1983        no   psychologist   performance   E.1.a.ii    7    22    16 
#> 52    38     40      Nachson 1983        no   psychologist   performance   E.1.a.ii    7    15    11 
#> 53    38     41      Nachson 1983        no   psychologist   performance   E.1.a.ii    7   242   156 
#> 54    38     41      Nachson 1983        no   psychologist   performance   E.1.a.ii    7   225   144 
#> 55    39     42    Shan-Ming 1985        no   psychologist   performance   E.1.a.ii   30     7    10 
#> 56    39     42    Shan-Ming 1985        no   psychologist   performance   E.1.a.ii   30     9     5 
#> 57    40     43 Hoogmaartens 1987        no          other questionnaire      E.2.a   23    10     2 
#> 58    41     44  Whittington 1987        no          other   performance    E.1.a.i   11   563   320 
#> 59    42     45      Costeff 1988        no          other   performance   E.1.a.ii   42     5     1 
#> 60    43     46        Brown 1988        no   psychologist   performance    E.1.a.i   23    19     6 
#> 61    44     47       Bryden 1988       yes   psychologist questionnaire   E.1.a.ii   22    46    32 
#> 62    45     48       Bryden 1989       yes   psychologist questionnaire      E.1.b   20    43    44 
#> 63    45     48       Bryden 1989       yes   psychologist questionnaire      E.1.b   20    46    26 
#> 64    46     49     Levander 1989       yes   psychologist questionnaire      E.1.c   22    12    25 
#> 65    46     49     Levander 1989       yes   psychologist questionnaire      E.1.c   22    13    28 
#> 66    47     50 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   25    61    21 
#> 67    47     50 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   25    80    33 
#> 68    47     51 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   22    11    11 
#> 69    47     51 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   22    12     6 
#> 70    47     52 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   30    24    12 
#> 71    47     52 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   30     8     1 
#> 72    47     53 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   25    11     2 
#> 73    47     53 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   25     9    10 
#> 74    47     54 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   25    18     7 
#> 75    47     54 Dargent-Pare 1992        no   psychologist questionnaire      E.2.a   25    12     1 
#>    rh.le rh.re      sex      yi     vi 
#> 1    130   760 combined  3.4384 0.0768 
#> 2    158   427     male  1.2512 0.0395 
#> 3    147   270   female  1.2627 0.0545 
#> 4     43   114 combined  1.0970 0.1613 
#> 5    597  1898 combined  1.2061 0.0222 
#> 6     38    80 combined  0.9257 0.1645 
#> 7     52   170 combined  2.6130 0.5202 
#> 8     15   169 combined  3.4906 0.6037 
#> 9     15    38 combined  1.4976 0.7127 
#> 10    24    39 combined  0.1412 0.7518 
#> 11    18    39 combined  0.5072 0.5873 
#> 12    29    89 combined  1.3612 0.5530 
#> 13    20    49 combined  1.3924 1.1356 
#> 14    19    45 combined  1.1838 0.7590 
#> 15   101   261 combined  4.9537 2.0500 
#> 16    59    98 combined  0.3130 0.2192 
#> 17    23    38 combined  1.2821 0.6503 
#> 18   137   327 combined  1.2856 0.1332 
#> 19    14    34 combined  0.5985 0.3694 
#> 20    31    56 combined  1.0196 0.3489 
#> 21  1587  3440 combined  0.9061 0.0098 
#> 22    11    16     male  0.9316 0.3884 
#> 23    16    17   female  0.6948 0.4576 
#> 24    27    75     male  1.0099 0.8496 
#> 25     6     9   female  0.3795 4.2591 
#> 26    19    31     male  1.0674 0.7053 
#> 27    10    25   female  3.0845 2.3567 
#> 28     4    25     male  1.7346 0.4720 
#> 29    13    17   female -0.1082 0.5073 
#> 30    69   146     male  0.2116 0.2078 
#> 31    28    40   female -0.6041 0.6136 
#> 32  2528  4418 combined  0.8794 0.0062 
#> 33    12    48 combined  1.3558 0.2911 
#> 34    27    57     male  1.2172 0.3028 
#> 35    29    65   female -0.5016 0.8977 
#> 36   118   155 combined  0.1691 0.1177 
#> 37    16    17     male  0.1498 0.2999 
#> 38    23    17   female  1.5153 0.3583 
#> 39    16    33 combined  1.8068 0.2956 
#> 40   172   352 combined  1.1593 0.1399 
#> 41   500  1884     male  2.0049 0.0145 
#> 42   575  1579   female  1.8152 0.0221 
#> 43    22    63     male  0.8369 0.4642 
#> 44    29    60   female  1.1702 0.5180 
#> 45    42    90 combined  0.4920 0.1013 
#> 46     8    11     male  0.3023 1.5379 
#> 47    35    54   female  1.2760 0.9989 
#> 48    31    75 combined  0.8215 0.1503 
#> 49   271   543 combined  2.0040 0.1549 
#> 50   180   353 combined  0.0464 0.1418 
#> 51    90   131     male  0.6838 0.1237 
#> 52   100   129   female  0.5520 0.1691 
#> 53  1203  1725     male  0.7982 0.0119 
#> 54  1298  1739   female  0.7374 0.0127 
#> 55    24   159     male  1.5369 0.2757 
#> 56    44   174   female  1.9130 0.3153 
#> 57    32    84 combined  2.3906 0.5378 
#> 58  2703  6927 combined  1.5052 0.0054 
#> 59    14    37 combined  2.2495 0.9441 
#> 60    29   132 combined  2.6008 0.2466 
#> 61    48   171 combined  1.6212 0.0787 
#> 62    78   193     male  0.8795 0.0634 
#> 63    65   191   female  1.6351 0.0797 
#> 64     4    35     male  1.3525 0.3696 
#> 65    12    33   female  0.2386 0.2190 
#> 66   150   708     male  2.6002 0.0708 
#> 67   224   811   female  2.1617 0.0480 
#> 68    50   280     male  1.7146 0.1973 
#> 69    39   225   female  2.3959 0.2636 
#> 70    92   307     male  1.8742 0.1349 
#> 71    36   133   female  3.0314 0.8192 
#> 72    70   198     male  2.5612 0.5062 
#> 73    93   178   female  0.5465 0.2168 
#> 74    56   236     male  2.3346 0.2093 
#> 75    73   211   female  3.1772 0.7650 
#> 

### multilevel model to account for heterogeneity at the study, sample, and subgroup levels
res <- rma.mv(yi, vi, random = ~ 1 | study/sample/sex, data=dat)
res
#> 
#> Multivariate Meta-Analysis Model (k = 75; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed            factor 
#> sigma^2.1  0.1828  0.4276     47     no             study 
#> sigma^2.2  0.2259  0.4753     54     no      study/sample 
#> sigma^2.3  0.0000  0.0000     75     no  study/sample/sex 
#> 
#> Test for Heterogeneity:
#> Q(df = 74) = 386.3457, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub      
#>   1.2681  0.1130  11.2219  <.0001  1.0466  1.4895  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  3.55  2.85  4.44  1.00 12.69 
#>