dat.cohen1981 {metadat}	R Documentation

Studies on the Relationship between Course Instructor Ratings and Student Achievement

Description

Results from 20 studies on the correlation between course instructor ratings and student achievement.

Usage


dat.cohen1981

Format

The data frame contains the following columns:

study 	 character 	 study author(s) and year

sample 	 character 	 course type

control 	 character 	 ability control

ni 	 numeric 	 sample size of the study (number of sections)

ri 	 numeric 	 observed correlation

Details

The studies included in this dataset examined to what extent students' ratings of a course instructor correlated with their achievement in the course. Instead of correlating individual ratings and achievement scores, the studies were carried out in multisection courses, in which the sections had different instructors but all sections used a common achievement measure (e.g., a final exam). The correlation coefficients reflect the correlation between the mean instructor rating and the mean achievement score of each section. Hence, the unit of analysis are the sections, not the individuals.

Note that this dataset (extracted from Table A.3 in Cooper & Hedges, 1994) only contains studies with at least 10 sections.

Concepts

education, correlation coefficients

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Cooper, H., & Hedges, L. V. (1994). Appendix A: Data Sets. In H. Cooper & L. V. Hedges (Eds.), The handbook of research synthesis (pp. 543-547). New York: Russell Sage Foundation.

References

Cohen, P. A. (1981). Student ratings of instruction and student achievement: A meta-analysis of multisection validity studies. Review of Educational Research, 51(3), 281–309. https://doi.org/10.3102/00346543051003281

Examples


### copy data into 'dat' and examine data
dat <- dat.cohen1981
dat[c(1,4,5)]

## Not run: 

### load metafor package
library(metafor)

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=ri, ni=ni, data=dat[c(1,4,5)])
dat

### meta-analysis of the transformed correlations using a random-effects model
res <- rma(yi, vi, data=dat, digits=2)
res

### predicted average correlation with 95% CI
predict(res, transf=transf.ztor)


## End(Not run)

[Package metadat version 1.2-0 Index]