dat.craft2003 {metadat}	R Documentation

Studies on the Relationship between the Competitive State Anxiety Inventory-2 and Sport Performance

Description

Results from 10 studies on the relationship between the Competitive State Anxiety Inventory-2 (CSAI-2) and sport performance.

Usage


dat.craft2003

Format

The data frame contains the following columns:

study 	 numeric 	 study number

ni 	 numeric 	 sample size

sport 	 character 	 type of sport (T = team sport, I = individual sport)

ri 	 numeric 	 correlation coefficient

var1 	 character 	 variable 1 of the correlation coefficient (see ‘Details’)

var2 	 character 	 variable 2 of the correlation coefficient (see ‘Details’)

Details

The 10 studies included in this dataset are a subset of the studies included in the meta-analysis by Craft et al. (2003) on the relationship between the Competitive State Anxiety Inventory-2 (CSAI-2) and sport performance.

The CSAI-2 has three subscales: cognitive anxiety (acog), somatic anxiety (asom), and self-confidence (conf). The studies included in this dataset administered the CSAI-2 prior to some sport competition and then measured sport performance based on the competition. Most studies provided all 6 correlations (3 for the correlations among the 3 subscales and 3 for the correlations between the subscales and sport performance), but 2 studies (with study numbers 6 and 17) only provided a subset.

Concepts

psychology, correlation coefficients, multivariate models

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Becker, B. J., & Aloe, A. M. (2019). Model-based meta-analysis and related approaches. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (3nd ed., pp. 339–363). New York: Russell Sage Foundation.

References

Craft, L. L., Magyar, T. M., Becker, B. J., & Feltz, D. L. (2003). The relationship between the Competitive State Anxiety Inventory-2 and sport performance: A meta-analysis. Journal of Sport and Exercise Psychology, 25(1), 44–65. https://doi.org/10.1123/jsep.25.1.44

Examples


### copy data into 'dat' and examine data
dat <- dat.craft2003
head(dat, 18)

## Not run: 

### load metafor package
library(metafor)

### construct dataset and var-cov matrix of the correlations
tmp <- rcalc(ri ~ var1 + var2 | study, ni=ni, data=dat)
V <- tmp$V
dat <- tmp$dat

### examine data for study 1
dat[dat$study == 1,]
V[dat$study == 1, dat$study == 1]

### examine data for study 6
dat[dat$study == 6,]
V[dat$study == 6, dat$study == 6]

### examine data for study 17
dat[dat$study == 17,]
V[dat$study == 17, dat$study == 17]

### multivariate random-effects model
res <- rma.mv(yi, V, mods = ~ var1.var2 - 1, random = ~ var1.var2 | study, struct="UN", data=dat)
res


## End(Not run)

[Package metadat version 1.2-0 Index]