dat.crede2010 {metadat}	R Documentation

Studies on the Relationship between Class Attendance and Grades in College Students

Description

Results from 68 studies on the relationship between class attendence and class performance and/or grade point average in college students.

Usage


dat.crede2010

Format

The data frame contains the following columns:

studyid 	 numeric 	 study number

year 	 numeric 	 publication year

source 	 character 	 study source (journal, dissertation, other)

sampleid 	 numeric 	 sample within study number

criterion 	 character 	 criterion variable (grade, gpa)

class 	 character 	 class type (science, nonscience)

ni 	 numeric 	 sample size

ri 	 numeric 	 observed correlation

Details

The 68 studies included in this dataset provide information about the relationship between class attendance of college students and their performance (i.e., grade) in the class and/or their overall grade point average. Some studies included multiple samples and hence the dataset actually contains 97 correlation coefficients.

The dataset was obtained via personal communication. Note that this dataset differs just slightly from the one used by Credé et al. (2010).

Concepts

education, correlation coefficients, multilevel models

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Personal communication.

References

Credé, M., Roch, S. G., & Kieszczynka, U. M. (2010). Class attendance in college: A meta-analytic review of the relationship of class attendance with grades and student characteristics. Review of Educational Research, 80(2), 272–295. https://doi.org/10.3102/0034654310362998

Examples


### copy data into 'dat' and examine data
dat <- dat.crede2010
head(dat, 18)

## Not run: 

### load metafor package
library(metafor)

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=ri, ni=ni, data=dat)

############################################################################

### meta-analysis for the relationship between attendance and grades
res <- rma(yi, vi, data=dat, subset=criterion=="grade")
res

### estimated average correlation with 95% CI/PI
predict(res, transf=transf.ztor, digits=2)

### examine if relationship between attendance and grades differs for nonscience/science classes
res <- rma(yi, vi, mods = ~ class, data=dat, subset=criterion=="grade")
res

### estimated average correlations for nonscience and science classes
predict(res, newmods=c(0,1), transf=transf.ztor, digits=2)

### examine if relationship between attendance and grades has changed over time
res <- rma(yi, vi, mods = ~ year, data=dat, subset=criterion=="grade")
res

############################################################################

### meta-analysis for the relationship between attendance and GPA
res <- rma(yi, vi, data=dat, subset=criterion=="gpa")
res

### estimated average correlation with 95% CI/PI
predict(res, transf=transf.ztor, digits=2)

### examine if relationship between attendance and GPA has changed over time
res <- rma(yi, vi, mods = ~ year, data=dat, subset=criterion=="gpa")
res

############################################################################

### use a multilevel model to examine the relationship between attendance and grades
res <- rma.mv(yi, vi, random = ~ 1 | studyid/sampleid, data=dat, subset=criterion=="grade")
res
predict(res, transf=transf.ztor, digits=2)

### use a multilevel model to examine the relationship between attendance and gpa
res <- rma.mv(yi, vi, random = ~ 1 | studyid/sampleid, data=dat, subset=criterion=="gpa")
res
predict(res, transf=transf.ztor, digits=2)


## End(Not run)

[Package metadat version 1.2-0 Index]