Studies on the Effects of Elevated CO2 Levels on Woody Plant Mass

Results from studies examining the effects of elevated CO2 levels on woody plant mass.

dat.curtis1998

Format

The data frame contains the following columns:

id	numeric	observation number

paper	numeric	paper number

genus	character	genus name

species	character	species name

fungrp	character	plant functional group

co2.ambi	numeric	ambient CO2 level (control group)

co2.elev	numeric	elevated CO2 level (treatment group)

units	character	units for CO2 exposure levels

time	numeric	maximum length of time (days) of CO2 exposure

pot	character	growing method (see ‘Details’)

method	character	CO2 exposure facility (see ‘Details’)

stock	character	planting stock code

xtrt	character	interacting treatment code (see ‘Details’)

level	character	interacting treatment level codes (see ‘Details’)

m1i	numeric	mean plant mass under elevated CO2 level (treatment group)

sd1i	numeric	standard deviation of plant mass underelevated CO2 level (treatment group)

n1i	numeric	number of observations under elevated CO2 level (treatment group)

m2i	numeric	mean plant mass under ambient CO2 level (control group)

sd2i	numeric	standard deviation of plant mass under ambient CO2 level (control group)

n2i	numeric	number of observations under ambient CO2 level (control group)

Details

The studies included in this dataset compared the total above- plus below-ground biomass (in grams) for plants that were either exposed to ambient (around 35 Pa) and elevated CO2 levels (around twice the ambient level). The co2.ambi and co2.elev variables indicate the CO2 levels in the control and treatment groups, respectively (with the units variable specifying the units for the CO2 exposure levels). Many of the studies also varied one or more additional environmental variables (defined by the xtrt and level variables):

NONE = no additional treatment factor

FERT = soil fertility (either a CONTROL, HIGH, or LOW level)

LIGHT = light treatment (always a LOW light level)

FERT+L = soil fertility and light (a LOW light and soil fertility level)

H2O = well watered vs drought (either a WW or DRT level)

TEMP = temperature treatment (either a HIGH or LOW level)

OZONE = ozone exposure (either a HIGH or LOW level)

UVB = ultraviolet-B radiation exposure (either a HIGH or LOW level)

In addition, the studies differed with respect to various design variables, including CO2 exposure duration (time), growing method (pot: number = pot size in liters; GRND = plants rooted in ground; HYDRO = solution or aeroponic culture), CO2 exposure facility (method: GC = growth chamber; GH = greenhouse; OTC = field-based open-top chamber), and planting stock (stock: SEED = plants started from seeds; SAP = plants started from cuttings). The goal of the meta-analysis was to examine the effects of elevated CO2 levels on plant physiology and growth and the interacting effects of the environmental (and design) variables.

Source

Hedges, L. V., Gurevitch, J., & Curtis, P. S. (1999). The meta-analysis of response ratios in experimental ecology. Ecology, 80(4), 1150--1156. https://doi.org/10.1890/0012-9658(1999)080[1150:TMAORR]2.0.CO;2 (data obtained from Ecological Archives, E080-008-S1, at: https://esapubs.org/archive/ecol/E080/008/)

References

Curtis, P. S., & Wang, X. (1998). A meta-analysis of elevated CO2 effects on woody plant mass, form, and physiology. Oecologia, 113(3), 299--313. https://doi.org/10.1007/s004420050381

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Examples

### copy data into 'dat' and examine data
dat <- dat.curtis1998
head(dat)
#>   id paper   genus     species fungrp co2.ambi co2.elev units time pot method stock xtrt   level
#> 1 21    44   ALNUS       RUBRA  N2FIX      350      650  ul/l   47 0.5     GC  SEED FERT    HIGH
#> 2 22    44   ALNUS       RUBRA  N2FIX      350      650  ul/l   47 0.5     GC  SEED FERT CONTROL
#> 3 27   121    ACER      RUBRUM  ANGIO      350      700   ppm   59 2.6     GH  SEED NONE       .
#> 4 32   121 QUERCUS      PRINUS  ANGIO      350      700   ppm   70 2.6     GH  SEED NONE       .
#> 5 35   121   MALUS   DOMESTICA  ANGIO      350      700   ppm   64 2.6     GH  SEED NONE       .
#> 6 38   121    ACER SACCHARINUM  ANGIO      350      700   ppm   50 2.6     GH  SEED NONE       .
#>       m1i      sd1i n1i    m2i      sd2i n2i
#> 1  6.8169 1.7699820   3 3.9450 1.1157970   5
#> 2  2.5961 0.6674662   5 2.2512 0.3275839   5
#> 3  2.9900 0.8560000   5 1.9300 0.5520000   5
#> 4  5.9100 1.7420000   5 6.6200 1.6310000   5
#> 5  4.6100 1.4070000   4 4.1000 1.2570000   4
#> 6 10.7800 1.1630000   5 6.4200 2.0260000   3

# \dontrun{

### load metafor package
library(metafor)

### calculate (log transformed) ratios of means and corresponding sampling variances
dat <- escalc(measure="ROM", m1i=m1i, sd1i=sd1i, n1i=n1i,
                             m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat)
head(dat)
#>   id paper   genus     species fungrp co2.ambi co2.elev units time pot method stock xtrt   level 
#> 1 21    44   ALNUS       RUBRA  N2FIX      350      650  ul/l   47 0.5     GC  SEED FERT    HIGH 
#> 2 22    44   ALNUS       RUBRA  N2FIX      350      650  ul/l   47 0.5     GC  SEED FERT CONTROL 
#> 3 27   121    ACER      RUBRUM  ANGIO      350      700   ppm   59 2.6     GH  SEED NONE       . 
#> 4 32   121 QUERCUS      PRINUS  ANGIO      350      700   ppm   70 2.6     GH  SEED NONE       . 
#> 5 35   121   MALUS   DOMESTICA  ANGIO      350      700   ppm   64 2.6     GH  SEED NONE       . 
#> 6 38   121    ACER SACCHARINUM  ANGIO      350      700   ppm   50 2.6     GH  SEED NONE       . 
#>       m1i      sd1i n1i    m2i      sd2i n2i      yi     vi 
#> 1  6.8169 1.7699820   3 3.9450 1.1157970   5  0.5470 0.0385 
#> 2  2.5961 0.6674662   5 2.2512 0.3275839   5  0.1425 0.0175 
#> 3  2.9900 0.8560000   5 1.9300 0.5520000   5  0.4378 0.0328 
#> 4  5.9100 1.7420000   5 6.6200 1.6310000   5 -0.1134 0.0295 
#> 5  4.6100 1.4070000   4 4.1000 1.2570000   4  0.1172 0.0468 
#> 6 10.7800 1.1630000   5 6.4200 2.0260000   3  0.5183 0.0355 

### meta-analysis using a random-effects model
res <- rma(yi, vi, method="DL", data=dat)
res
#> 
#> Random-Effects Model (k = 102; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0216 (SE = 0.0088)
#> tau (square root of estimated tau^2 value):      0.1471
#> I^2 (total heterogeneity / total variability):   86.87%
#> H^2 (total variability / sampling variability):  7.61
#> 
#> Test for Heterogeneity:
#> Q(df = 101) = 769.0185, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub 
#>   0.2531  0.0185  13.6965  <.0001  0.2168  0.2893  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### average ratio of means with 95% CI
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.29  1.24  1.34  0.96  1.72 
#> 

### meta-analysis for plants grown under nutrient stress
res <- rma(yi, vi, method="DL", data=dat, subset=(xtrt=="FERT" & level=="LOW"))
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.14  1.04  1.26  0.82  1.59 
#> 

### meta-analysis for plants grown under low light conditions
res <- rma(yi, vi, method="DL", data=dat, subset=(xtrt=="LIGHT" & level=="LOW"))
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.57  1.34  1.85  0.94  2.63 
#> 

# }