Studies on Low Dosage Tricyclic Antidepressants for the Treatment of Depression

Results on depression severity from 17 studies comparing low dosage tricyclic antidepressants (TCA) and placebo for the treatment of depression.

dat.furukawa2003

Format

The data frame contains the following columns:

author	character	First author with information on dosage in parentheses

Ne	integer	number of patients in low TCA group

Me	numeric	depression severity (low TCA)

Se	numeric	standard deviation (low TCA)

Nc	integer	number of patients in placebo group

Mc	numeric	depression severity (placebo)

Sc	numeric	standard deviation (placebo)

Details

Furukawa et al. (2003) carried out a systematic review comparing low dosage tricyclic antidepressants (TCA) with placebo for the treatment of depression. They reported the effect on presence/absence of depression and on depression severity at various time points. Here we focus on depression severity at four weeks. Most studies used some version of the Hamilton Depression Rating Scale, however, some studies used the Montgomery-Asberg Depression Rating Scale. Accordingly, it is not possible to pool the estimated effects directly.

This data set is used as an example in Schwarzer et al. (2015).

Source

Furukawa, T. A., McGuire, H., & Barbui, C. (2003). Low dosage tricyclic antidepressants for depression. Cochrane Database of Systematic Reviews, 3, CD003197. https://doi.org/10.1002/14651858.CD003197

References

Schwarzer, G., Carpenter, J. R., & Rücker, G. (2015). Meta-analysis with R. Cham, Switzerland: Springer.

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

standardized mean differences

Examples

### Show first five studies
head(dat.furukawa2003, 5)
#>             author Ne   Me   Se Nc   Mc   Sc
#> 1  Blashki(75&150) 13  6.4  5.4 18 11.4  9.6
#> 2   Hormazabal(86) 17 11.0  8.2 16 19.0  8.2
#> 3 Jacobson(75-100) 10 17.5  8.8  6 23.0  8.8
#> 4      Jenkins(75)  7 12.3  9.9  7 20.0 10.5
#> 5   Lecrubier(100) 73 15.7 10.6 73 18.7 10.6

### Load meta package
suppressPackageStartupMessages(library("meta"))

### Use RevMan5 settings
oldset <- settings.meta("RevMan5", digits = 2)

### Conduct random effects meta-analysis with Hedges' g as effect measure
mc2 <- metacont(Ne, Me, Se, Nc, Mc, Sc, common = FALSE,
                data = dat.furukawa2003, sm = "SMD")
mc2
#> Number of studies: k = 17
#> Number of observations: o = 902 (o.e = 479, o.c = 423)
#> 
#>                        SMD         95%-CI     z  p-value
#> Random effects model -0.59 [-0.87, -0.30] -4.04 < 0.0001
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 0.231 [0.138, 0.981]; tau = 0.4806 [0.3710, 0.9906]
#>  I^2 = 73% [55%, 83%]; H = 1.91 [1.50, 2.43]
#> 
#> Test of heterogeneity:
#>      Q d.f.  p-value
#>  58.27   16 < 0.0001
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q
#> - Hedges' g (bias corrected standardised mean difference)

### Use previous settings
settings.meta(oldset)