Studies on the Effectiveness of Self-Management Education and Regular Medical Review for Adults with Asthma

Results from 15 trials examining the effectiveness of self-management education and regular medical review for adults with asthma.

dat.gibson2002

Format

The data frame contains the following columns:

author	character	first author of study

year	numeric	publication year

n1i	numeric	number of participants in the intervention group

m1i	numeric	mean number of days off work/school in the intervention group

sd1i	numeric	standard deviation of the number of days off work/school in the intervention group

n2i	numeric	number of participants in the control/comparison group

m2i	numeric	mean number of days off work/school in the control/comparison group

sd2i	numeric	standard deviation of the number of days off work/school in the control/comparison group

ai	numeric	number of participants who had one or more days off work/school in the intervention group

bi	numeric	number of participants who no days off work/school in the intervention group

ci	numeric	number of participants who had one or more days off work/school in the control/comparison group

di	numeric	number of participants who no days off work/school in the control/comparison group

type	numeric	numeric code for the intervention type (see ‘Details’)

Details

Asthma management guidelines typically recommend for patients to receive education and regular medical review. While self-management programs have been shown to increase patient knowledge, it is less clear to what extent they actually impact health outcomes. The systematic review by Gibson et al. (2002) examined the effectiveness of self-management education and regular medical review for adults with asthma. In each study, participants receiving a certain management intervention were compared against those in a control/comparison group with respect to a variety of health outcomes. One of the outcomes examined in a number of studies was the number of days off work/school.

The majority of studies reporting this outcome provided means and standard deviations allowing a meta-analysis of standardized mean differences. Seven studies also reported the number of participants who had one or more days off work/school in each group. These studies could be meta-analyzed using, for example, (log) risk ratios. Finally, one could also consider a combined analysis based on standardized mean differences computed from the means and standard deviations where available and using probit transformed risk differences (which also provide estimates of the standardized mean difference) for the remaining studies.

Some degree of patient education was provided in all studies. In addition, the type variable indicates what additional intervention components were included in each study:

optimal self-management (writing action plan, self-monitoring, regular medical review),

self-monitoring and regular medical review,

self-monitoring only,

regular medical review only,

written action plan only.

Source

Gibson, P. G., Powell, H., Wilson, A., Abramson, M. J., Haywood, P., Bauman, A., Hensley, M. J., Walters, E. H., & Roberts, J. J. L. (2002). Self-management education and regular practitioner review for adults with asthma. Cochrane Database of Systematic Reviews, 3, CD001117. https://doi.org/10.1002/14651858.CD001117

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, primary care, risk ratios, standardized mean differences

Examples

### copy data into 'dat' and examine data
dat <- dat.gibson2002
dat
#>            author year n1i   m1i  sd1i n2i   m2i  sd2i ai bi ci di type
#> 1            Cote 1997  50  2.20 12.73  54  5.20 12.50 NA NA NA NA    1
#> 2           Ghosh 1998 140 17.60 24.20 136 34.10 38.80 NA NA NA NA    1
#> 3         Hayward 1996  23  0.38  0.56  19  0.23  0.29 NA NA NA NA    1
#> 4           Heard 1999  97  2.09  5.93  94  2.66  4.95 34 63 36 58    1
#> 5  Ignacio-Garcia 1995  35  4.92  6.05  35 20.00 26.34 24 11 29  6    1
#> 6          Knoell 1998  45  0.85  4.75  55  2.31  9.16 NA NA NA NA    1
#> 7       Lahdensuo 1996  56  2.80  9.00  59  4.80  7.20 13 43 25 34    1
#> 8       Sommaruga 1995  20 24.10 11.80  20 31.80 17.90 NA NA NA NA    1
#> 9          Zeiger 1991 128  1.40  3.30 143  2.30  7.60 NA NA NA NA    1
#> 10         Garret 1994 119  6.23 12.20 100  5.71  8.57 58 42 57 33    2
#> 11           Neri 1996  32  2.10  8.00  33  5.10 14.00  7 25 13 20    3
#> 12         Hilton 1986  86  0.73  1.48 100  0.47  1.20 NA NA NA NA    4
#> 13      Gallefoss 1999  25  8.00 32.00  24 26.00 70.00 NA NA NA NA    5
#> 14           Yoon 1993  28    NA    NA  28    NA    NA  5 23  4 24    1
#> 15         Brewin 1995  12    NA    NA  33    NA    NA  0 12 16 17    3

# \dontrun{

### load metafor package
library(metafor)

### compute standardized mean differences and corresponding sampling variances
dat <- escalc(measure="SMD", m1i=m1i, sd1i=sd1i, n1i=n1i, m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat)
dat
#> 
#>            author year n1i   m1i  sd1i n2i   m2i  sd2i ai bi ci di type      yi     vi 
#> 1            Cote 1997  50  2.20 12.73  54  5.20 12.50 NA NA NA NA    1 -0.2361 0.0388 
#> 2           Ghosh 1998 140 17.60 24.20 136 34.10 38.80 NA NA NA NA    1 -0.5105 0.0150 
#> 3         Hayward 1996  23  0.38  0.56  19  0.23  0.29 NA NA NA NA    1  0.3209 0.0973 
#> 4           Heard 1999  97  2.09  5.93  94  2.66  4.95 34 63 36 58    1 -0.1038 0.0210 
#> 5  Ignacio-Garcia 1995  35  4.92  6.05  35 20.00 26.34 24 11 29  6    1 -0.7804 0.0615 
#> 6          Knoell 1998  45  0.85  4.75  55  2.31  9.16 NA NA NA NA    1 -0.1930 0.0406 
#> 7       Lahdensuo 1996  56  2.80  9.00  59  4.80  7.20 13 43 25 34    1 -0.2445 0.0351 
#> 8       Sommaruga 1995  20 24.10 11.80  20 31.80 17.90 NA NA NA NA    1 -0.4978 0.1031 
#> 9          Zeiger 1991 128  1.40  3.30 143  2.30  7.60 NA NA NA NA    1 -0.1504 0.0148 
#> 10         Garret 1994 119  6.23 12.20 100  5.71  8.57 58 42 57 33    2  0.0484 0.0184 
#> 11           Neri 1996  32  2.10  8.00  33  5.10 14.00  7 25 13 20    3 -0.2589 0.0621 
#> 12         Hilton 1986  86  0.73  1.48 100  0.47  1.20 NA NA NA NA    4  0.1937 0.0217 
#> 13      Gallefoss 1999  25  8.00 32.00  24 26.00 70.00 NA NA NA NA    5 -0.3277 0.0828 
#> 14           Yoon 1993  28    NA    NA  28    NA    NA  5 23  4 24    1      NA     NA 
#> 15         Brewin 1995  12    NA    NA  33    NA    NA  0 12 16 17    3      NA     NA 
#> 

### fit an equal-effects model to the standardized mean differences (as in Gibson et al., 2002)
res <- rma(yi, vi, data=dat, method="EE")
#> Warning: Studies with NAs omitted from model fitting.
print(res, digits=2)
#> 
#> Equal-Effects Model (k = 13)
#> 
#> I^2 (total heterogeneity / total variability):   55.36%
#> H^2 (total variability / sampling variability):  2.24
#> 
#> Test for Heterogeneity:
#> Q(df = 12) = 26.88, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub     ​ 
#>    -0.18  0.05  -3.77  <.01  -0.28  -0.09  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### compute log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, bi=bi, ci=ci, di=di, data=dat)
dat
#> 
#>            author year n1i   m1i  sd1i n2i   m2i  sd2i ai bi ci di type      yi     vi 
#> 1            Cote 1997  50  2.20 12.73  54  5.20 12.50 NA NA NA NA    1      NA     NA 
#> 2           Ghosh 1998 140 17.60 24.20 136 34.10 38.80 NA NA NA NA    1      NA     NA 
#> 3         Hayward 1996  23  0.38  0.56  19  0.23  0.29 NA NA NA NA    1      NA     NA 
#> 4           Heard 1999  97  2.09  5.93  94  2.66  4.95 34 63 36 58    1 -0.0886 0.0362 
#> 5  Ignacio-Garcia 1995  35  4.92  6.05  35 20.00 26.34 24 11 29  6    1 -0.1892 0.0190 
#> 6          Knoell 1998  45  0.85  4.75  55  2.31  9.16 NA NA NA NA    1      NA     NA 
#> 7       Lahdensuo 1996  56  2.80  9.00  59  4.80  7.20 13 43 25 34    1 -0.6017 0.0821 
#> 8       Sommaruga 1995  20 24.10 11.80  20 31.80 17.90 NA NA NA NA    1      NA     NA 
#> 9          Zeiger 1991 128  1.40  3.30 143  2.30  7.60 NA NA NA NA    1      NA     NA 
#> 10         Garret 1994 119  6.23 12.20 100  5.71  8.57 58 42 57 33    2 -0.0880 0.0137 
#> 11           Neri 1996  32  2.10  8.00  33  5.10 14.00  7 25 13 20    3 -0.5883 0.1582 
#> 12         Hilton 1986  86  0.73  1.48 100  0.47  1.20 NA NA NA NA    4      NA     NA 
#> 13      Gallefoss 1999  25  8.00 32.00  24 26.00 70.00 NA NA NA NA    5      NA     NA 
#> 14           Yoon 1993  28    NA    NA  28    NA    NA  5 23  4 24    1  0.2231 0.3786 
#> 15         Brewin 1995  12    NA    NA  33    NA    NA  0 12 16 17    3 -2.5351 1.9543 
#> 

### fit an equal-effects model to the log risk ratios
res <- rma(yi, vi, data=dat, method="EE")
#> Warning: Studies with NAs omitted from model fitting.
print(res, digits=2)
#> 
#> Equal-Effects Model (k = 7)
#> 
#> I^2 (total heterogeneity / total variability):   18.22%
#> H^2 (total variability / sampling variability):  1.22
#> 
#> Test for Heterogeneity:
#> Q(df = 6) = 7.34, p-val = 0.29
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub   ​ 
#>    -0.17  0.08  -2.31  0.02  -0.32  -0.03  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub 
#>  0.84  0.72  0.97 
#> 

### note: Gibson et al. (2002) used the Mantel-Haenszel method for their analysis
rma.mh(measure="RR", ai=ai, bi=bi, ci=ci, di=di, data=dat, digits=2)
#> Warning: Tables with NAs omitted from model fitting.
#> Warning: Some yi/vi values are NA.
#> 
#> Equal-Effects Model (k = 7)
#> 
#> I^2 (total heterogeneity / total variability):  25.74%
#> H^2 (total variability / sampling variability): 1.35
#> 
#> Test for Heterogeneity: 
#> Q(df = 6) = 8.08, p-val = 0.23
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.24  0.08  -2.91  <.01  -0.40  -0.08 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.79   0.67   0.92 
#> 

### compute standardized mean differences where possible and otherwise probit transformed
### risk differences (which also provide estimates of the standardized mean differences)
dat <- escalc(measure="SMD", m1i=m1i, sd1i=sd1i, n1i=n1i,
                             m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat, add.measure=TRUE)
dat <- escalc(measure="PBIT", ai=ai, bi=bi, ci=ci, di=di, data=dat, replace=FALSE, add.measure=TRUE)
dat
#> 
#>            author year n1i   m1i  sd1i n2i   m2i  sd2i ai bi ci di type      yi     vi measure 
#> 1            Cote 1997  50  2.20 12.73  54  5.20 12.50 NA NA NA NA    1 -0.2361 0.0388     SMD 
#> 2           Ghosh 1998 140 17.60 24.20 136 34.10 38.80 NA NA NA NA    1 -0.5105 0.0150     SMD 
#> 3         Hayward 1996  23  0.38  0.56  19  0.23  0.29 NA NA NA NA    1  0.3209 0.0973     SMD 
#> 4           Heard 1999  97  2.09  5.93  94  2.66  4.95 34 63 36 58    1 -0.1038 0.0210     SMD 
#> 5  Ignacio-Garcia 1995  35  4.92  6.05  35 20.00 26.34 24 11 29  6    1 -0.7804 0.0615     SMD 
#> 6          Knoell 1998  45  0.85  4.75  55  2.31  9.16 NA NA NA NA    1 -0.1930 0.0406     SMD 
#> 7       Lahdensuo 1996  56  2.80  9.00  59  4.80  7.20 13 43 25 34    1 -0.2445 0.0351     SMD 
#> 8       Sommaruga 1995  20 24.10 11.80  20 31.80 17.90 NA NA NA NA    1 -0.4978 0.1031     SMD 
#> 9          Zeiger 1991 128  1.40  3.30 143  2.30  7.60 NA NA NA NA    1 -0.1504 0.0148     SMD 
#> 10         Garret 1994 119  6.23 12.20 100  5.71  8.57 58 42 57 33    2  0.0484 0.0184     SMD 
#> 11           Neri 1996  32  2.10  8.00  33  5.10 14.00  7 25 13 20    3 -0.2589 0.0621     SMD 
#> 12         Hilton 1986  86  0.73  1.48 100  0.47  1.20 NA NA NA NA    4  0.1937 0.0217     SMD 
#> 13      Gallefoss 1999  25  8.00 32.00  24 26.00 70.00 NA NA NA NA    5 -0.3277 0.0828     SMD 
#> 14           Yoon 1993  28    NA    NA  28    NA    NA  5 23  4 24    1  0.1467 0.1627    PBIT 
#> 15         Brewin 1995  12    NA    NA  33    NA    NA  0 12 16 17    3 -1.7320 0.4546    PBIT 
#> 

### fit a random-effects model to these estimates
res <- rma(yi, vi, data=dat)
print(res, digits=2)
#> 
#> Random-Effects Model (k = 15; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.04 (SE = 0.03)
#> tau (square root of estimated tau^2 value):      0.20
#> I^2 (total heterogeneity / total variability):   52.75%
#> H^2 (total variability / sampling variability):  2.12
#> 
#> Test for Heterogeneity:
#> Q(df = 14) = 32.82, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub    ​ 
#>    -0.20  0.08  -2.67  <.01  -0.35  -0.05  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### meta-regression model examining if there are systematic differences based on the
### type of measure used (there are only 2 studies where measure="PBIT", so this isn't
### very conclusive here, but shown for illustration purposes)
res <- rma(yi, vi, mods = ~ measure, data=dat)
print(res, digits=2)
#> 
#> Mixed-Effects Model (k = 15; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.04 (SE = 0.03)
#> tau (square root of estimated tau^2 value):             0.21
#> I^2 (residual heterogeneity / unaccounted variability): 56.02%
#> H^2 (unaccounted variability / sampling variability):   2.27
#> R^2 (amount of heterogeneity accounted for):            0.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 13) = 32.60, p-val < .01
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 0.28, p-val = 0.59
#> 
#> Model Results:
#> 
#>             estimate    se   zval  pval  ci.lb  ci.ub   ​ 
#> intrcpt        -0.40  0.38  -1.06  0.29  -1.15   0.34    
#> measureSMD      0.21  0.39   0.53  0.59  -0.56   0.97    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
predict(res, newmods=1, digits=2)
#> 
#>   pred   se ci.lb ci.ub pi.lb pi.ub 
#>  -0.20 0.08 -0.35 -0.04 -0.63  0.24 
#> 

# }