Case-Control Studies on the Impact of Alcohol Consumption on Breast Cancer Risk

Results from 16 case-control studies evaluating the impact of alcohol consumption on breast cancer risk.

dat.greenland1992

Format

The data frame contains the following columns:

author	character	first author

year	integer	publication year

b	numeric	adjusted log risk ratio

SE	numeric	standard error

Details

Greenland and Longecker (1992) describe a method to combine trend estimates from summarized dose-response data. A meta-analysis of 16 case-control studies evaluating the impact of alcohol consumption on breast cancer risk was used as an illustrative example. The estimates are the increase in the log relative risk of breast cancer associated with an average daily alcohol consumption of 1g.

Data are provided in Greenland and Longecker (1992), Table 3.

Source

Greenland, S., & Longnecker, M. P. (1992). Methods for trend estimation from summarized dose-response data, with applications to meta-analysis. American Journal of Epidemiology, 135(11), 1301–1309. https://doi.org/10.1093/oxfordjournals.aje.a116237

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

epidemiology, risk ratios

Examples

### Show first 6 rows of the dataset
head(dat.greenland1992)
#>             author year       b      SE
#> 1  Hiatt and Bawol 1984 0.00434 0.00247
#> 2     Hiatt et al. 1988 0.01090 0.00410
#> 3    Willett t al. 1987 0.02840 0.00564
#> 4 Schatzkin et al. 1987 0.11800 0.04760
#> 5    Harvey et al. 1987 0.01210 0.00429
#> 6 Rosenberg et al. 1982 0.08700 0.02320

### Load meta package
suppressPackageStartupMessages(library(meta))

### Inverse variance method
mg <- metagen(b, SE, data = dat.greenland1992,
  studlab = paste(author, year), sm = "RR",
  random = FALSE, overall.hetstat = FALSE,
  backtransf = FALSE)

### Overall result
print(mg, digits = 5)
#> Number of studies: k = 16
#> 
#>                       logRR             95%-CI    z  p-value
#> Common effect model 0.00823 [0.00564; 0.01081] 6.24 < 0.0001
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method

### Standard error for common effect estimate
round(mg$seTE.common, 5)
#> [1] 0.00132