Studies on the Risk of Lung Cancer in Women Exposed to Environmental Tobacco Smoke

Results from 37 studies on the risk of lung cancer in women exposed to environmental tobacco smoke (ETS) from their smoking spouse.

dat.hackshaw1998

Format

The data frame contains the following columns:

study	numeric	study number

author	character	first author of study

year	numeric	publication year

country	character	country where study was conducted

design	character	study design (either cohort or case-control)

cases	numeric	number of lung cancer cases

or	numeric	odds ratio

or.lb	numeric	lower bound of 95% CI for the odds ratio

or.ub	numeric	upper bound of 95% CI for the odds ratio

yi	numeric	log odds ratio

vi	numeric	corresponding sampling variance

Details

The dataset includes the results from 37 studies (4 cohort, 33 case-control) examining if women (who are lifelong nonsmokers) have an elevated risk for lung cancer due to exposure to environmental tobacco smoke (ETS) from their smoking spouse. Values of the log odds ratio greater than 0 indicate an increased risk of cancer in exposed women compared to women not exposed to ETS from their spouse.

Note that the log odds ratios and corresponding sampling variances were back-calculated from the reported odds ratios and confidence interval (CI) bounds (see ‘Examples’). Since the reported values were rounded to some extent, this introduces some minor inaccuracies into the back-calculations. The overall estimate reported in Hackshaw et al. (1997) and Hackshaw (1998) can be fully reproduced though.

Source

Hackshaw, A. K., Law, M. R., & Wald, N. J. (1997). The accumulated evidence on lung cancer and environmental tobacco smoke. British Medical Journal, 315(7114), 980–988. https://doi.org/10.1136/bmj.315.7114.980

Hackshaw, A. K. (1998). Lung cancer and passive smoking. Statistical Methods in Medical Research, 7(2), 119–136. https://doi.org/10.1177/096228029800700203

Author

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Concepts

medicine, oncology, epidemiology, smoking, odds ratios

Examples

### copy data into 'dat' and examine data
dat <- dat.hackshaw1998
head(dat, 10)
#> 
#>    study       author year   country       design cases   or or.lb or.ub      yi     vi 
#> 1      1    Garfinkel 1981       USA       cohort   153 1.18  0.90  1.54  0.1655 0.0188 
#> 2      2     Hirayama 1984     Japan       cohort   200 1.45  1.02  2.08  0.3716 0.0330 
#> 3      3       Butler 1988       USA       cohort     8 2.02  0.48  8.56  0.7031 0.5402 
#> 4      4     Cardenas 1997       USA       cohort   150 1.20  0.80  1.60  0.1823 0.0313 
#> 5      5         Chan 1982 Hong Kong case-control    84 0.75  0.43  1.30 -0.2877 0.0797 
#> 6      6       Correa 1983       USA case-control    22 2.07  0.81  5.25  0.7275 0.2273 
#> 7      7 Trichopolous 1983    Greece case-control    62 2.13  1.19  3.83  0.7561 0.0889 
#> 8      8      Buffler 1984       USA case-control    41 0.80  0.34  1.90 -0.2231 0.1927 
#> 9      9        Kabat 1984       USA case-control    24 0.79  0.25  2.45 -0.2357 0.3390 
#> 10    10          Lam 1985 Hong Kong case-control    60 2.01  1.09  3.72  0.6981 0.0981 
#> 

### load metafor package
library(metafor)

### random-effects model using the log odds ratios
res <- rma(yi, vi, data=dat, method="DL")
res
#> 
#> Random-Effects Model (k = 37; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0170 (SE = 0.0171)
#> tau (square root of estimated tau^2 value):      0.1305
#> I^2 (total heterogeneity / total variability):   24.21%
#> H^2 (total variability / sampling variability):  1.32
#> 
#> Test for Heterogeneity:
#> Q(df = 36) = 47.4979, p-val = 0.0952
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub      
#>   0.2139  0.0471  4.5390  <.0001  0.1215  0.3062  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average odds ratio with CI (and prediction interval)
predict(res, transf=exp, digits=2)
#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.24  1.13  1.36  0.94  1.63 
#> 

### illustrate how the log odds ratios and corresponding sampling variances
### can be back-calculated based on the reported odds ratios and CI bounds
dat$yi <- NULL
dat$vi <- NULL
dat <- data.frame(dat)
head(dat, 10)
#>    study       author year   country       design cases   or or.lb or.ub
#> 1      1    Garfinkel 1981       USA       cohort   153 1.18  0.90  1.54
#> 2      2     Hirayama 1984     Japan       cohort   200 1.45  1.02  2.08
#> 3      3       Butler 1988       USA       cohort     8 2.02  0.48  8.56
#> 4      4     Cardenas 1997       USA       cohort   150 1.20  0.80  1.60
#> 5      5         Chan 1982 Hong Kong case-control    84 0.75  0.43  1.30
#> 6      6       Correa 1983       USA case-control    22 2.07  0.81  5.25
#> 7      7 Trichopolous 1983    Greece case-control    62 2.13  1.19  3.83
#> 8      8      Buffler 1984       USA case-control    41 0.80  0.34  1.90
#> 9      9        Kabat 1984       USA case-control    24 0.79  0.25  2.45
#> 10    10          Lam 1985 Hong Kong case-control    60 2.01  1.09  3.72
dat <- conv.wald(out=or, ci.lb=or.lb, ci.ub=or.ub, data=dat, transf=log)
head(dat, 10)
#> 
#>    study       author year   country       design cases   or or.lb or.ub      yi     vi 
#> 1      1    Garfinkel 1981       USA       cohort   153 1.18  0.90  1.54  0.1655 0.0188 
#> 2      2     Hirayama 1984     Japan       cohort   200 1.45  1.02  2.08  0.3716 0.0330 
#> 3      3       Butler 1988       USA       cohort     8 2.02  0.48  8.56  0.7031 0.5402 
#> 4      4     Cardenas 1997       USA       cohort   150 1.20  0.80  1.60  0.1823 0.0313 
#> 5      5         Chan 1982 Hong Kong case-control    84 0.75  0.43  1.30 -0.2877 0.0797 
#> 6      6       Correa 1983       USA case-control    22 2.07  0.81  5.25  0.7275 0.2273 
#> 7      7 Trichopolous 1983    Greece case-control    62 2.13  1.19  3.83  0.7561 0.0889 
#> 8      8      Buffler 1984       USA case-control    41 0.80  0.34  1.90 -0.2231 0.1927 
#> 9      9        Kabat 1984       USA case-control    24 0.79  0.25  2.45 -0.2357 0.3390 
#> 10    10          Lam 1985 Hong Kong case-control    60 2.01  1.09  3.72  0.6981 0.0981 
#>