dat.hannum2020 {metadat}	R Documentation

Studies Comparing Objective and Subjective Olfactory Loss in COVID-19 Patients

Description

Results from 35 studies measuring olfactory loss in COVID-19 patients using either objective or subjective measures.

Usage


dat.hannum2020

Format

The data frame contains the following columns:

authorName 	 character 	 (first) author of study

DOI 	 character 	 article DOI number

ni 	 numeric 	 number of Covid-19 positive patients in the study

xi 	 numeric 	 number of Covid-19 positive patients in the study with olfactory loss

percentOlfactoryLoss 	 numeric 	 percent of the sample with olfactory loss

objectivity 	 character 	 objective or subjective measure used

measured 	 character 	 outcome measure

testType 	 character 	 type of test used

country 	 character 	 country where patients were treated

patientType 	 character 	 type of patient information and location where being treated

Details

One of the symptoms of COVID-19 infection is olfactory loss (loss of smell) either recently acquired anosmia (complete loss of smell) or hyposmia (partial loss of smell). One challenge to reaching this symptom is the wide range of reported prevalence for this symptom ranging from 5 percent to 98 percent. In this dataset studies were grouped into one of two groups based on the type of method used to measure smell loss (either subjective measures, such as self-reported smell loss, or objective measures using rated stimuli).

Concepts

medicine, covid-19, proportions

Author(s)

W. Kyle Hamilton whamilton@ucmerced.edu https://kylehamilton.com

Source

Ramirez VA , Hannum ME, Lipson SJ, Herriman RD, Toskala AK, Lin C, Joseph PV, Reed DR. 2020. COVID-19 Smell Loss Prevalence Tracker. Available from: https://vicente-ramirez.shinyapps.io/COVID19_Olfactory_Dashboard/ and https://github.com/vramirez4/OlfactoryLoss (accessed August 11, 2021)

References

Hannum, M. E., Ramirez, V. A., Lipson, S. J., Herriman, R. D., Toskala, A. K., Lin, C., Joseph, P. V., & Reed, D. R. (2020). Objective sensory testing methods reveal a higher prevalence of olfactory loss in COVID-19 positive patients compared to subjective methods: A systematic review and meta-analysis. Chemical Senses, 45(9), 865–874. https://doi.org/10.1093/chemse/bjaa064

Examples


# copy data into 'dat' and examine data
dat <- dat.hannum2020
dat

## Not run: 
   
# load metafor package
library(metafor)

# compute effect size
dat <- escalc(measure="PR", xi=xi, ni=ni, data=dat)

# split data into objective and subjective datasets
dat_split <- split(dat, dat$objectivity)
dat_objective <- dat_split[["Objective"]]
dat_subjective <- dat_split[["Subjective"]]

# random-effects model all studies
res_all <- rma(yi, vi, data=dat)
print(res_all, digits=2)

# random-effects model objective
res_objective <- rma(yi, vi, data=dat_objective)
print(res_objective, digits=2)

# random-effects model subjective
res_subjective <- rma(yi, vi, data=dat_subjective)
print(res_subjective, digits=2)


## End(Not run)

[Package metadat version 1.2-0 Index]