dat.hine1989 {metadat}	R Documentation

Studies on Prophylactic Use of Lidocaine After a Heart Attack

Description

Results from 6 studies evaluating mortality from prophylactic use of lidocaine in acute myocardial infarction.

Usage


dat.hine1989

Format

The data frame contains the following columns:

study 	 numeric 	 study number

source 	 character 	 source of data

n1i 	 numeric 	 number of patients in lidocaine group

n2i 	 numeric 	 number of patients in control group

ai 	 numeric 	 number of deaths in lidocaine group

ci 	 numeric 	 number of deaths in control group

Details

Hine et al. (1989) conducted a meta-analysis of death rates in randomized controlled trials in which prophylactic lidocaine was administered to patients with confirmed or suspected acute myocardial infarction. The dataset describes the mortality at the end of the assigned treatment period for control and intravenous lidocaine treatment groups for six studies. The question of interest is whether there is a detrimental effect of lidocaine. Because the studies were conducted to compare rates of arrhythmias following a heart attack, the studies, taken individually, are too small to detect important differences in mortality rates.

The data in this dataset were obtained from Table I in Normand (1999, p. 322).

Concepts

medicine, cardiology, risk differences

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Normand, S. T. (1999). Meta-analysis: Formulating, evaluating, combining, and reporting. Statistics in Medicine, 18(3), 321–359. https://doi.org/10.1002/(sici)1097-0258(19990215)18:3<321::aid-sim28>3.0.co;2-p

References

Hine, L. K., Laird, N., Hewitt, P., & Chalmers, T. C. (1989). Meta-analytic evidence against prophylactic use of lidocaine in acute myocardial infarction. Archives of Internal Medicine, 149(12), 2694–2698. https://doi.org/10.1001/archinte.1989.00390120056011

Examples


### copy data into 'dat' and examine data
dat <- dat.hine1989
dat

## Not run: 

### load metafor package
library(metafor)

### calculate risk differences and corresponding sampling variances
dat <- escalc(measure="RD", n1i=n1i, n2i=n2i, ai=ai, ci=ci, data=dat)
dat

### meta-analysis of risk differences using a random-effects model
res <- rma(yi, vi, data=dat)
res


## End(Not run)

[Package metadat version 1.2-0 Index]