Individual Participant Data (IPD) from Five Studies on Uro-Vaxom in Urinary Tract Infections

Individual participant data (IPD) of 601 patients from 5 studies comparing Uro-Vaxom versus placebo in recurrent urinary tract infections (UTI) based on the numbers of recurrences experienced.

dat.huber2000

Format

The data frame contains the following columns (one row per patient):

study	factor	study identifier

treatment	factor	treatment (Uro-Vaxom or Placebo)

recurrences	numeric	number of recurrences (relapses)

Details

Huber et al. (2000) jointly analyzed data from five randomized, placebo-controlled studies investigating the use of OM-89 (Uro-Vaxom) in recurrent urinary tract infections (UTI). Treatment efficacy was evaluated based on the numbers of recurrences experienced by study participants.

The eventual analysis was carried out based on nonparametric Wilcoxon-Mann-Whitney (WMW) tests applied for each study individually and then synthesizing the five resulting test statistic values. The WMW test statistics may be expressed in terms of common language effect sizes (CLES), that is, the estimated probability that a patient in the treatment group has a better outcome than a control patient.

Source

Bauer, H. W., Rahlfs, V. W., Lauener, P. A., & Bleßmann, G. S. S. (2002). Prevention of recurrent urinary tract infections with immuno-active E. coli fractions: A meta-analysis of five placebo-controlled double-blind studies. International Journal of Antimicrobial Agents, 19(6), 451–456. https://doi.org/10.1016/s0924-8579(02)00106-1

See also

Author

Christian Röver, christian.roever@med.uni-goettingen.de

Concepts

medicine, proportions, individual participant data)

Examples

# show data
head(dat.huber2000)
#>         study treatment recurrences
#> 1 Frey (1986) Uro-Vaxom           0
#> 2 Frey (1986) Uro-Vaxom           0
#> 3 Frey (1986) Uro-Vaxom           0
#> 4 Frey (1986) Uro-Vaxom           0
#> 5 Frey (1986) Uro-Vaxom           0
#> 6 Frey (1986) Uro-Vaxom           0
str(dat.huber2000)
#> 'data.frame':	601 obs. of  3 variables:
#>  $ study      : Factor w/ 5 levels "Frey (1986)",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ treatment  : Factor w/ 2 levels "Placebo","Uro-Vaxom": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ recurrences: int  0 0 0 0 0 0 0 0 0 0 ...
summary(dat.huber2000)
#>              study         treatment    recurrences  
#>  Frey (1986)    : 58   Placebo  :299   Min.   :0.00  
#>  Tammen (1990)  :120   Uro-Vaxom:302   1st Qu.:0.00  
#>  Schulman (1993):160                   Median :1.00  
#>  Magasi (1994)  :112                   Mean   :1.19  
#>  Pisani (2000)  :151                   3rd Qu.:2.00  
#>                                        Max.   :8.00  
table(dat.huber2000[,1:2])
#>                  treatment
#> study             Placebo Uro-Vaxom
#>   Frey (1986)          31        27
#>   Tammen (1990)        59        61
#>   Schulman (1993)      78        82
#>   Magasi (1994)        54        58
#>   Pisani (2000)        77        74
table(dat.huber2000[,c(2,3,1)])
#> , , study = Frey (1986)
#> 
#>            recurrences
#> treatment    0  1  2  3  4  5  6  7  8
#>   Placebo   10  5  9  3  3  0  1  0  0
#>   Uro-Vaxom 14  8  3  1  1  0  0  0  0
#> 
#> , , study = Tammen (1990)
#> 
#>            recurrences
#> treatment    0  1  2  3  4  5  6  7  8
#>   Placebo   10 19 10 16  3  1  0  0  0
#>   Uro-Vaxom 23 27 10  1  0  0  0  0  0
#> 
#> , , study = Schulman (1993)
#> 
#>            recurrences
#> treatment    0  1  2  3  4  5  6  7  8
#>   Placebo   14 21 15 14  9  4  0  1  0
#>   Uro-Vaxom 39 22  8  4  1  5  1  2  0
#> 
#> , , study = Magasi (1994)
#> 
#>            recurrences
#> treatment    0  1  2  3  4  5  6  7  8
#>   Placebo   12 20 14  7  1  0  0  0  0
#>   Uro-Vaxom 39 14  2  3  0  0  0  0  0
#> 
#> , , study = Pisani (2000)
#> 
#>            recurrences
#> treatment    0  1  2  3  4  5  6  7  8
#>   Placebo   39 21  7  4  1  1  2  0  2
#>   Uro-Vaxom 53 13  2  0  2  1  1  1  1
#> 

# compute WMW statistics for all 5 studies
wmw <- tapply(dat.huber2000, dat.huber2000$study,
              function(x){wilcox.test(x$recurrences[x$treatment=="Placebo"],
                                      x$recurrences[x$treatment=="Uro-Vaxom"])$statistic},
              simplify=TRUE)
#> Warning: cannot compute exact p-value with ties
wmw
#>     Frey (1986)   Tammen (1990) Schulman (1993)   Magasi (1994)   Pisani (2000) 
#>           554.5          2570.5          4340.5          2363.5          3453.5 

# determine treatment group sizes
patients <- table(dat.huber2000[,1:2])
patients
#>                  treatment
#> study             Placebo Uro-Vaxom
#>   Frey (1986)          31        27
#>   Tammen (1990)        59        61
#>   Schulman (1993)      78        82
#>   Magasi (1994)        54        58
#>   Pisani (2000)        77        74

# convert WMW statistics into estimated probability of superiority
# ("common language effect size")
prob <- wmw / (patients[,1]*patients[,2])
round(prob, 3)
#>     Frey (1986)   Tammen (1990) Schulman (1993)   Magasi (1994)   Pisani (2000) 
#>           0.662           0.714           0.679           0.755           0.606