dat.laopaiboon2015 {metadat}	R Documentation

Studies on the Effectiveness of Azithromycin for Treating Lower Respiratory Tract Infections

Description

Results from 15 studies on the effectiveness of azithromycin versus amoxycillin or amoxycillin/clavulanic acid (amoxyclav) in the treatment of acute lower respiratory tract infections.

Usage


dat.laopaiboon2015

Format

The data frame contains the following columns:

author 	 character 	 author(s)

year 	 numeric 	 publication year

ai 	 numeric 	 number of clinical failures in the group treated with azithromycin

n1i 	 numeric 	 number of patients in the group treated with azithromycin

ci 	 numeric 	 number of clinical failures in the group treated with amoxycillin or amoxyclav

n2i 	 numeric 	 number of patients in the group treated with amoxycillin or amoxyclav

age 	 character 	 whether the trial included adults or children

diag.ab 	 numeric 	 trial included patients with a diagnosis of acute bacterial bronchitis

diag.cb 	 numeric 	 trial included patients with a diagnosis of chronic bronchitis with acute exacerbation

diag.pn 	 numeric 	 trial included patients with a diagnosis of pneumonia

ctrl 	 character 	 antibiotic in control group (amoxycillin or amoxyclav)

Details

Azithromycin is an antibiotic useful for the treatment of a number of bacterial infections. Laopaiboon et al. (2015) conducted a meta-analysis of trials comparing the effectiveness of azithromycin versus amoxycillin or amoxycillin/clavulanic acid (amoxyclav) in the treatment of acute lower respiratory tract infections, including acute bacterial bronchitis, acute exacerbations of chronic bronchitis, and pneumonia. The results from 15 trials are included in this dataset.

Concepts

medicine, risk ratios

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Laopaiboon, M., Panpanich, R., & Swa Mya, K. (2015). Azithromycin for acute lower respiratory tract infections. Cochrane Database of Systematic Reviews, 3, CD001954. https://doi.org/10.1002/14651858.CD001954.pub4

Examples


### copy data into 'dat' and examine data
dat <- dat.laopaiboon2015
dat

## Not run: 

### load metafor package
library(metafor)

### analysis using the Mantel-Haenszel method
rma.mh(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, digits=3)

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat)

### random-effects model
res <- rma(yi, vi, data=dat)
res

### average risk ratio with 95% CI
predict(res, transf=exp)


## End(Not run)

[Package metadat version 1.2-0 Index]