dat.lee2004 {metadat}	R Documentation

Studies on Acupoint P6 Stimulation for Preventing Nausea

Description

Results from studies examining the effectiveness of wrist acupuncture point P6 stimulation for preventing postoperative nausea.

Usage


dat.lee2004

Format

The data frame contains the following columns:

id 	 numeric 	 trial id number

study 	 character 	 first author

year 	 numeric 	 study year

ai 	 numeric 	 number of patients experiencing nausea in the treatment group

n1i 	 numeric 	 total number of patients in treatment group

ci 	 numeric 	 number of patients experiencing nausea in the sham group

n2i 	 numeric 	 total number of patients in the sham group

Details

Postoperative nausea and vomiting are common complications following surgery and anaesthesia. As an alternative to drug therapy, acupuncture has been studied as a potential treatment in several trials. The dataset contains the results from 16 clinical trials examining the effectiveness of wrist acupuncture point P6 stimulation for preventing postoperative nausea.

Concepts

medicine, alternative medicine, risk ratios

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Lee, A., & Done, M. L. (2004). Stimulation of the wrist acupuncture point P6 for preventing postoperative nausea and vomiting. Cochrane Database of Systematic Reviews, 3, CD003281. https://doi.org/10.1002/14651858.CD003281.pub2

Examples


### copy data into 'dat' and examine data
dat <- dat.lee2004
dat

## Not run: 

### load metafor package
library(metafor)

### meta-analysis based on log risk ratios
res <- rma(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat)
res
predict(res, transf=exp, digits=2)


## End(Not run)

[Package metadat version 1.2-0 Index]