dat.lim2014 {metadat}	R Documentation

Studies on the Association Between Maternal Size, Offspring Size, and Number of Offsprings

Description

Results from studies examining the association between maternal size, offspring size, and number of offsprings.

Usage


dat.lim2014

Format

The object is a list containing data frames m_o_size, m_o_fecundity, o_o_unadj, and o_o_adj that contain the following columns and the corresponding phylogenetic trees called m_o_size_tree, m_o_fecundity_tree, o_o_unadj_tree, and o_o_adj_tree:

article 	 numeric 	 article id

author 	 character 	 study author(s)

year 	 numeric 	 publication year

species 	 character 	 species

amniotes 	 character 	 whether the species was amniotic

environment 	 character 	 whether the species were wild or captive

reprounit 	 character 	 whether the data were based on lifetime reproductive output or a single reproductive event (only in m_o_size and m_o_fecundity)

ri 	 numeric 	 correlation coefficient

ni 	 numeric 	 sample size

Details

The object dat.lim2014 includes 4 datasets:

m_o_size 	 on the correlation between maternal size and offspring size

m_o_fecundity 	 on the correlation between maternal size and number of offsprings

o_o_unadj 	 on the correlation between offspring size and number of offsprings

o_o_adj 	 on the correlation between offspring size and number of offsprings adjusted for maternal size

Objects m_o_size_tree, m_o_fecundity_tree, o_o_unadj_tree, and o_o_adj_tree are the corresponding phylogenetic trees for the species included in each of these datasets.

Concepts

ecology, evolution, correlation coefficients, multilevel models, phylogeny

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Lim, J. N., Senior, A. M., & Nakagawa, S. (2014). Heterogeneity in individual quality and reproductive trade-offs within species. Evolution, 68(8), 2306–2318. https://doi.org/10.1111/evo.12446

References

Cinar, O., Nakagawa, S., & Viechtbauer, W. (in press). Phylogenetic multilevel meta-analysis: A simulation study on the importance of modelling the phylogeny. Methods in Ecology and Evolution. https://doi.org/10.1111/2041-210X.13760

Hadfield, J. D., & Nakagawa, S. (2010). General quantitative genetic methods for comparative biology: Phylogenies, taxonomies and multi-trait models for continuous and categorical characters. Journal of Evolutionary Biology, 23(3), 494–508. https://doi.org/10.1111/j.1420-9101.2009.01915.x

Nakagawa, S., & Santos, E. S. A. (2012). Methodological issues and advances in biological meta-analysis. Evolutionary Ecology, 26(5), 1253–1274. https://doi.org/10.1007/s10682-012-9555-5

Examples


### copy data into 'dat' and examine data
dat <- dat.lim2014$o_o_unadj
dat[1:14, -c(2:3)]

## Not run: 

### load metafor package
library(metafor)

### load ape package
library(ape, warn.conflicts=FALSE)

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=ri, ni=ni, data=dat)

### copy tree to 'tree'
tree <- dat.lim2014$o_o_unadj_tree

### compute branch lengths
tree <- compute.brlen(tree)

### compute phylogenetic correlation matrix
A <- vcv(tree, corr=TRUE)

### make copy of the species variable
dat$species.phy <- dat$species

### create effect size id variable
dat$esid <- 1:nrow(dat)

### fit multilevel phylogenetic meta-analytic model
res <- rma.mv(yi, vi,
   random = list(~ 1 | article, ~ 1 | esid, ~ 1 | species, ~ 1 | species.phy),
   R=list(species.phy=A), data=dat)
res


## End(Not run)

[Package metadat version 1.2-0 Index]