dat.linde2005 {metadat}	R Documentation

Studies on the Effectiveness of St. John's Wort for Treating Depression

Description

Results from 26 studies on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression.

Usage


dat.linde2005

Format

The data frame contains the following columns:

id 	 numeric 	 study number

study 	 character 	 study author(s)

year 	 numeric 	 publication year

country 	 character 	 study location

ni 	 numeric 	 total sample size

major 	 numeric 	 sample restricted to patients who met criteria for major depression

baseline 	 numeric 	 HRSD baseline score

version 	 numeric 	 HRSD version (17 or 21 items)

duration 	 numeric 	 study duration (in weeks)

prep 	 character 	 Hypericum extract preparation

dosage 	 numeric 	 dosage (in mg)

response 	 numeric 	 definition of response (see ‘Details’)

ai 	 numeric 	 number of responses in treatment group

n1i 	 numeric 	 number of patients in treatment group

ci 	 numeric 	 number of responses in placebo group

n2i 	 numeric 	 number of patients in placebo group

group 	 numeric 	 stratification variable used by the authors (see ‘Details’)

Details

The dataset includes the results from 26 double-blind placebo-controlled trials on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression (note that 2 studies did not provide sufficient response information).

Data were extracted from Table 1 and Figure 3 from Linde et al. (2005). For study duration, the assessment week (instead of the total study duration) was coded for Philipp et al. (1999) and Montgomery et al. (2000). For dosage, the midpoint was coded when a range of values was given.

The definition of what constitutes a response differed across studies and is coded as follows:

HRSD score reduction of at least 50% or HRSD score after therapy <10,

HRSD reduction of at least 50%,

based on HRSD scale but exact definition not reported,

global patient assessment of efficacy,

at least ‘much improved’ on the Clinical Global Impression sub-scale for global improvement.

The group variable corresponds to the variable used by Linde et al. (2005) to stratify their analyses and is coded as follows:

smaller trials restricted to major depression,

larger trials restricted to major depression,

smaller trials not restricted to major depression,

larger trials not restricted to major depression.

Concepts

medicine, psychiatry, risk ratios

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Linde, K., Berner, M., Egger, M., & Mulrow, C. (2005). St John's wort for depression: Meta-analysis of randomised controlled trials. British Journal of Psychiatry, 186(2), 99–107. https://doi.org/10.1192/bjp.186.2.99

References

Viechtbauer, W. (2007). Accounting for heterogeneity via random-effects models and moderator analyses in meta-analysis. Zeitschrift für Psychologie / Journal of Psychology, 215(2), 104–121. https://doi.org/10.1027/0044-3409.215.2.104

Examples


### copy data into 'dat' and examine data
dat <- dat.linde2005
head(dat)

## Not run: 

### load metafor package
library(metafor)

### remove studies with no response information and study with no responses in either group
dat <- dat[-c(5,6,26),]

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, ci=ci, n1i=n1i, n2i=n2i, data=dat)
head(dat)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(yi, vi, data=dat, method="DL")
res

### mixed-effects meta-regression model with stratification variable
res <- rma(yi, vi, mods = ~ factor(group) - 1, data=dat, method="DL")
res

### predicted average risk ratio for each level of the stratification variable
predict(res, newmods=diag(4), transf=exp, digits=2)


## End(Not run)

[Package metadat version 1.2-0 Index]