Studies on Antidepressants for the Primary Care Setting

Results from 93 trials examining 22 interventions (including placebo and usual care) for the primary care of depression.

dat.linde2016

Format

The data frame contains the following columns:

id	integer	study ID

author	character	first author

year	numeric	year of publication

resp	numeric	number of responders

n	numeric	number of patients

int	character	intervention label

int.long	character	intervention label (full name)

Details

This dataset comes from a network meta-analysis of 22 treatments of depression in primary care (Linde et al., 2016), based on 93 trials (79 two-arm trials, 13 three-arm trials, and one four-arm trial). The primary outcome was response after treatment (yes/no), defined as a reduction from baseline by at least 50% on a depression scale. The dataset contains log odds ratios with standard errors for all pairwise comparisons.

The interventions comprised both medical and psychological treatments, also in combination, including placebo and usual care (UC) (Linde et al., 2016). Pharmacological interventions were tricyclic antidepressants (TCA), selective serotonin reuptake inhibitors (SSRI), serotonin-noradrenaline reuptake inhibitors (SNRI), noradrenaline reuptake inhibitors (NRI), low- dose serotonin (5-HT2) antagonists and reuptake inhibitors (low-dose SARI), noradrenergic and specific serotonergic agents (NaSSa), reversible inhibitors of monoaminoxidase A (rMAO-A), hypericum extracts, and an individualized drug. Psychological interventions were cognitive behavioral therapy (CBT; four forms: face-to-face CBT, remote therapist-led CBT, guided self-help CBT, and no or minimal contact CBT), face-to-face problem-solving therapy (PST), face-to-face interpersonal psychotherapy, face-to-face psychodynamic therapy, and “other face-to-face therapy”. Combination therapies were face-to-face CBT + SSRI, face-to-face PST + SSRI, and face-to-face interpersonal psychotherapy + SSRI.

The dataset was used as an example in Rücker et al. (2020) to illustrate component network meta-analysis using frequentist methods.

Source

Linde, K., Rücker, G., Schneider, A., & Kriston, L. (2016). Questionable assumptions hampered interpretation of a network meta-analysis of primary care depression treatments. Journal of Clinical Epidemiology, 71, 86–96. https://doi.org/10.1016/j.jclinepi.2015.10.010

References

Rücker, G., Petropoulou, M., & Schwarzer, G. (2020). Network meta-analysis of multicomponent interventions. Biometrical Journal, 62(3), 808–821. https://doi.org/10.1002/bimj.201800167

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

medicine, psychology, psychiatry, odds ratios, network meta-analysis, component network meta-analysis

Examples

### Show results of first three studies (first study has three treatment arms)
head(dat.linde2016, 7)
#>   id    author year resp  n              int         int.long
#> 1  1 Lecrubier 1997   60 78             SNRI             SNRI
#> 2  1 Lecrubier 1997   48 76          Placebo          Placebo
#> 3  1 Lecrubier 1997   49 75              TCA              TCA
#> 4  2      Smit 2006   16 44 Face-to-face CBT Face-to-face CBT
#> 5  2      Smit 2006   19 72               UC               UC
#> 6  4   Blashki 1971   20 35              TCA              TCA
#> 7  4   Blashki 1971    8 23          Placebo          Placebo

### Load netmeta package
suppressPackageStartupMessages(library(netmeta))

### Print odds ratios and confidence limits with two digits
oldset <- settings.meta(digits = 2)

### Define order of treatments in printouts and forest plots
trts <- c("SSRI", "Face-to-face CBT", "Face-to-face interpsy", "Face-to-face
  PST", "Face-to-face CBT + SSRI", "Face-to-face interpsy + SSRI",
  "Face-to-face PST + SSRI", "Face-to-face psychodyn", "Other face-to-face",
  "TCA", "SNRI", "NRI", "Low-dose SARI", "NaSSa", "rMAO-A", "Ind drug",
  "Hypericum", "Remote CBT", "Self-help CBT", "No contact CBT", "UC",
  "Placebo")

### Use pairwise() to transform data to comparison-based format
pw <- pairwise(treat = int,
  event = resp, n = n,
  studlab = paste(author, year),
  data = dat.linde2016,
  reference = "plac",
  sm = "OR")

### Conduct random effects network meta-analysis
nma <- netmeta(pw, reference.group = "placebo",
  seq = trts, common = FALSE)
#> Error: Argument 'seq' must be a permutation of the following values:
#>   'Face-to-face CBT' - 'Face-to-face CBT + SSRI' - 'Face-to-face PST' - 'Face-to-face PST + SSRI' - 'Face-to-face interpsy' - 'Face-to-face interpsy + SSRI' - 'Face-to-face psychodyn' - 'Hypericum' - 'Ind drug' - 'Low-dose SARI' - 'NRI' - 'NaSSa' - 'No contact CBT' - 'Other face-to-face' - 'Placebo' - 'Remote CBT' - 'SNRI' - 'SSRI' - 'Self-help CBT' - 'TCA' - 'UC' - 'rMAO-A'

### Network graph
netgraph(nma, seq = "o")
#> Error: object 'nma' not found

### Show results
nma
#> Error: object 'nma' not found
forest(nma, xlim = c(0.2, 50))
#> Error: object 'nma' not found

### Additive component network meta-analysis with placebo as inactive treatment
cnma <- netcomb(nma, inactive = "placebo")
#> Error: object 'nma' not found
cnma
#> Error: object 'cnma' not found
forest(cnma, xlim = c(0.2, 50))
#> Error: object 'cnma' not found

### Use previous settings
settings.meta(oldset)