dat.nakagawa2007 {metadat}	R Documentation

Assessing the Function of House Sparrows' Bib Size Using a Flexible Meta-Analysis Method

Description

A meta-analysis on the association between the size of a male's bib and their social status in house sparrows (Passer domesticus).

Usage


dat.nakagawa2007

Format

The data frame contains the following columns:

StudyID 	 character 	 identity of primary study

Place 	 character 	 location of study population

Correlation 	 numeric 	 correlation coefficient

SampleSize 	 integer 	 sample size of population

Details

Each study measures the association between a sparrows bib size and its social status. Effects are quantified as correlation coefficients.

Concepts

ecology, correlation coefficients

Author(s)

Daniel Noble, daniel.noble@anu.edu.au

Source

Nakagawa, S., Ockendon, N., Gillespie, D. O. S, Hatchwell, B. J., & Burke, T. (2007). Assessing the function of house sparrows' bib size using a flexible meta-analysis method. Behavioral Ecology, 18(5), 831–840. https://doi.org/10.1093/beheco/arm050

Examples


### copy data into 'dat' and examine data
dat <- dat.nakagawa2007
dat

## Not run: 

### load metafor package
library(metafor)

### calculate Zr
dat <- escalc(measure="ZCOR", ri=Correlation, ni=SampleSize, data=dat)

### fit meta-analytic model
res <- rma.mv(yi, vi, random = list(~ 1 | StudyID), data=dat)
res


## End(Not run)

[Package metadat version 1.2-0 Index]