dat.pignon2000 {metadat}	R Documentation

Studies on the Effectiveness of Locoregional Treatment plus Chemotherapy for Head and Neck Squamous-Cell Carcinoma

Description

Results from studies examining mortality risk in patients with nonmetastatic head and neck squamous-cell carcinoma receiving either locoregional treatment plus chemotherapy versus locoregional treatment alone.

Usage


dat.pignon2000

Format

The data frame contains the following columns:

id 	 numeric 	 study id number

trial 	 character 	 trial abbreviation

OmE 	 numeric 	 observed minus expected number of deaths in the locoregional treatment plus chemotherapy group

V 	 numeric 	 corresponding variance

grp 	 numeric 	 timing of chemotherapy: 1 = adjuvant, 2 = neoadjuvant, 3 = concomitant

Details

The purpose of this meta-analysis was to examine the mortality risk in patients with nonmetastatic head and neck squamous-cell carcinoma receiving either locoregional treatment plus chemotherapy versus locoregional treatment alone. For 65 trials, the dataset provides the observed minus expected number of deaths and corresponding variances in the locoregional treatment plus chemotherapy group. Based on these values, we can estimate the log hazard ratios with OmE/V and the corresponding sampling variance with 1/V.

The trials were also divided according to the timing of the chomotherapy: (1) adjuvant, after the locoregional treatment, (2) neoadjuvant, before the locoregional treatment, and (3) concomitant, chemotherapy given concomitantly or alternating with radiotherapy.

Concepts

medicine, oncology, hazard ratios

Author(s)

Wolfgang Viechtbauer, wvb@metafor-project.org, https://www.metafor-project.org

Source

Pignon, J. P., Bourhis, J., Domenge, C., & Designe, L. (2000). Chemotherapy added to locoregional treatment for head and neck squamous-cell carcinoma: Three meta-analyses of updated individual data. Lancet, 355(9208), 949–955. https://doi.org/10.1016/S0140-6736(00)90011-4

Examples


### copy data into 'dat' and examine data
dat <- dat.pignon2000
head(dat)

## Not run: 

### load metafor package
library(metafor)

### calculate log hazard ratios and sampling variances
dat$yi <- with(dat, OmE/V)
dat$vi <- with(dat, 1/V)
head(dat)

### meta-analysis based on all 65 trials
res <- rma(yi, vi, data=dat, method="EE", digits=2)
res
predict(res, transf=exp)

### only adjuvant trials
res <- rma(yi, vi, data=dat, method="EE", subset=grp==1, digits=2)
res
predict(res, transf=exp)

### only neoadjuvant trials
res <- rma(yi, vi, data=dat, method="EE", subset=grp==2, digits=2)
res
predict(res, transf=exp)

### only concomitant trials
res <- rma(yi, vi, data=dat, method="EE", subset=grp==3, digits=2)
res
predict(res, transf=exp)


## End(Not run)

[Package metadat version 1.2-0 Index]