Trials on Single-Agent Purine Analogues for the Treatment of Chronic Lymphocytic Leukaemia

Results from 4 trials evaluating single-agent purine analogues in patients with chronic lymphocytic leukaemia.

dat.steurer2006

Format

The data frame contains the following columns:

author	character	first author

year	integer	publication year

Ne	integer	number of patients (purine antagonists)

Nc	integer	number of patients (alkylator-based)

HR	numeric	hazard ratio

lowHR	numeric	lower limit

uppHR	numeric	upper limit

lnHR	numeric	log hazard ratio

selnHR	numeric	standard error

Details

Steurer et al. (2006) conducted a Cochrane review to evaluate the effect of single-agent purine analogues for the treatment of chronic lymphocytic leukaemia. This dataset contains data from the main outcome, overall survival. Note, the hazard ratios and confidence limits have been reported in the Cochrane review with only two significant figures and were recalculated using the reported log hazard ratios and standard errors.

This data set is used as an example in Schwarzer et al. (2015).

Source

Steurer, M., Pall, G., Richards, S., Schwarzer, G., Bohlius, J., & Greil, R. (2006). Purine antagonists for chronic lymphocytic leukaemia. Cochrane Database of Systematic Reviews, 3, CD004270. https://doi.org/10.1002/14651858.cd004270.pub2

References

Schwarzer, G., Carpenter, J. R., & Rücker, G. (2015). Meta-analysis with R. Cham, Switzerland: Springer.

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

medicine, oncology, hazard ratios

Examples

### Show dataset
dat.steurer2006
#>       author year  Ne  Nc     HR  lowHR  uppHR    lnHR selnHR
#> 1 FCG on CLL 1996  53  52 0.5532 0.2813 1.0878 -0.5920 0.3450
#> 2  Leporrier 2001 341 597 0.9239 0.7919 1.0780 -0.0791 0.0787
#> 3        Rai 2000 195 200 0.7890 0.5950 1.0463 -0.2370 0.1440
#> 4      Robak 2000 133 117 1.1770 0.6386 2.1695  0.1630 0.3120

### Load meta package
suppressPackageStartupMessages(library(meta))

### Use RevMan 5 settings
oldset <- settings.meta("RevMan5")

### Conduct common effect meta-analysis
m1 <- metagen(lnHR, selnHR, data = dat.steurer2006,
  studlab = paste(author, year),
  sm = "HR", random = FALSE, n.e = Ne, n.c = Nc)
m1
#> Number of studies: k = 4
#> Number of observations: o = 1688 (o.e = 722, o.c = 966)
#> 
#>                      HR       95%-CI     z p-value
#> Fixed effect model 0.89 [0.78, 1.01] -1.82  0.0688
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 0.006 [0.000, 0.855]; tau = 0.0778 [0.0000, 0.9245]
#>  I^2 = 17% [0%, 87%]; H = 1.10 [1.00, 2.81]
#> 
#> Test of heterogeneity:
#>     Q d.f. p-value
#>  3.62    3  0.3049
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q

### Same analysis using lower and upper confidence limits
m2 <- metagen(HR, lower = lowHR, upper = uppHR,
  data = dat.steurer2006, transf = FALSE,
  studlab = paste(author, year),
  sm = "HR", random = FALSE, n.e = Ne, n.c = Nc)
m2
#> Number of studies: k = 4
#> Number of observations: o = 1688 (o.e = 722, o.c = 966)
#> 
#>                      HR       95%-CI     z p-value
#> Fixed effect model 0.89 [0.78, 1.01] -1.82  0.0687
#> 
#> Quantifying heterogeneity (with 95%-CIs):
#>  tau^2 = 0.006 [0.000, 0.855]; tau = 0.0777 [0.0000, 0.9244]
#>  I^2 = 17% [0%, 87%]; H = 1.10 [1.00, 2.81]
#> 
#> Test of heterogeneity:
#>     Q d.f. p-value
#>  3.62    3  0.3050
#> 
#> Details of meta-analysis methods:
#> - Inverse variance method
#> - DerSimonian-Laird estimator for tau^2
#> - Jackson method for confidence interval of tau^2 and tau
#> - Calculation of I^2 based on Q

### Use previous settings
settings.meta(oldset)