Trials on Serum Cholesterin Concentration Lowering and Risk of Ischaemic Heart Disease

Results from 28 trials evaluating effect of serum cholesterin concentration lowering on risk of ischaemic heart disease

dat.thompson1999

Format

The data frame contains the following columns:

studyid	integer	study ID

ihd.cont	integer	number of ischaemic heart disease (control group)

noihd.cont	integer	number of non-events (control group)

ihd.exp	integer	number of ischaemic heart disease (treated group)

noihd.exp	integer	number of non-events (treated group)

OR	numeric	odds ratio

logOR	numeric	log odds ratio

varlogOR	numeric	variance of log odds ratio

cholr	numeric	cholesterol reduction (mmol/l)

Details

Thompson and Sharp (1999) compare several meta-regression approaches to explain heterogeneity in meta-analysis. The data set used is originally from Law et al. (1994), but with minor modifications.

Source

Thompson, S. G., & Sharp, S. J. (1999). Explaining heterogeneity in meta-analysis: A comparison of methods. Statistics in Medicine, 18(20), 2693–2708. https://doi.org/10.1002/(sici)1097-0258(19991030)18:20<2693::aid-sim235>3.0.co;2-v

References

Law, M. R., Wald, N. J., & Thompson, S. G. (1994). By how much and how quickly does reduction in serum cholesterol concentration lower risk of ischaemic heart disease? British Medical Journal, 308(6925), 367–372. https://doi.org/10.1136/bmj.308.6925.367

Author

Guido Schwarzer, guido.schwarzer@uniklinik-freiburg.de, https://github.com/guido-s/

Concepts

medicine, cardiology, odds ratios, meta-regression

Examples

### Show first five studies
head(dat.thompson1999, 5)
#>   studyid ihd.cont noihd.cont ihd.exp noihd.exp   OR  logOR varlogOR cholr
#> 1       1      210       5086     173      5158 0.81 -0.208   0.0109  0.55
#> 2       2       85        168      54       190 0.56 -0.577   0.0415  0.68
#> 3       3       75        292      54       296 0.71 -0.342   0.0387  0.85
#> 4       4      936       1853     676      1546 0.87 -0.144   0.0037  0.55
#> 5       5       69        215      42       103 1.27  0.239   0.0527  0.59

### Load meta package
suppressPackageStartupMessages(library(meta))

### Conduct meta-analysis
m <- metabin(ihd.exp, ihd.exp + noihd.exp, ihd.cont, ihd.cont + noihd.cont,
  data = dat.thompson1999, sm = "OR", method = "Inverse")

### Thompson & Sharp (1999), Table III
### (1) None
metareg(m, cholr, method.tau = "FE")
#> 
#> Fixed-Effects with Moderators Model (k = 28)
#> 
#> I^2 (residual heterogeneity / unaccounted variability): 31.34%
#> H^2 (unaccounted variability / sampling variability):   1.46
#> R^2 (amount of heterogeneity accounted for):            20.86%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 26) = 37.8663, p-val = 0.0623
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 11.8241, p-val = 0.0006
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb    ci.ub      
#> intrcpt    0.1208  0.0972   1.2424  0.2141  -0.0698   0.3113      
#> cholr     -0.4752  0.1382  -3.4386  0.0006  -0.7461  -0.2044  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
### (3a) Additive (MM)
metareg(m, cholr, method.tau = "DL")
#> 
#> Mixed-Effects Model (k = 28; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0165 (SE = 0.0160)
#> tau (square root of estimated tau^2 value):             0.1283
#> I^2 (residual heterogeneity / unaccounted variability): 31.34%
#> H^2 (unaccounted variability / sampling variability):   1.46
#> R^2 (amount of heterogeneity accounted for):            43.97%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 26) = 37.8663, p-val = 0.0623
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 8.3733, p-val = 0.0038
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb    ci.ub     
#> intrcpt    0.1595  0.1367   1.1668  0.2433  -0.1084   0.4275     
#> cholr     -0.5206  0.1799  -2.8937  0.0038  -0.8732  -0.1680  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>