dat.white2020 {metadat}	R Documentation

Studies on the Relationship between Sexual Signal Expression and Individual Quality

Description

Results from 41 studies examining the relationship between measures of individual quality and the expression of structurally coloured sexual signals.

Usage


dat.white2020

Format

The object is a data frame which contains the following columns:

study_id 	 character 	 study-level ID

obs 	 character 	 observation-level ID

exp_obs 	 character 	 whether the study is observational or experimental

control 	 numeric 	 whether the study did (1) or did not (0) include a non-sexual control trait

class 	 character 	 class of the study organisms

genus 	 character 	 class of the study organisms

species 	 character 	 species of the study organisms

sex 	 character 	 sex of the study organisms

iridescent 	 numeric 	 whether the colour signals were iridescent (1) or not (0)

col_var 	 character 	 the colour variable quantified

col_component 	 character 	 whether the colour variable is chromatic or achromatic

quality_measure 	 character 	 the measure of individual quality used

region 	 character 	 the body region from which colour was sampled

n 	 numeric 	 study sample size

r 	 numeric 	 Pearson's correlation coefficient

Details

The 186 rows in this dataset come from 41 experimental and observational studies reporting on the correlation between measures of individual quality (age, body condition, immune function, parasite resistance) and the expression of structurally coloured sexual signals across 28 species. The purpose of this meta-analysis was to test whether structural colour signals show heightened condition-dependent expression, as predicted by evolutionary models of 'honest' signalling.

Concepts

ecology, evolution, correlation coefficients

Author(s)

Thomas E. White, thomas.white@sydney.edu.au

Source

White, T. E. (2020). Structural colours reflect individual quality: A meta-analysis. Biology Letters, 16(4), 20200001. https://doi.org/10.1098/rsbl.2020.0001

Examples


### copy data into 'dat' and examine data
dat <- dat.white2020
head(dat, 10)

## Not run: 

### load metafor package
library(metafor)

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=r, ni=n, data=dat)

### fit multilevel meta-analytic model
res <- rma.mv(yi, vi, random = list(~ 1 | study_id, ~ 1 | obs), data=dat)
res


## End(Not run)

[Package metadat version 1.2-0 Index]