Search function for the datasets

Function to search among the existing datasets.

datsearch(pattern, concept=TRUE, matchall=TRUE, fixed=TRUE, pkgdown=FALSE)

Arguments

pattern
character string or vector of strings specifying the search pattern(s) to search for within the datasets.

concept
logical indicating whether the search should be confined to the concept terms (TRUE by default) or whether a full-text search should be conducted.

matchall
logical indicating whether only the datasets matching all patterns (if multiple are specified) are returned (TRUE by default) or whether datasets matching any one of the patterns are returned.

fixed
logical indicating whether a pattern is a string to be matched as is (TRUE by default). If FALSE, a pattern is a regular expression that grep will search for. Only relevant when concept=FALSE (i.e., when doing a full-text search).

pkgdown
logical indicating whether the local help file or the pkgdown docs (at https://wviechtb.github.io/metadat/index.html) should be shown for the chosen dataset (FALSE by default).

Details

The function can be used to search all existing datasets in the metadat package based on their concept terms (see below) or based on a full-text search of their corresponding help files. One can also specify multiple search patterns (by using a string vector as the pattern). If matchall=TRUE (the default), only the datasets matching all patterns are returned. If matchall=FALSE, the datasets matching any one of the patterns are returned.

If one or multiple datasets are found, the user is prompted to choose one of the matching datasets of interest and the corresponding help file is then shown.

Concept Terms

Each dataset is tagged with one or multiple concept terms that refer to various aspects of a dataset, such as the field/topic of research, the outcome measure used for the analysis, the model(s) used for analyzing the data, and the methods/concepts that can be illustrated with the dataset.

In terms of ‘fields/topics’, the following tags have been used one or multiple times: alternative medicine, cardiology, climate change, covid-19, criminology, dentistry, ecology, education, epidemiology, evolution, medicine, obstetrics, oncology, primary care, psychiatry, psychology, social work, sociology.

In terms of ‘outcome measures’, the following tags have been used one or multiple times: correlation coefficients, Cronbach's alpha, hazard ratios, incidence rates, mean differences, odds ratios, proportions, ratios of means, risk differences, risk ratios, (semi-)partial correlations, standardized mean change, standardized mean differences.

In terms of ‘methods/concepts’, the following tags have been used one or multiple times: cluster-robust inference, cumulative meta-analysis, diagnostic accuracy studies, dose response models, generalized linear models, longitudinal models, Mantel-Haenszel method, meta-regression, model checks, multilevel models, multivariate models, network meta-analysis, outliers, Peto's method, phylogeny, publication bias, reliability generalization, single-arm studies, spatial correlation.

Author

Daniel Noble daniel.noble@anu.edu.au Wolfgang Viechtbauer wvb@metafor-project.org

Examples

# note: the examples below are not run since they require interactivity

if (FALSE) {

   # find all datasets tagged with the concept term 'standardized mean differences'
   datsearch("standardized mean differences")

   # find all datasets tagged with the concept terms 'odds ratio' AND 'multilevel'
   datsearch(c("odds ratio", "multilevel"))

   # do a full-text search for the term 'infarct'
   datsearch("infarct", concept=FALSE)

   # do a full-text search for 'rma.mv(' (essentially finds all datasets where the rma.mv()
   # function was used in the examples section of a help file)
   datsearch("rma.mv(", concept=FALSE)

}